### 2022-10-27 #107 @whedon
[PRE REVIEW]: DisjunctiveProgramming.jl: Generalized Disjunctive Programming Models and Algorithms for JuMP
**Submitting author:** @hdavid16 (<a href="http://orcid.org/0000-0002-7374-1120">Hector Perez</a>)
**Repository:** <a href="https://github.com/hdavid16/DisjunctiveProgramming.jl" target ="_blank">https://github.com/hdavid16/DisjunctiveProgramming.jl</a>
**Version:**
**Managing EiC:** Mathieu Besançon
