### 2022-11-04 5a03217
reduce allocaitons
