### 2022-11-04 a7bf324
attempt to improve speed using by storing disaggregated vars in a Set instead of vector
