### 2022-11-04 2592c94
avoid stackoverflow error for large disjunctions
