### 2022-11-09 591d71c
reduce allocations: don't use variable_by_name, but store this in .ext Dict
