### 2022-12-05 #107 @pulsipher
> @pulsipher would you be available to review

I can, my only potential conflict of interest would be that I have helped develop the package.
