### 2022-12-05 #107 @hdavid16
> I can, my only potential conflict of interest would be that I have helped develop the package.

I will note that the contents of the paper is prior to any contributions from @pulsipher to the repo.
