### 2022-12-11 #117 @whedon
[REVIEW]: DisjunctiveProgramming.jl: Generalized Disjunctive Programming Models and Algorithms for JuMP
**Submitting author:** @hdavid16 (<a href="http://orcid.org/0000-0002-7374-1120">Hector Perez</a>)
**Repository:** <a href="https://github.com/hdavid16/DisjunctiveProgramming.jl" target ="_blank">https://github.com/hdavid16/DisjunctiveProgramming.jl</a>
**Version:** v0.4.1
