### 2022-12-11 #117 @matbesancon
@sbolusani @joaquimg the reviewer's guide is here: https://juliacon.github.io/proceedings-guide/reviewer/
