### 2022-12-23 afd934a
Reference updates
Fixed doi's and one duplicate reference
