+++ paper/ref.bib
@@ -4,6 +4,7 @@
   booktitle={International Symposium on Practical Aspects of Declarative Languages},
   pages={134--148},
   year={2010},
+  doi="10.1007/978-3-642-11503-5\_12",
   organization={Springer}
 }
 @article{huangfu2018parallelizing,
@@ -14,6 +15,7 @@
   number={1},
   pages={119--142},
   year={2018},
+  doi={10.1007/s12532-017-0130-5},
   publisher={Springer}
 }
 @article{chen2022pyomo,
@@ -34,6 +36,7 @@
   number={4-5},
   pages={555--565},
   year={1999},
+  doi={10.1016/s0098-1354(98)00293-2},
   publisher={Elsevier}
 }
 @article{trespalacios_grossmann_2016,
@@ -49,7 +52,7 @@ pages={209–222}
 
 @Inbook{E.Grossmann2009,
 author="E. Grossmann, Ignacio",
-title="Logic-based outer approximationLogic-Based Outer Approximation",
+title="Logic-based outer approximation",
 bookTitle="Encyclopedia of Optimization",
 year="2009",
 publisher="Springer US",
@@ -68,7 +71,7 @@ volume = {67},
 pages = {13-25},
 year = {2014},
 issn = {0098-1354},
-doi = {https://doi.org/10.1016/j.compchemeng.2014.03.014},
+doi = {10.1016/j.compchemeng.2014.03.014},
 url = {https://www.sciencedirect.com/science/article/pii/S0098135414000957},
 author = {Miguel A. Navarro-Amorós and Rubén Ruiz-Femenia and José A. Caballero},
 keywords = {Process synthesis, Generalized Disjunctive Programming, Modular simulators, Logic-based optimization algorithm},
@@ -106,6 +109,7 @@ volume={3},
 number={3},
 pages={219--260},
 year={2011},
+doi={10.1007/s12532-011-0026-8},
 publisher={Springer}
 }
 
@@ -119,7 +123,7 @@ pages = {841-846},
 year = {2022},
 booktitle = {32nd European Symposium on Computer Aided Process Engineering},
 issn = {1570-7946},
-doi = {https://doi.org/10.1016/B978-0-323-95879-0.50141-7},
+doi = {10.1016/B978-0-323-95879-0.50141-7},
 url = {https://www.sciencedirect.com/science/article/pii/B9780323958790501417},
 author = {Seolhee Cho and Ignacio E. Grossmann},
 keywords = {Reliability, Expansion Planning, Power systems, Optimization},
@@ -133,7 +137,7 @@ volume = {157},
 pages = {107616},
 year = {2022},
 issn = {0098-1354},
-doi = {https://doi.org/10.1016/j.compchemeng.2021.107616},
+doi = {10.1016/j.compchemeng.2021.107616},
 url = {https://www.sciencedirect.com/science/article/pii/S009813542100394X},
 author = {Ying Chen and Yixin Ye and Zhihong Yuan and Ignacio E. Grossmann and Bingzhen Chen},
 keywords = {Reliability-based superstructure optimization, Stochastic programming, Endogenous and exogenous uncertainties, Logic-based outer approximation algorithm},
@@ -147,7 +151,7 @@ volume = {163},
 pages = {107856},
 year = {2022},
 issn = {0098-1354},
-doi = {https://doi.org/10.1016/j.compchemeng.2022.107856},
+doi = {10.1016/j.compchemeng.2022.107856},
 url = {https://www.sciencedirect.com/science/article/pii/S0098135422001946},
 author = {Fahad Matovu and Shuhaimi Mahadzir and Rasel Ahmed and Nor Erniza Mohammad Rozali},
[…]
