### 2022-12-24 #117 @hdavid16
> @hdavid16 can you check the DOIs listed above, that's minor but can be done in parallel

DOIs have been updated in the repo now.
