### 2022-12-31 bc193fb
update docstring and kwarg types
