### 2023-01-03 f1be88a
single disjunct constraints are allowed (always selected as long as choose at least 1 is used for the binaries)
