### 2023-01-05 e1bfb33
speed up by using filter! and disag. vars. Dict
