### 2023-01-05 7b49496
check that disjunction is not empty
