+++ src/macros.jl
@@ -85,6 +85,7 @@ Add disjunction and reformulate.
 function add_disjunction!(m::Model,disj...;reformulation::Symbol,M=missing,ϵ=1e-6,name=missing)
     #run checks
     @assert reformulation in [:big_m, :hull] "Invalid reformulation method passed to keyword argument `:reformulation`. Valid options are :big_m (Big-M Reformulation) and :hull (Hull Reformulation)."
+    @assert !all(isnothing.(disj)) "Must have at least 1 disjunct that is not `nothing`."
     # @assert length(disj) > 1 "At least 2 disjuncts must be included. If there is an empty disjunct, use `nothing`."
     
     #create indicator variable
