+++ Project.toml
@@ -1,7 +1,7 @@
 name = "DisjunctiveProgramming"
 uuid = "0d27d021-0159-4c7d-b4a7-9ccb5d9366cf"
 authors = ["hdavid16 <hperez16@gmail.com>"]
-version = "0.3.5"
+version = "0.3.6"
 
 [deps]
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
