### 2023-01-10 4d0f337
fix bug in how m.ext is updated
