### 2023-01-10 7c1bf50
return new NL constraints; change name split func
