### 2023-01-10 ee34be8
track nested perspective functions to avoid issues
