### 2023-01-21 b11f39c
fix Interval arithmetic can't be applied to quadratic constraints to insert big M value #62
