+++ src-old/logic.jl
@@ -1,290 +1,0 @@
-"""
-    choose!(m::Model, n::Int, vars::VariableRef...; mode)
-
-Add constraint to select n elements from the list of variables. Options for mode
-are `:at_least`, `:at_most`, `:exactly`.
-"""
-function choose!(m::Model, n::Int, vars::VariableRef...; mode=:exactly)
-    @assert length(vars) >= n "Not enough variables passed."
-    @assert all(is_valid.(m, vars)) "Invalid VariableRefs passed."
-    add_selection!(m, n, vars...; mode)
-end
-function choose!(m::Model, vars::VariableRef...; mode=:exactly)
-    @assert all(is_valid.(m, vars)) "Invalid VariableRefs passed."
-    n = vars[1] #first variable is the n
-    add_selection!(m, n, vars...; mode)
-end
-function add_selection!(m::Model, n, vars::VariableRef...; mode::Symbol)
-    display(n)
-    if mode == :exactly
-        display(@constraint(m, sum(vars) == n))
-    elseif mode == :at_least
-        @constraint(m, sum(vars) ≥ n)
-    elseif mode == :at_most
-        @constraint(m, sum(vars) ≥ n)
-    end
-end
-
-"""
-    add_proposition!(m::Model, expr::Expr)
-
-Convert logical proposition expression into conjunctive normal form.
-"""
-function add_proposition!(m::Model, expr::Expr)
-    expr_name = Symbol("{$expr}") #get name to register reformulated logical proposition
-    replace_Symvars!(expr, m; logical_proposition = true) #replace all JuMP variables with Symbolic variables
-    clause_list = to_cnf!(expr)
-    #replace symbolic variables with JuMP variables and boolean operators with their algebraic counterparts
-    for clause in clause_list
-        replace_JuMPvars!(clause, m)
-        replace_logic_operators!(clause)
-    end
-    #generate and simplify JuMP expressions for the lhs of the algebraic constraints
-    lhs = eval.(clause_list)
-    drop_zeros!.(lhs)
-    unique!(lhs)
-    #generate JuMP constraints for the logical proposition
-    if length(lhs) == 1
-        m[expr_name] = @constraint(m, lhs[1] >= 1, base_name = string(expr_name))
-    else
-        m[expr_name] = @constraint(m, [i = eachindex(lhs)], lhs[i] >= 1, base_name = string(expr_name))
-    end
-end
-
-"""
-    to_cnf!(expr::Expr)
-
-Convert an expression of symbolic Boolean variables and operators to CNF.
-"""
-function to_cnf!(expr::Expr)
-    check_logical_proposition(expr) #check that valid boolean symbols and variables are used in the logical proposition
-    eliminate_equivalence!(expr) #eliminate ⇔
-    eliminate_implication!(expr) #eliminmate ⇒
-    move_negations_inwards!(expr) #expand ¬
-    clause_list = distribute_and_over_or_recursively!(expr) #distribute ∧ over ∨ recursively
-    @assert !isempty(clause_list) "Conversion to CNF failed."
-
-    return clause_list
-end
-
-"""
-    check_logical_proposition(expr::Expr)
-
-Validate logical proposition provided.
-"""
-function check_logical_proposition(expr::Expr)
-    #NOTE: this is quick and dirty (uses Suppressor.jl). A more robust approach should traverse the expression tree to verify that only valid boolean symbols and model variables are used.
-    dump_str = @capture_out dump(expr, maxdepth = typemax(Int)) #caputre dump
-    dump_arr = split(dump_str,"\n") #split by \n
-    filter!(i -> occursin("1:",i), dump_arr) #filter only first args in each subexpression
[…]
