### 2023-05-03 c826cef
add methods for disjunctive constraints
