+++ src/constraints.jl
@@ -20,3 +20,49 @@ function JuMP.add_constraint(
     idx = _MOIUC.add_item(gdp_data(model).disjunctions, constr_data)
     return DisjunctiveConstraintRef(model, idx)
 end
+
+"""
+
+"""
+JuMP.owner_model(cref::DisjunctiveConstraintRef) = cref.model
+
+"""
+
+"""
+JuMP.index(cref::DisjunctiveConstraintRef) = cref.index
+
+"""
+
+"""
+function JuMP.is_valid(model::JuMP.Model, cref::DisjunctiveConstraintRef)
+    return model === JuMP.owner_model(cref)
+end
+
+"""
+
+"""
+function JuMP.name(cref::DisjunctiveConstraintRef)
+    constr_data = gdp_data(JuMP.owner_model(cref))
+    return constr_data.disjunctions[JuMP.index(cref)].name
+end
+
+"""
+
+"""
+function JuMP.set_name(cref::DisjunctiveConstraintRef, name::String)
+    constr_data = gdp_data(JuMP.owner_model(cref))
+    constr_data.disjunctions[JuMP.index(cref)].name = name
+    return
+end
+
+"""
+
+"""
+function JuMP.delete(model::JuMP.Model, cref::DisjunctiveConstraintRef)
+    @assert JuMP.is_valid(model, cref) "Disjunctive constraint does not belong to model."
+    constr_data = gdp_data(JuMP.owner_model(cref))
+    dict = constr_data.disjunctions[JuMP.index(cref)]
+    # TODO check if used by a disjunction and/or a proposition
+    delete!(dict, index(cref))
+    return 
+end
