+++ src/DisjunctiveProgramming.jl
@@ -15,6 +15,9 @@ include("model.jl")
 include("variables.jl")
 include("constraints.jl")
 include("propositions.jl")
+include("reformulate.jl")
+include("bigm.jl")
+include("hull.jl")
 include("macros.jl")
 include("optimize.jl")
 
