### 2023-05-11 708a8cb
extend copy_extension_data
