+++ Project.toml
@@ -1,22 +1,18 @@
 name = "DisjunctiveProgramming"
 uuid = "0d27d021-0159-4c7d-b4a7-9ccb5d9366cf"
 authors = ["hdavid16 <hperez16@gmail.com>"]
-version = "0.2.1"
+version = "0.4.0"
 
 [deps]
 IntervalArithmetic = "d1acc4aa-44c8-5952-acd4-ba5d80a2a253"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
-Suppressor = "fd094767-a336-5f1f-9728-57cf17d0bbfb"
-Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
 LeftChildRightSiblingTrees = "1d6d02ad-be62-4b6b-8a6d-2f90e265016e"
 
 [compat]
 IntervalArithmetic = "0.20"
 JuMP = "1"
 LeftChildRightSiblingTrees = "0.1"
-Suppressor = "0.2"
-Symbolics = "4"
-julia = "1"
+julia = "1.6"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
