+++ examples/ex1.jl
@@ -46,11 +46,7 @@ print(m)
 #  x + 7 Y[1] <= 10.0
 #  x + Y[2] <= 10.0
 #  x >= -10.0
-#  Y[1] >= 0.0
-#  Y[2] >= 0.0
 #  x <= 10.0
-#  Y[1] <= 1.0
-#  Y[2] <= 1.0
 #  Y[1] binary
 #  Y[2] binary
 
@@ -69,13 +65,9 @@ print(m)
 #  -3 Y[1] + x_Y[1] <= 0.0
 #  -9 Y[2] + x_Y[2] <= 0.0
 #  x >= -10.0
-#  Y[1] >= 0.0
-#  Y[2] >= 0.0
 #  x_Y[1] >= -10.0
 #  x_Y[2] >= -10.0
 #  x <= 10.0
-#  Y[1] <= 1.0
-#  Y[2] <= 1.0
 #  x_Y[1] <= 10.0
 #  x_Y[2] <= 10.0
 #  Y[1] binary
