### 2023-05-11 5ddb008
remove bounds on binary indicator vars
