### 2023-05-11 1f594eb
extend methods for LogicVariableRef
