### 2023-05-11 96af35f
change LogicalIndex to LogicalVariableIndex (since JuMP uses VariableIndex)
