### 2023-05-11 a33702c
extend getindex to allow copying GDPModel
