### 2023-05-12 fa53c52
copy_extension_data needs more work
