### 2023-07-11 6952e33
use Sets instead of Vectors for querying variables
