+++ src/hull.jl
@@ -39,56 +39,50 @@ end
 """
 
 """
-function _disaggregate_variables(model::JuMP.Model, disj::DisjunctiveConstraintData)
-    disj_vars = _get_variables(disj)
-    disag_var_dict = gdp_data(model).disaggregated_variables
-    ind_var_dict = gdp_data(model).indicator_variables
-    for d in disj.constraint.disjuncts
-        #create binary variable for logic variable (indicator)
-        bvar = JuMP.@variable(model, 
-            base_name = string(d.indicator), 
-            binary = true, 
-        )
-        ind_var_dict[Symbol(d.indicator,"_Bin")] = bvar
-    end
-    for var in disj_vars
-        JuMP.is_binary(var) && continue #skip binary variables
-        if !JuMP.has_lower_bound(var) | !JuMP.has_upper_bound(var)
+function _update_variable_bounds!(var_bounds_dict::Dict{JuMP.VariableRef, Tuple{Float64, Float64}}, var_set::Set{JuMP.VariableRef})
+    for var in var_set
+        if (var in keys(var_bounds_dict)) | JuMP.is_binary(var) #skip if binary or if bounds already stored
+            continue
+        elseif !JuMP.has_lower_bound(var) | !JuMP.has_upper_bound(var)
             error("Variable $var must have a lower and upper bound defined when using the Hull reformulation.")
-        end
-        sum_disag_vars = JuMP.AffExpr(0) #sum of disaggregated variables
-        for d in disj.constraint.disjuncts
-            #create disaggregated var
+        else
             lb = min(0, JuMP.lower_bound(var))
             ub = max(0, JuMP.upper_bound(var))
-            disag_var_name = string(var,"_",d.indicator)
-            disag_var = JuMP.@variable(model, 
-                base_name = disag_var_name,
-                lower_bound = lb,
-                upper_bound = ub,
-            )
-            disag_var_dict[Symbol(disag_var_name)] = disag_var
-            #create bounding constraints
-            bvar = ind_var_dict[Symbol(d.indicator,"_Bin")]
-            JuMP.@constraint(model, lb*bvar - disag_var ≤ 0)
-            JuMP.@constraint(model, disag_var - ub*bvar ≤ 0)
-            #update aggregation constraint
-            JuMP.add_to_expression!(sum_disag_vars, 1, disag_var)
+            var_bounds_dict[var] = (lb, ub)
         end
-        JuMP.@constraint(model, var == sum_disag_vars)
     end
-
 end
 
 """
 
 """
-function _disaggregated_constraint(model::JuMP.Model, con::JuMP.ScalarConstraint{JuMP.AffExpr,T}, lvar::LogicalVariableRef) where {T}
+function _disaggregate_variable(model::JuMP.Model, d::Disjunct, var::JuMP.VariableRef, bvar::JuMP.VariableRef)
+    #create disaggregated var
+    lb, ub = gdp_data(model).variable_bounds[var]
+    disag_var_name = string(var, "_", d.indicator)
+    disag_var = JuMP.@variable(model, 
+        base_name = disag_var_name,
+        lower_bound = lb,
+        upper_bound = ub,
+    )
+    #store ref to disaggregated variable
+    gdp_data(model).disaggregated_variables[Symbol(disag_var_name)] = disag_var
+    #create bounding constraints
+    JuMP.@constraint(model, lb*bvar - disag_var ≤ 0)
+    JuMP.@constraint(model, disag_var - ub*bvar ≤ 0)
+
+    return disag_var
+end
+
+"""
+
+"""
[…]
