+++ src/bigm.jl
@@ -3,27 +3,19 @@
 
 Apply interval arithmetic on a linear constraint to infer the tightest Big-M value from the bounds on the constraint.
 """
-function _calculate_tight_M(
-    con::JuMP.ScalarConstraint{JuMP.AffExpr, <: _MOI.LessThan}
-    )
+function _calculate_tight_M(con::JuMP.ScalarConstraint{JuMP.AffExpr, S}) where {S <: _MOI.LessThan}
     return _interval_arithmetic_LessThan(con, -con.set.upper)
 end
-function _calculate_tight_M(
-    con::JuMP.ScalarConstraint{JuMP.AffExpr, <: _MOI.GreaterThan}
-    )
+function _calculate_tight_M(con::JuMP.ScalarConstraint{JuMP.AffExpr, S}) where {S <: _MOI.GreaterThan}
     return _interval_arithmetic_GreaterThan(con, -con.set.lower)
 end
-function _calculate_tight_M(
-    con::JuMP.ScalarConstraint{JuMP.AffExpr, <: _MOI.Interval}
-    )
+function _calculate_tight_M(con::JuMP.ScalarConstraint{JuMP.AffExpr, S}) where {S <: _MOI.Interval}
     return [
         _interval_arithmetic_GreaterThan(con, -con.set.lower),
         _interval_arithmetic_LessThan(con, -con.set.upper)
     ]
 end
-function _calculate_tight_M(
-    con::JuMP.ScalarConstraint{JuMP.AffExpr, <: _MOI.EqualTo}
-    )
+function _calculate_tight_M(con::JuMP.ScalarConstraint{JuMP.AffExpr, S}) where {S <: _MOI.EqualTo}
     return [
         _interval_arithmetic_GreaterThan(con, -con.set.value),
         _interval_arithmetic_LessThan(con, -con.set.value)
@@ -34,7 +26,7 @@ _calculate_tight_M(con::JuMP.ScalarConstraint) = Inf
 """
 
 """
-function _interval_arithmetic_LessThan(con::JuMP.ScalarConstraint{JuMP.AffExpr,T}, M::Float64) where {T}
+function _interval_arithmetic_LessThan(con::JuMP.ScalarConstraint{JuMP.AffExpr, T}, M::Float64) where {T}
     for (var,coeff) in con.func.terms
         JuMP.is_binary(var) && continue #skip binary variables
         if coeff > 0
@@ -52,7 +44,7 @@ end
 """
 
 """
-function _interval_arithmetic_GreaterThan(con::JuMP.ScalarConstraint{JuMP.AffExpr,T}, M::Float64) where {T}
+function _interval_arithmetic_GreaterThan(con::JuMP.ScalarConstraint{JuMP.AffExpr, T}, M::Float64) where {T}
     for (var,coeff) in con.func.terms
         JuMP.is_binary(var) && continue #skip binary variables
         if coeff < 0
