### 2023-07-12 0d1866f
close #64; add epsilon value for Hull
