### 2023-07-12 a5451de
update fallbacks for quadexpr
