### 2023-07-13 4e926df
implement copy_extension_data
