+++ src/model.jl
@@ -42,14 +42,17 @@ function JuMP.copy_extension_data(data::GDPData, new_model::JuMP.AbstractModel, 
     #TODO need to copy disjunction constraints over to the new model
     #   (maybe something similar to what is done in copy_to in MOI.jl/src/Utilities/copy.jl).
     #   Otherwise, the constraints inside the disjuncts are not owned by the new model...
-    
-    # new_model.ext[:GDP] = GDPData(
-    #     data.logical_variables, 
-    #     data.disjunctions, 
-    #     data.propositions, 
-    #     data.solution_method, 
-    #     data.ready_to_optimize
-    # )
+
+    new_model.ext[:GDP] = GDPData(
+        data.logical_variables, 
+        _copy_disjunctions(data.disjunctions, new_model), 
+        data.propositions, 
+        data.solution_method, 
+        data.ready_to_optimize,
+        data.disaggregated_variables,
+        data.indicator_variables,
+        data.variable_bounds
+    )
 end
 
 # Determine if the model is ready to call `optimize!` without a optimize hook
