### 2023-07-13 #117 @hdavid16
Hi @sbolusani and @joaquimg, I'm sure you are all super busy. But just want to remind you to review this conference paper. It's been almost a year since the last JuliaCon where this work was presented...
