### 2023-07-18 929413c
add big-m support for NLP; fix bugs
