+++ src/hull.jl
@@ -1,40 +1,125 @@
 """
 
 """
-function _get_variables(disj::DisjunctiveConstraintData)
+function _get_disjunction_variables(disj::DisjunctiveConstraintData)
     vars = Set{JuMP.VariableRef}()
     for d in disj.constraint.disjuncts
-        union!(vars, _get_variables(d))
+        _interrogate_variables(v -> push!(vars, v), d)
     end
     return vars
 end
-function _get_variables(d::Disjunct)
-    vars = Set{JuMP.VariableRef}()
+# function _get_variables(disj::DisjunctiveConstraintData)
+#     vars = Set{JuMP.VariableRef}()
+#     for d in disj.constraint.disjuncts
+#         union!(vars, _get_variables(d))
+#     end
+#     return vars
+# end
+# function _get_variables(d::Disjunct)
+#     vars = Set{JuMP.VariableRef}()
+#     for con in d.constraints
+#         union!(vars, _get_variables(con))
+#     end
+#     return vars
+# end
+# function _get_variables(con::JuMP.AbstractArray{T}) where {T <: JuMP.AbstractConstraint}
+#     vars = Set{JuMP.VariableRef}()
+#     for c in con
+#         union!(vars, _get_variables(c))
+#     end
+#     return vars
+# end
+# function _get_variables(con::JuMP.ScalarConstraint)
+#     _get_variables(con.func)
+# end
+# function _get_variables(expr::JuMP.AffExpr)
+#     return keys(expr.terms)
+# end
+# function _get_variables(expr::JuMP.QuadExpr)
+#     vars = Set(keys(expr.aff.terms))
+#     for (pair, _) in expr.terms
+#         push!(vars, pair.a, pair.b)
+#     end
+#     return vars
+# end
+# function _get_variables(expr::JuMP.NonlinearExpr)
+#     vars = Set{JuMP.VariableRef}()
+#     _interrogate_variables(v -> push!(vars, v), expr)
+#     return vars
+# end
+# function _get_variables(expr)
+#     error("`_all_function_variables` not defined for expression of type $(typeof(expr)).")
+# end
+
+################################################################################
+#                              VARIABLE ITERATION
+################################################################################
+## Create helper methods to interrogate the variables of an expr w/ a function
+# Constant
+function _interrogate_variables(interrogator::Function, c::Number)
+    return
+end
+
+# VariableRef
+function _interrogate_variables(interrogator::Function, var::JuMP.VariableRef)
+    interrogator(var)
+    return
+end
+
+# AffExpr
+function _interrogate_variables(interrogator::Function, aff::JuMP.AffExpr)
+    for (var, _) in aff.terms
+        interrogator(var)
+    end
+    return
+end
+
+# QuadExpr
+function _interrogate_variables(interrogator::Function, quad::JuMP.QuadExpr)
+    for (pair, _) in quad.terms
+        interrogator(pair.a)
+        interrogator(pair.b)
+    end
+    _interrogate_variables(interrogator, quad.aff)
+    return
+end
+
+# NonlinearExpr
+function _interrogate_variables(interrogator::Function, nlp::JuMP.NonlinearExpr)
+    for arg in nlp.args
+        _interrogate_variables(interrogator, arg)
+    end
+    # See InfiniteOpt.jl for alternate method that avoids stackoverflow errors with deeply nested expressions:
+    # https://github.com/infiniteopt/InfiniteOpt.jl/blob/cb6dd6ae40fe0144b1dd75da0739ea6e305d5357/src/expressions.jl#L520-L534
+    return
+end
+
+# Constraint
+function _interrogate_variables(interrogator::Function, con::JuMP.ScalarConstraint)
[…]

[…]
