### 2023-07-18 4f901d9
first pass on hull reform for nlp
