+++ examples/ex3.jl
@@ -46,7 +46,10 @@ print(m_bigm)
 #  (exp(x) + -1000000000 Y[2] + 1000000000) ≥ 3
 
 ##
-
+m_hull = copy(m)
+DisjunctiveProgramming._reformulate(m_hull, Hull())
+print(m_hull)
+##
 @disjunction(
     m,
     begin
@@ -87,3 +90,11 @@ print(m)
 #  (1.0000000000000002e-6 + 2.999999 * z[2]) + (-1.0e-6 + -0.999999 * z[2]) * exp(x_z2 / (1.0e-6 + 0.999999 * z[2])) <= 0
 #  -1.0 * x_z1 + -3.0 * z[1] <= 0
 #  -1.0 * x_z2 + 5.0 * z[2] <= 0
+
+(
+    (0.999999 Y[1] + 1.0e-6) * (
+        (0.999999 Y[1] + 1.0e-6) * exp((x_Y[1] / (0.999999 Y[1] + 1.0e-6))) - 2.0
+    )
+)
+-
+-2.0e-6 Y[1] + 2.0e-6
