### 2023-07-18 8ff3e83
first pass on nlp hull reform
