+++ src/reformulate.jl
@@ -93,10 +93,7 @@ function _reformulate(
         M = method.value
     end
     JuMP.add_constraint(model,
-        JuMP.build_constraint(error, 
-            con.func - M*(1-bvar), 
-            con.set
-        ),
+        JuMP.build_constraint(error, con.func - M*(1-bvar), con.set)
     )
 end
 function _reformulate(
@@ -111,10 +108,7 @@ function _reformulate(
         M = method.value
     end
     JuMP.add_constraint(model,
-        JuMP.build_constraint(error, 
-            con.func + M*(1-bvar), 
-            con.set
-        ),
+        JuMP.build_constraint(error, con.func + M*(1-bvar), con.set)
     )
 end
 function _reformulate(
@@ -132,16 +126,10 @@ function _reformulate(
         M[2] = method.value
     end
     JuMP.add_constraint(model,
-        JuMP.build_constraint(error, 
-            con.func + first(M)*(1-bvar), 
-            _MOI.GreaterThan(con.set.lower)
-        ),
+        JuMP.build_constraint(error, con.func + first(M)*(1-bvar), _MOI.GreaterThan(con.set.lower))
     )
     JuMP.add_constraint(model,
-        JuMP.build_constraint(error, 
-            con.func - last(M)*(1-bvar), 
-            _MOI.LessThan(con.set.upper)
-        ),
+        JuMP.build_constraint(error, con.func - last(M)*(1-bvar), _MOI.LessThan(con.set.upper))
     )
 end
 function _reformulate(
@@ -159,16 +147,10 @@ function _reformulate(
         M[2] = method.value
     end
     JuMP.add_constraint(model,
-        JuMP.build_constraint(error, 
-            con.func + first(M)*(1-bvar), 
-            _MOI.GreaterThan(con.set.value)
-        ),
+        JuMP.build_constraint(error, con.func + first(M)*(1-bvar), _MOI.GreaterThan(con.set.value))
     )
     JuMP.add_constraint(model,
-        JuMP.build_constraint(error, 
-            con.func - last(M)*(1-bvar), 
-            _MOI.LessThan(con.set.value)
-        ),
+        JuMP.build_constraint(error, con.func - last(M)*(1-bvar), _MOI.LessThan(con.set.value))
     )
 end
 
@@ -177,78 +159,91 @@ end
 """
 function _reformulate(
     model::JuMP.Model, 
-    ::Hull,
-    con::JuMP.ScalarConstraint{JuMP.AffExpr, S}, 
+    method::Hull,
+    con::JuMP.ScalarConstraint{T, S}, 
     bvar::JuMP.VariableRef
-) where {S <: _MOI.LessThan}
+) where {T <: Union{JuMP.AffExpr, JuMP.QuadExpr}, S <: _MOI.LessThan}
     #TODO: need to pass _error to build_constraint
-    con_func = _disaggregated_constraint(model, con, bvar)
-    con_func.terms[bvar] = -con.set.upper
+    con_func = _disaggregate_expression(model, con.func, bvar, method)
+    con_func -= con.set.upper * bvar
 
     JuMP.add_constraint(model,
-        JuMP.build_constraint(error,
-            con_func,
-            _MOI.LessThan(0)
-        )
+        JuMP.build_constraint(error, con_func, _MOI.LessThan(0))
     )
 end
 function _reformulate(
     model::JuMP.Model, 
-    ::Hull,
-    con::JuMP.ScalarConstraint{JuMP.AffExpr, S}, 
+    method::Hull,
[…]
