### 2023-07-19 e6e12dd
hull reformulation for nonlinear models added
