### 2023-07-19 17d124b
add selector logical constrain reformulations
