+++ src/reformulate.jl
@@ -1,33 +1,37 @@
 """
 
+Create binary (indicator) variables for logic variables.
 """
-function _reformulate(model::JuMP.Model, method::AbstractReformulationMethod)
-    for (_, disj) in disjunctions(model)
-        _reformulate(model, method, disj)
-    end
-end
-
-"""
-
-"""
-function _reformulate(model::JuMP.Model, method::Union{BigM,Indicator}, disj::DisjunctiveConstraintData)
+function _reformulate_logical_variables(model::JuMP.Model)
     ind_var_dict = gdp_data(model).indicator_variables
-    for d in disj.constraint.disjuncts
-        #create binary variable for logic variable (indicator)
+    for (idx, _) in logical_variables(model)
+        lvar = LogicalVariableRef(model, idx)
         bvar = JuMP.@variable(model, 
-            base_name = string(d.indicator), 
+            base_name = string(lvar), 
             binary = true,
         )
-        ind_var_dict[Symbol(d.indicator, "_Bin")] = bvar
-        #reformualte disjunct
-        _reformulate(model, method, d, bvar)
+        ind_var_dict[lvar] = bvar
     end
 end
 
 """
 
 """
-function _reformulate(model::JuMP.Model, method::Hull, disj::DisjunctiveConstraintData)
+function _reformulate_disjunctive_constraints(model::JuMP.Model, method::AbstractReformulationMethod)
+    for (_, disj) in disjunctive_constraints(model)
+        _reformulate_disjunctive_constraints(model, method, disj)
+    end
+end
+
+function _reformulate_disjunctive_constraints(model::JuMP.Model, method::Union{BigM,Indicator}, disj::DisjunctiveConstraintData)
+    @show ind_var_dict = gdp_data(model).indicator_variables
+    for d in disj.constraint.disjuncts
+        bvar = ind_var_dict[d.indicator]
+        _reformulate_disjunctive_constraints(model, method, d, bvar)
+    end
+end
+
+function _reformulate_disjunctive_constraints(model::JuMP.Model, method::Hull, disj::DisjunctiveConstraintData)
     ind_var_dict = gdp_data(model).indicator_variables
     var_bounds_dict = gdp_data(model).variable_bounds
     disj_vars = _get_disjunction_variables(disj)
@@ -35,12 +39,7 @@ function _reformulate(model::JuMP.Model, method::Hull, disj::DisjunctiveConstrai
     _update_variable_bounds!(var_bounds_dict, disj_vars) #update variable bounds dict
     #reformulate each disjunct
     for d in disj.constraint.disjuncts
-        #create binary variable for logic variable (indicator)
-        bvar = JuMP.@variable(model, 
-            base_name = string(d.indicator), 
-            binary = true,
-        )
-        ind_var_dict[Symbol(d.indicator,"_Bin")] = bvar
+        bvar = ind_var_dict[d.indicator]
         #create disaggregated variables for that disjunct
         for var in disj_vars
             JuMP.is_binary(var) && continue #skip binary variables
@@ -50,7 +49,7 @@ function _reformulate(model::JuMP.Model, method::Hull, disj::DisjunctiveConstrai
             JuMP.add_to_expression!(sum_disag_vars[var], 1, disag_var)
         end
         #reformulate disjunct
-        _reformulate(model, method, d, bvar)
+        _reformulate_disjunctive_constraints(model, method, d, bvar)
     end
     #create sum constraint for disaggregated variables
[…]
