+++ src/DisjunctiveProgramming.jl
@@ -15,7 +15,7 @@ include("copy.jl")
 include("model.jl")
 include("variables.jl")
 include("constraints.jl")
-include("propositions.jl")
+include("logic.jl")
 include("reformulate.jl")
 include("bigm.jl")
 include("hull.jl")
