+++ src/model.jl
@@ -32,26 +32,41 @@ end
 """
 
 """
-function disjunctions(model::JuMP.Model)
+function disjunctive_constraints(model::JuMP.Model)
     is_gdp_model(model) || error("Cannot access disjunctions from a regular `JuMP.Model`.")
-    return model.ext[:GDP].disjunctions
+    return model.ext[:GDP].disjunctive_constraints
+end
+
+"""
+
+"""
+function logical_constraints(model::JuMP.Model)
+    is_gdp_model(model) || error("Cannot access logical constraints from a regular `JuMP.Model`.")
+    return model.ext[:GDP].logical_constraints
+end
+
+"""
+
+"""
+function logical_variables(model::JuMP.Model)
+    is_gdp_model(model) || error("Cannot access logical variables from a regular `JuMP.Model`.")
+    return model.ext[:GDP].logical_variables
 end
 
 function JuMP.copy_extension_data(data::GDPData, new_model::JuMP.AbstractModel, model::JuMP.AbstractModel)
-    model_data = gdp_data(model)
     #TODO need to copy disjunction constraints over to the new model
     #   (maybe something similar to what is done in copy_to in MOI.jl/src/Utilities/copy.jl).
     #   Otherwise, the constraints inside the disjuncts are not owned by the new model...
 
     new_model.ext[:GDP] = GDPData(
         data.logical_variables, 
-        _copy_disjunctions(data.disjunctions, new_model), 
-        data.propositions, 
+        _copy(data.logical_constraints, new_model), 
+        _copy(data.disjunctive_constraints, new_model), 
         data.solution_method, 
         data.ready_to_optimize,
-        data.disaggregated_variables,
-        data.indicator_variables,
-        data.variable_bounds
+        _copy(data.disaggregated_variables, new_model), 
+        _copy(data.indicator_variables, new_model), 
+        _copy(data.variable_bounds, new_model)
     )
 end
 
