### 2023-07-19 3ebf603
add CNF functions
