+++ src/logic.jl
@@ -1,40 +1,142 @@
+"""
+    _eliminate_equivalence!(expr)
 
+Eliminate equivalence logical operator.
+"""
+function _eliminate_equivalence(lvar::LogicalVariableRef)
+    return lvar
+end
+function _eliminate_equivalence(lexpr::LogicalExpr)
+    if lexpr.head == :⇔
+        if length(lexpr.args) != 2 
+            error("The equivalence operator must have two clauses.")
+        end
+        A = _eliminate_equivalence(lexpr.args[1])
+        B = _eliminate_equivalence(lexpr.args[2])
+        new_lexpr = LogicalExpr(:∧, Any[
+            LogicalExpr(:⇒, Any[A, B]),
+            LogicalExpr(:⇒, Any[B, A])
+        ])
+    else
+        new_lexpr = lexpr
+    end
 
+    return new_lexpr
+end
 
+"""
+    eliminate_implication!(expr)
 
+Eliminate implication logical operator.
+"""
+function _eliminate_implication(lvar::LogicalVariableRef)
+    return lvar
+end
+function _eliminate_implication(lexpr::LogicalExpr)
+    if lexpr.head == :⇒
+        if length(lexpr.args) != 2 
+            error("The implication operator must have two clauses.")
+        end
+        A = _eliminate_implication(lexpr.args[1])
+        B = _eliminate_implication(lexpr.args[2])
+        new_lexpr = LogicalExpr(:∨, Any[
+            LogicalExpr(:¬, Any[A]),
+            B
+        ])
+    else
+        new_lexpr = lexpr
+    end
 
+    return new_lexpr
+end
 
+"""
+    _move_negations_inwards(expr)
 
-# """
-#     choose!(m::JuMP.Model, n::Union{Int,LogicalVariable}, vars::LogicalVariable...; mode::Symbol=:exactly, name::String="")
+Move negation inwards in logical proposition expression.
+"""
+function _move_negations_inward(lvar::LogicalVariableRef)
+    return lvar
+end
+function _move_negations_inward(lexpr::LogicalExpr)
+    if lexpr.head == :¬
+        if length(lexpr.args) != 1
+            error("The negation operator can only have 1 clause.")
+        end
+        new_lexpr = _negate(lexpr.args[1])
+    else
+        new_lexpr = lexpr
+    end
 
-# Add constraint to select n elements from the list of variables. Options for mode
-# are `:at_least`, `:at_most`, `:exactly`. Alternately, if `n` is a Binary variable, 
-# it becomes the RHS of the constraint. If `name` is provided, it will be converted 
-# to a Symbol and stored in the object dictionary.
-# """
-# function choose!(m::JuMP.Model, n::Int, vars::LogicalVariable...; mode::Symbol=:exactly, name::String="")
-#     @assert length(vars) >= n "Not enough variables passed."
-#     @assert all(is_valid.(m, vars)) "Invalid VariableRefs passed."
-#     add_selection!(m, n, vars...; mode, name)
-# end
-# function choose!(m::JuMP.Model, var::LogicalVariable, vars::LogicalVariable...; mode::Symbol=:exactly, name::String="")
-#     @assert all(is_valid.(m, vcat(var,vars...))) "Invalid VariableRefs passed."
-#     add_selection!(m, var, vars...; mode, name)
-# end
-# function add_selection!(m::JuMP.Model, n, vars::LogicalVariable...; mode::Symbol, name::String)
-#     if mode == :exactly
-#         con = @constraint(m, sum(vars) == n)
-#     elseif mode == :at_least
-#         con = @constraint(m, sum(vars) ≥ n)
-#     elseif mode == :at_most
[…]
