### 2023-07-20 82c0397
update logic transformation; add _flatten
