+++ src/logic.jl
@@ -18,7 +18,9 @@ function _eliminate_equivalence(lexpr::LogicalExpr)
             LogicalExpr(:⇒, Any[B, A])
         ])
     else
-        new_lexpr = lexpr
+        new_lexpr = LogicalExpr(lexpr.head, Any[
+            _eliminate_equivalence(arg) for arg in lexpr.args
+        ])
     end
 
     return new_lexpr
@@ -44,7 +46,9 @@ function _eliminate_implication(lexpr::LogicalExpr)
             B
         ])
     else
-        new_lexpr = lexpr
+        new_lexpr = LogicalExpr(lexpr.head, Any[
+            _eliminate_implication(arg) for arg in lexpr.args
+        ])
     end
 
     return new_lexpr
@@ -65,7 +69,9 @@ function _move_negations_inward(lexpr::LogicalExpr)
         end
         new_lexpr = _negate(lexpr.args[1])
     else
-        new_lexpr = lexpr
+        new_lexpr = LogicalExpr(lexpr.head, Any[
+            _move_negations_inward(arg) for arg in lexpr.args
+        ])
     end
 
     return new_lexpr
@@ -97,14 +103,12 @@ end
 Negate OR boolean operator.
 """
 function _negate_or(lexpr::LogicalExpr)
-    if length(lexpr.args) != 2 
-        error("The OR operator must have two clauses.")
+    if length(lexpr.args) < 2 
+        error("The OR operator must have at least two clauses.")
     end
-    A = lexpr.args[1]
-    B = lexpr.args[2]
     return LogicalExpr(:∧, Any[ #flip OR to AND
-        _move_negations_inward(LogicalExpr(:¬, Any[A])),
-        _move_negations_inward(LogicalExpr(:¬, Any[B]))
+        _move_negations_inward(LogicalExpr(:¬, Any[arg]))
+        for arg in lexpr.args
     ])
 end
 
@@ -114,14 +118,12 @@ end
 Negate AND boolean operator.
 """
 function _negate_and(lexpr::LogicalExpr)
-    if length(lexpr.args) != 2 
-        error("The AND operator must have two clauses.")
+    if length(lexpr.args) < 2 
+        error("The AND operator must have at least two clauses.")
     end
-    A = lexpr.args[1]
-    B = lexpr.args[2]
     return LogicalExpr(:∨, Any[ #flip AND to OR
-        _move_negations_inward(LogicalExpr(:¬, Any[A])),
-        _move_negations_inward(LogicalExpr(:¬, Any[B]))
+        _move_negations_inward(LogicalExpr(:¬, Any[arg]))
+        for arg in lexpr.args
     ])
 end
 
@@ -137,6 +139,130 @@ function _negate_negation(lexpr::LogicalExpr)
     return _move_negations_inward(lexpr.args[1])
 end
 
+
+"""
+    _distribute_and_over_or(expr)
+
+Distribute AND over OR boolean operators.
+"""
+function _distribute_and_over_or(lvar::LogicalVariableRef)
+    return lvar
+end
+function _distribute_and_over_or(lexpr::LogicalExpr)
+    if lexpr.head == :∨
+        if length(lexpr.args) < 2 
+            error("The OR operator must have at least two clauses.")
+        end
+        new_lexpr = _distribute_and_over_or_left(lexpr)
+        new_lexpr = _distribute_and_over_or_right(new_lexpr)
+    else
+        new_lexpr = LogicalExpr(lexpr.head, Any[
+            _distribute_and_over_or(arg) for arg in lexpr.args
+        ])
+    end
+
+    return new_lexpr
+end
+function _distribute_and_over_or_left(lexpr::LogicalExpr)
[…]
