### 2023-07-20 a4e1e49
fix distribute AND over OR
