### 2023-07-20 25d700f
remove JuMP macros; add reformulate proposition
