### 2023-07-20 f58a242
wrapper for VariableInfo
