### 2023-07-20 da2fef5
complete proposition transformation functions
