+++ src/logic.jl
@@ -2,6 +2,9 @@
 
 """
 function _to_cnf(lexpr::LogicalExpr)
+    #NOTE: some redundant constraints may be created in the process.
+    #   For example A ∨ ¬B ∨ B is always true and is reformulated 
+    #   to the redundant constraint A ≥ 0.
     lexpr |> 
         _eliminate_equivalence |>  
         _eliminate_implication |> 
@@ -223,7 +226,10 @@ function _distribute_and_over_or_right(lexpr::LogicalExpr)
 end
 
 """
+    _flatten(lexpr::LogicalRef)
 
+Flatten netsed OR / AND operators and replace them with their n-ary form.
+For example, ∨(∨(A, B), C) is replaced with ∨(A, B, C).
 """
 function _flatten(lvar::LogicalVariableRef)
     return lvar
@@ -234,7 +240,7 @@ function _flatten(lexpr::LogicalExpr)
         for arg in lexpr.args
             if arg isa LogicalVariableRef
                 push!(nary_args, arg)
-            elseif arg.head == :¬
+            elseif arg.head == :¬ && arg.args[1] isa LogicalVariableRef
                 push!(nary_args, arg)
             elseif arg.head == lexpr.head
                 arg_flat = _flatten(arg)
@@ -254,184 +260,3 @@ function _flatten(lexpr::LogicalExpr)
 
     return new_lexpr
 end
-# function JuMP.flatten(expr::LogicalExpr)
-#     root = LogicalExpr(expr.head, Any[])
-#     nodes_to_visit = Any[(root, arg) for arg in reverse(expr.args)]
-#     while !isempty(nodes_to_visit)
-#         parent, arg = pop!(nodes_to_visit)
-#         if !(arg isa LogicalExpr)
-#             # Not a nonlinear expression, so can use recursion.
-#             push!(parent.args, JuMP.flatten(arg))
-#         elseif parent.head in (:∨, :∧) && arg.head == parent.head
-#             # A special case: the arg can be lifted to an n-ary argument of the
-#             # parent.
-#             for n in reverse(arg.args)
-#                 push!(nodes_to_visit, (parent, n))
-#             end
-#         else
-#             # The default case for nonlinear expressions. Put the args on the
-#             # stack, so that we may walk them later.
-#             for n in reverse(arg.args)
-#                 push!(nodes_to_visit, (arg, n))
-#             end
-#             empty!(arg.args)
-#             push!(parent.args, arg)
-#         end
-#     end
-#     return root
-# end
-
-# """
-#     add_proposition!(m::JuMP.Model, expr::Expr; name::String = "")
-
-# Convert logical proposition expression into conjunctive normal form.
-# """
-# function add_proposition!(m::JuMP.Model, expr::Expr; name::String = "")
-#     if isempty(name)
-#         name = "{$expr}" #get name to register reformulated logical proposition
-#     end
-#     replace_Symvars!(expr, m; logical_proposition = true) #replace all JuMP variables with Symbolic variables
-#     clause_list = to_cnf!(expr)
-#     #replace symbolic variables with JuMP variables and boolean operators with their algebraic counterparts
-#     for clause in clause_list
-#         replace_JuMPvars!(clause, m)
-#         replace_logic_operators!(clause)
-#     end
-#     #generate and simplify JuMP expressions for the lhs of the algebraic constraints
-#     lhs = eval.(clause_list)
-#     drop_zeros!.(lhs)
-#     unique!(lhs)
-#     #generate JuMP constraints for the logical proposition
-#     if length(lhs) == 1
[…]
