### 2023-07-20 d0c5ddd
define and use logical operators to build LogicExpr
