+++ examples/ex4.jl
@@ -1,3 +1,6 @@
+using JuMP
+using DisjunctiveProgramming
+
 # Example with proposition reformulation
 # Proposition:
 # ¬((Y[1] ∧ ¬Y[2]) ⇔ (Y[3] ∨ Y[4]))
@@ -5,16 +8,7 @@
 m = GDPModel()
 @variable(m, Y[1:4], LogicalVariable)
 logic_1 = LogicalConstraint(
-    LogicalExpr(:¬, Any[
-        LogicalExpr(:⇔, Any[
-                LogicalExpr(:∧, Any[
-                    Y[1], 
-                    LogicalExpr(:¬, Any[Y[2]])
-                ]),
-                LogicalExpr(:∨, Any[Y[3], Y[4]])
-            ]
-        )
-    ])
+    ¬((Y[1] ∧ ¬Y[2]) ⇔ (Y[3] ∨ Y[4]))
 )
 selector = add_constraint(m,
     logic_1,
