### 2023-08-01 7f44c7f
first implementation using @constraint macro for disjunct constraints and logic constraints
