+++ src/variables.jl
@@ -57,7 +57,7 @@ function JuMP.add_variable(
     )
     is_gdp_model(model) || error("Can only add logical variables to `GDPModel`s.")
     data = LogicalVariableData(v, name)
-    idx = _MOIUC.add_item(gdp_data(model).logical_variables, data)
+    idx = _MOIUC.add_item(_logical_variables(model), data)
     return LogicalVariableRef(model, idx)
 end
 
