+++ src/constraints.jl
@@ -1,29 +1,54 @@
 """
 
 """
+# function JuMP.build_constraint(_error::Function, func::_MOI.AbstractVectorFunction, set::_MOI.AbstractVectorSet, tag::Type{DisjunctConstraint})
+#     # TODO add error checking
+#     # NOTE: assumes function is of the form MOI.ACTIVATE_ON_ONE (not negating the Logical Variable)
+#     lvar = func[1].terms |> keys |> first
+#     con = func[2]
+
+#     return DisjunctConstraint(func, set.set, lvar)
+# end
+function JuMP._build_indicator_constraint(
+    _error::Function,
+    lvar::LogicalVariableRef,
+    con::JuMP.ScalarConstraint,
+    ::Type{_MOI.Indicator{A}},
+) where {A}
+    return DisjunctConstraint(con.func, con.set, lvar)
+end
+
+function JuMP.build_constraint(_error::Function, func::JuMP.AbstractJuMPScalar, set::_MOI.AbstractScalarSet, lvar::LogicalVariableRef)
+    return DisjunctConstraint(func, set, lvar)
+end
+function JuMP.build_constraint(_error::Function, func::JuMP.VariableRef, lb::Number, ub::Number, lvar::LogicalVariableRef)
+    #for the case lb <= x <= ub
+    return DisjunctConstraint(1*func, _MOI.Interval{Float64}(lb,ub), lvar)
+end
+
+function JuMP.build_constraint(_error::Function, func::JuMP.AbstractJuMPScalar, set::_MOI.AbstractScalarSet, tag::Type{DisjunctConstraint})
+    return DisjunctConstraint(func, set, nothing)
+end
+function JuMP.build_constraint(_error::Function, func::JuMP.VariableRef, lb::Number, ub::Number, tag::Type{DisjunctConstraint})
+    #for the case lb <= x <= ub
+    return DisjunctConstraint(1*func, _MOI.Interval{Float64}(lb,ub), nothing)
+end
+
 function JuMP.build_constraint(_error::Function, disjuncts::Vector{<:Disjunct})
     # TODO add error checking
-    return DisjunctiveConstraint(disjuncts)
+    return Disjunction(disjuncts)
 end
 
-function JuMP.build_constraint(_error::Function, logical_expression::LogicalExpr)
+# function JuMP.parse_constraint_call(_error::Function, is_vectorized::Bool, ::Val{:exactly}, args...)
+#     parse_code = :()
+#     build_code = :(JuMP.build_constraint($(_error), $exactly($(args...)), LogicalConstraint))
+#     return parse_code, build_code 
+# end
+
+function JuMP.build_constraint(_error::Function, func::JuMP.AbstractJuMPScalar, set::_MOI.AbstractScalarSet, tag::Type{LogicalConstraint})
     # TODO add error checking
-    return JuMP.build_constraint(_error, logical_expression, _MOI.EqualTo(true))
-end
-
-"""
-
-"""
-function JuMP.add_constraint(
-    model::JuMP.Model, 
-    c::DisjunctiveConstraint, 
-    name::String = ""
-    )
-    is_gdp_model(model) || error("Can only add disjunctions to `GDPModel`s.")
-    # TODO maybe check the variables in the disjuncts belong to the model
-    constr_data = DisjunctiveConstraintData(c, name)
-    idx = _MOIUC.add_item(gdp_data(model).disjunctive_constraints, constr_data)
-    return DisjunctiveConstraintRef(model, idx)
+    # NOTE: excatly(1, Y[1], Y[2]) == true creates a constraint of the form LogicalExpr - 1.0 in MOI.EqualTo(0)
+    return LogicalConstraint(func.args[1], _MOI.EqualTo(true))
 end
 
 function JuMP.add_constraint(
[…]
