### 2023-08-01 2781545
use parse_constraint_call for logical constraints
