+++ examples/ex4.jl
@@ -7,7 +7,7 @@ using DisjunctiveProgramming
 
 m = GDPModel()
 @variable(m, Y[1:4], LogicalVariable)
-@constraint(m, logic, ¬((Y[1] ∧ ¬Y[2]) ⇔ (Y[3] ∨ Y[4])) == true, LogicalConstraint)
+@constraint(m, logic, ¬((Y[1] ∧ ¬Y[2]) ⇔ (Y[3] ∨ Y[4])))
 DisjunctiveProgramming._reformulate_logical_variables(m)
 DisjunctiveProgramming._reformulate_logical_constraints(m)
 print(m)
