### 2023-08-01 0f29b5f
use func in set for logical constraints
