### 2023-08-01 6eef5d4
allow creating a constraint with exactly, atleast, atmost functions and parse to func in set notation
