### 2023-08-01 73d759b
bug fix to allow CP set to have a LogicalVariableRef
