+++ src/variables.jl
@@ -1,3 +1,6 @@
+################################################################################
+#                              LOGICAL VARIABLES
+################################################################################
 """
     _variable_info(binary::Bool=false, lower_bound::Float64=-Inf, upper_bound::Float64=Inf)
 
