+++ src/datatypes.jl
@@ -1,36 +1,6 @@
-"""
-
-"""
-# Logical sets
-struct MOIIsTrue <: _MOI.AbstractScalarSet end
-struct IsTrue <: JuMP.AbstractScalarSet end
-JuMP.moi_set(set::IsTrue) = MOIIsTrue()
-
-abstract type MOIFirstOrderSet <: _MOI.AbstractVectorSet end
-struct MOIAtLeast{T} <: MOIFirstOrderSet
-    value::T
-    dimension::Int
-end
-struct MOIAtMost{T} <: MOIFirstOrderSet
-    value::T
-    dimension::Int
-end
-struct MOIExactly{T} <: MOIFirstOrderSet
-    value::T
-    dimension::Int
-end
-struct AtLeast{T} <: JuMP.AbstractVectorSet
-    value::T
-end
-struct AtMost{T} <: JuMP.AbstractVectorSet
-    value::T
-end
-struct Exactly{T} <: JuMP.AbstractVectorSet
-    value::T
-end
-JuMP.moi_set(set::AtLeast, dim::Int) = MOIAtLeast(set.value, dim)
-JuMP.moi_set(set::AtMost, dim::Int) = MOIAtMost(set.value, dim)
-JuMP.moi_set(set::Exactly, dim::Int) = MOIExactly(set.value, dim)
+################################################################################
+#                              LOGICAL VARIABLES
+################################################################################
 
 """
     LogicalVariable <: JuMP.AbstractVariable
@@ -84,7 +54,99 @@ struct LogicalVariableRef <: JuMP.AbstractVariableRef
     index::LogicalVariableIndex
 end
 
-#######################################################################
+################################################################################
+#                              LOGICAL CONSTRAINTS
+################################################################################
+"""
+
+"""
+# Logical sets
+struct MOIIsTrue <: _MOI.AbstractScalarSet end
+struct IsTrue <: JuMP.AbstractScalarSet end
+JuMP.moi_set(set::IsTrue) = MOIIsTrue()
+
+abstract type MOIFirstOrderSet <: _MOI.AbstractVectorSet end
+struct MOIAtLeast{T} <: MOIFirstOrderSet
+    value::T
+    dimension::Int
+end
+struct MOIAtMost{T} <: MOIFirstOrderSet
+    value::T
+    dimension::Int
+end
+struct MOIExactly{T} <: MOIFirstOrderSet
+    value::T
+    dimension::Int
+end
+struct AtLeast{T} <: JuMP.AbstractVectorSet
+    value::T
+end
+struct AtMost{T} <: JuMP.AbstractVectorSet
+    value::T
+end
+struct Exactly{T} <: JuMP.AbstractVectorSet
+    value::T
+end
+JuMP.moi_set(set::AtLeast, dim::Int) = MOIAtLeast(set.value, dim)
+JuMP.moi_set(set::AtMost, dim::Int) = MOIAtMost(set.value, dim)
+JuMP.moi_set(set::Exactly, dim::Int) = MOIExactly(set.value, dim)
+
+
+const LogicalExpr = JuMP.NonlinearExpr{LogicalVariableRef}
+
+"""
+    LogicalConstraint <: JuMP.AbstractConstraint
+
+A type for a logical constraint that is comprised of an expression on LogicalVariables 
+    with logic operators.
+
+**Fields**
+- `expression::LogicalExpr`: The logical constraint.
+"""
+struct LogicalConstraint{T,S} <: JuMP.AbstractConstraint
+    func::T
+    set::S
+end
+
+"""
+    LogicalConstraintData
+
+A type for storing [`LogicalConstraint`](@ref)s and any meta-data they possess.
+
+**Fields**
[…]
