+++ src/macros.jl
@@ -1,0 +1,243 @@
+################################################################################
+#                                BASIC HELPERS
+################################################################################
+# Macro error function
+# inspired from https://github.com/jump-dev/JuMP.jl/blob/709d41b78e56efb4f2c54414266b30932010bd5a/src/macros.jl#L923-L928
+function _macro_error(macroname, args, source, str...)
+    error("At $(source.file):$(source.line): `@$macroname($(join(args, ", ")))`: ", 
+          str...)
+end
+
+# Escape when needed
+# taken from https://github.com/jump-dev/JuMP.jl/blob/709d41b78e56efb4f2c54414266b30932010bd5a/src/macros.jl#L895-L897
+_esc_non_constant(x::Number) = x
+_esc_non_constant(x::Expr) = isexpr(x,:quote) ? x : esc(x)
+_esc_non_constant(x) = esc(x)
+
+# Extract the name from a macro expression 
+# Inspired from https://github.com/jump-dev/JuMP.jl/blob/45ce630b51fb1d72f1ff8fed35a887d84ef3edf7/src/Containers/macro.jl#L8-L17
+_get_name(c::Symbol) = c
+_get_name(c::Nothing) = ()
+_get_name(c::AbstractString) = c
+function _get_name(c::Expr)
+    if isexpr(c, :string)
+        return c
+    else
+        return c.args[1]
+    end
+end
+
+# Given a base_name and idxvars, returns an expression that constructs the name
+# of the object.
+# Inspired from https://github.com/jump-dev/JuMP.jl/blob/709d41b78e56efb4f2c54414266b30932010bd5a/src/macros.jl#L930-L946
+function _name_call(base_name, idxvars)
+    if isempty(idxvars) || base_name == ""
+        return base_name
+    end
+    ex = Expr(:call, :string, base_name, "[")
+    for i in eachindex(idxvars)
+        # Converting the arguments to strings before concatenating is faster:
+        # https://github.com/JuliaLang/julia/issues/29550.
+        esc_idxvar = esc(idxvars[i])
+        push!(ex.args, :(string($esc_idxvar)))
+        i < length(idxvars) && push!(ex.args, ",")
+    end
+    push!(ex.args, "]")
+    return ex
+end
+
+# Process macro arugments 
+function _extract_kwargs(args)
+    arg_list = collect(args)
+    if !isempty(args) && isexpr(args[1], :parameters)
+        p = popfirst!(arg_list)
+        append!(arg_list, p.args)
+    end
+    extra_kwargs = filter(x -> isexpr(x, :(=)) && x.args[1] != :container &&
+                          x.args[1] != :base_name, arg_list)
+    container_type = :Auto
+    base_name = nothing
+    for kw in arg_list
+        if isexpr(kw, :(=)) && kw.args[1] == :container
+            container_type = kw.args[2]
+        elseif isexpr(kw, :(=)) && kw.args[1] == :base_name
+            base_name = esc(kw.args[2])
+        end
+    end
+    pos_args = filter!(x -> !isexpr(x, :(=)), arg_list)
+    return pos_args, extra_kwargs, container_type, base_name
+end
+
+# Add on keyword arguments to a function call expression and escape the expressions
+# Adapted from https://github.com/jump-dev/JuMP.jl/blob/d9cd5fb16c2d0a7e1c06aa9941923492fc9a28b5/src/macros.jl#L11-L36
+function _add_kwargs(call, kwargs)
+    for kw in kwargs
+        push!(call.args, esc(Expr(:kw, kw.args...)))
+    end
+    return
+end
+
+# Add on positional args to a function call and escape
[…]
