+++ src/constraints.jl
@@ -1,216 +1,468 @@
-"""
+################################################################################
+#                         BOILERPLATE EXTENSION METHODS
+################################################################################
+for (RefType, loc) in ((:DisjunctConstraintRef, :disjunct_constraints), 
+                       (:DisjunctionRef, disjunctions), 
+                       (:LogicalConstraintRef, logical_constraints))
+    @eval begin
+        @doc """
+            JuMP.owner_model(cref::$($RefType))
 
-"""
-JuMP.owner_model(cref::DisjunctConstraintRef) = cref.model
+        Return the model associated with `cref`.
+        """
+        JuMP.owner_model(cref::$RefType) = cref.model
 
-"""
+        @doc """
+            JuMP.index(cref::$($RefType))
 
-"""
-JuMP.owner_model(cref::DisjunctionRef) = cref.model
+        Return the index associated with `cref`.
+        """
+        JuMP.index(cref::$RefType) = cref.index
 
-"""
+        @doc """
+            JuMP.is_valid(model::JuMP.Model, cref::$($RefType))
 
-"""
-JuMP.index(cref::DisjunctConstraintRef) = cref.index
+        Return `Bool` whether the reference is valid.
+        """
+        function JuMP.is_valid(model::JuMP.Model, cref::$RefType) # TODO: generalize for AbstractModel
+            return model === JuMP.owner_model(cref)
+        end
 
-"""
+        # Get the ConstraintData object
+        function _constraint_data(cref::$RefType)
+            return gdp_data(JuMP.owner_model(cref)).$loc[JuMP.index(cref)]
+        end
 
-"""
-JuMP.index(cref::DisjunctionRef) = cref.index
+        @doc """
+            JuMP.name(cref::$($RefType))
 
-"""
+        Return the name associated with `cref`.
+        """
+        function JuMP.name(cref::$RefType)
+            return _constraint_data(cref).name
+        end
 
-"""
-function JuMP.is_valid(model::JuMP.Model, cref::DisjunctionRef)
-    return model === JuMP.owner_model(cref)
+        @doc """
+            JuMP.set_name(cref::$($RefType), name::String)
+
+        Return the name associated with `cref`.
+        """
+        function JuMP.set_name(cref::$RefType, name::String)
+            _constraint_data(cref).name = name
+            _set_ready_to_optimize(JuMP.owner_model(cref), false)
+            return
+        end
+
+        @doc """
+            JuMP.constraint_object(cref::$($RefType))
+
+        Return the constraint object associated with `cref`.
+        """
+        function JuMP.constraint_object(cref::$RefType)
+            return _constraint_data(cref).constraint
+        end
+
+        # Extend comparisons
+        function Base.:(==)(cref1::$RefType, cref2::$RefType)
+            return cref1.model === cref2.model && cref1.index == cref2.index
+        end
+
+        # Extend copy
+        Base.copy(cref::$RefType) = cref
+        #NOTE: copying a GDPmodel gives an error for any named DisjunctConstraints since 
+        #   it doesn't know how to copy these.
+        #   However, they are copied in the GDPData ext Dict.
+    end
+end
+
+# TODO: extend JuMP.delete for all the constraint types
+
+# function JuMP.delete(model::JuMP.Model, cref::DisjunctionRef)
+#     @assert JuMP.is_valid(model, cref) "Disjunctive constraint does not belong to model."
+#     constr_data = gdp_data(JuMP.owner_model(cref))
+#     dict = constr_data.disjunctions[JuMP.index(cref)]
+#     # TODO check if used by a disjunction and/or a proposition
+#     delete!(dict, index(cref))
[…]
