+++ examples/ex1.jl
@@ -9,19 +9,19 @@ using DisjunctiveProgramming
 m = GDPModel()
 @variable(m, -5 ≤ x ≤ 10)
 @variable(m, Y[1:2], LogicalVariable)
-@constraint(m, disjunct_1_con, 0 ≤ x ≤ 3, Y[1])
-@constraint(m, disjunct_2_con_a, 5 ≤ x, Y[2])
-@constraint(m, disjunct_2_con_b, x ≤ 9, Y[2])
+@constraint(m, disjunct_1_con, 0 ≤ x ≤ 3, DisjunctConstraint(Y[1]))
+@constraint(m, disjunct_2_con_a, 5 ≤ x, DisjunctConstraint(Y[2]))
+@constraint(m, disjunct_2_con_b, x ≤ 9, DisjunctConstraint(Y[2]))
 @disjunction(m, disjunction, [Y[1], Y[2]])
 
-# Disjunction Method 2: Same as Method 1, but using Indicator Constraint notation
-m = GDPModel()
-@variable(m, -5 ≤ x ≤ 10)
-@variable(m, Y[1:2], LogicalVariable)
-@constraint(m, disjunct_1_con, Y[1] => {0 ≤ x ≤ 3})
-@constraint(m, disjunct_2_con_a, Y[2] => {5 ≤ x})
-@constraint(m, disjunct_2_con_b, Y[2] => {x ≤ 9})
-disjunction = add_disjunction(m, Y, "Disjunction")
+# Disjunction Method 2: Same as Method 1, but using Indicator Constraint notation --> not currently supported
+# m = GDPModel()
+# @variable(m, -5 ≤ x ≤ 10)
+# @variable(m, Y[1:2], LogicalVariable)
+# @constraint(m, disjunct_1_con, Y[1] => {0 ≤ x ≤ 3})
+# @constraint(m, disjunct_2_con_a, Y[2] => {5 ≤ x})
+# @constraint(m, disjunct_2_con_b, Y[2] => {x ≤ 9})
+# disjunction = add_disjunction(m, Y, "Disjunction")
 
 # Disjunction Method 3: Create Logical Variables from Disjunctions
 m = GDPModel()
@@ -29,17 +29,14 @@ m = GDPModel()
 @constraint(m, disjunct_1_con, 0 ≤ x ≤ 3, DisjunctConstraint)
 @constraint(m, disjunct_2_con_a, 5 ≤ x, DisjunctConstraint)
 @constraint(m, disjunct_2_con_b, x ≤ 9, DisjunctConstraint)
-disjunction = add_disjunction(m,
-    [[disjunct_1_con], [disjunct_2_con_a, disjunct_2_con_b]],
-    "Y"
-)
-Y = disjunction_indicators(disjunction)
+@disjunction(m, d2, [[disjunct_1_con], [disjunct_2_con_a, disjunct_2_con_b]])
+Y = disjunction_indicators(d2)
 
 # Logical Constraint Method 1: Use func in set notation
 @constraint(m, exclussinve, Y in Exactly(1))
 
-# Logical Constraint Method 2: Use NonlinearExpr that gets parsed to func in set notation
-@constraint(m, exclussive, exactly(1, Y))
+# Logical Constraint Method 2: Use NonlinearExpr that gets parsed to func in set notation --> not currently supported
+# @constraint(m, exclussive, exactly(1, Y))
 
 # Reformulate logical variables and logical constraints
 DisjunctiveProgramming._reformulate_logical_variables(m)
