### 2023-08-07 82aa16c
first pass at adding reformulations and copy to new structure
