### 2023-08-08 36da978
2nd pass: fix bugs. add parse methods for logical operators
