+++ examples/ex4.jl
@@ -8,9 +8,7 @@ using DisjunctiveProgramming
 m = GDPModel()
 @variable(m, Y[1:4], LogicalVariable)
 @constraint(m, ¬((Y[1] ∧ ¬Y[2]) ⇔ (Y[3] ∨ Y[4])) == true)
-@constraint(m, ¬((Y[1] ∧ ¬Y[2]) ⇔ (Y[3] ∨ Y[4])) in MOI.EqualTo(true))
 @constraint(m, (Y[1] ∧ ¬Y[2]) ⇔ (Y[3] ∨ Y[4]) == false)
-@constraint(m, (Y[1] ∧ ¬Y[2]) ⇔ (Y[3] ∨ Y[4]) in MOI.EqualTo(false))
 DisjunctiveProgramming._reformulate_logical_variables(m)
 DisjunctiveProgramming._reformulate_logical_constraints(m)
 print(m)
