### 2023-08-08 eadaa2e
add dispatch methods for non Bool sets in build_constraint
