### 2023-08-11 #117 @joaquimg
From a first pass:

> Installation: Does installation proceed as outlined in the documentation?

On Julia 1.9.0, no a clean enviroment:
```julia
julia> using Pkg

julia> Pkg.add("DisjunctiveProgramming")
```
Installs, but returns warnings:

```
        Info Packages marked with ⌅ have new versions available but compatibility constraints restrict them from upgrading. To see why use `status --outdated -m`
Precompiling project...
  45 dependencies successfully precompiled in 482 seconds. 94 already precompiled.
  3 dependencies had warnings during precompilation:
┌ Symbolics [0c5d862f-8b57-4792-8d23-62f2024744c7]
│  WARNING: Method definition isapprox(IntervalSets.AbstractInterval{T} where T, IntervalSets.AbstractInterval{T} where T) in module IntervalSets at C:\JG\Julia\packages\IntervalSets\6RTOk\src\IntervalSets.jl:144 overwritten in module DomainSets at C:\JG\Julia\packages\DomainSets\aafhp\src\domains\interval.jl:52.
│    ** incremental compilation may be fatally broken for this module **
│
│  WARNING: Method definition isapprox(IntervalSets.AbstractInterval{T} where T, IntervalSets.AbstractInterval{T} where T) in module IntervalSets at C:\JG\Julia\packages\IntervalSets\6RTOk\src\IntervalSets.jl:144 overwritten in module DomainSets at C:\JG\Julia\packages\DomainSets\aafhp\src\domains\interval.jl:52.
│    ** incremental compilation may be fatally broken for this module **
└
┌ DisjunctiveProgramming [0d27d021-0159-4c7d-b4a7-9ccb5d9366cf]
│  WARNING: Method definition isapprox(IntervalSets.AbstractInterval{T} where T, IntervalSets.AbstractInterval{T} where T) in module IntervalSets at C:\JG\Julia\packages\IntervalSets\6RTOk\src\IntervalSets.jl:144 overwritten in module DomainSets at C:\JG\Julia\packages\DomainSets\aafhp\src\domains\interval.jl:52.
│    ** incremental compilation may be fatally broken for this module **
└
┌ DomainSets [5b8099bc-c8ec-5219-889f-1d9e522a28bf]
│  WARNING: Method definition isapprox(IntervalSets.AbstractInterval{T} where T, IntervalSets.AbstractInterval{T} where T) in module IntervalSets at C:\JG\Julia\packages\IntervalSets\6RTOk\src\IntervalSets.jl:144 overwritten in module DomainSets at C:\JG\Julia\packages\DomainSets\aafhp\src\domains\interval.jl:52.
│    ** incremental compilation may be fatally broken for this module **
└
```

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

The example in the README could be end-to-end and contain a solver, a call to `optimize` and some expected results.

Also, the line `@proposition(m, y[1] ∨ y[2])` is failing on windows.

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I could not find docs other than the README. There is a docs folder, but no link in the README.
Also, there are broken links in the README.

> Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

I could not find a CI.yml with the automated tests.

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

No.

> Authors: Does the paper.tex file include a list of authors with their affiliations?

Only the last author has an email. He is supposed to be the corresponding author?

> References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

Many references are missing DOI´s.
