### 2023-08-11 #117 @joaquimg
@matbesancon , about:
> Page limit: Is the page limit for full papers respected by the submitted document?

what is the page limit? I could not find it easily. Sorry, it is probably obvious.
