+++ src/logic.jl
@@ -4,6 +4,7 @@
 
 # Define all the logic functions/operators that use 2+ arguments
 for (name, func) in (
+    (:⇔, :⇔), (:⇔, :(<-->)), (:⇔, :iff), (:⇔, :equals), # \Leftrightarrow + tab
     (:∨, :∨), (:∨, :lor), # \vee + tab
     (:∧, :∧), (:∧, :land), # \wedge + tab
 )
@@ -21,7 +22,6 @@ end
 
 # Define all the logic functions/operators that use 2 arguments
 for (name, func) in (
-    (:⇔, :⇔), (:⇔, :(<-->)), (:⇔, :iff), (:⇔, :equals), # \Leftrightarrow + tab
     (:⇒, :⇒), (:⇒, :implies), # \Rightarrow + tab
     (:Ξ, :Ξ), (:Ξ, :exactly), # \Xi + tab
     (:Λ, :Λ), (:Λ, :atmost), # \Lambda + tab
@@ -76,12 +76,18 @@ function _eliminate_equivalence(lvar::LogicalVariableRef)
     return lvar
 end
 function _eliminate_equivalence(lexpr::_LogicalExpr)
-    if lexpr.head == :⇔
-        if length(lexpr.args) != 2 
-            error("The equivalence operator must have two clauses.")
-        end
+    if lexpr.head in (:⇔, :(<-->), :iff, :equals)
         A = _eliminate_equivalence(lexpr.args[1])
-        B = _eliminate_equivalence(lexpr.args[2])
+        if length(lexpr.args) > 2 
+            nested = _LogicalExpr(:⇔, Vector{Any}(lexpr.args[2:end]))
+            B = _eliminate_equivalence(nested)
+        elseif length(lexpr.args) == 2
+            B = _eliminate_equivalence(lexpr.args[2])
+        else
+            error("The equivalence logic operator `⇔` must have at least two arguments.")
+        end
+        
+        
         new_lexpr = _LogicalExpr(:∧, Any[
             _LogicalExpr(:⇒, Any[A, B]),
             _LogicalExpr(:⇒, Any[B, A])
@@ -104,7 +110,7 @@ function _eliminate_implication(lvar::LogicalVariableRef)
     return lvar
 end
 function _eliminate_implication(lexpr::_LogicalExpr)
-    if lexpr.head == :⇒
+    if lexpr.head in (:⇒, :implies)
         if length(lexpr.args) != 2 
             error("The implication operator must have two clauses.")
         end
@@ -132,7 +138,7 @@ function _move_negations_inward(lvar::LogicalVariableRef)
     return lvar
 end
 function _move_negations_inward(lexpr::_LogicalExpr)
-    if lexpr.head == :¬
+    if lexpr.head in (:¬, :lneg)
         if length(lexpr.args) != 1
             error("The negation operator can only have 1 clause.")
         end
@@ -153,11 +159,11 @@ function _negate(lvar::LogicalVariableRef)
     return _LogicalExpr(:¬, Any[lvar])
 end
 function _negate(lexpr::_LogicalExpr)
-    if lexpr.head == :∨
+    if lexpr.head in (:∨, :lor)
         new_lexpr = _negate_or(lexpr)
-    elseif lexpr.head == :∧
+    elseif lexpr.head in (:∧, :land)
         new_lexpr = _negate_and(lexpr)
-    elseif lexpr.head == :¬
+    elseif lexpr.head in (:¬, :lneg)
         new_lexpr = _negate_negation(lexpr)
     else
         #TODO: maybe catch error here if other operator is present
@@ -217,65 +223,29 @@ Distribute AND over OR boolean operators.
[…]
