### 2023-08-13 #117 @joaquimg
Also for @matbesancon ,
Per the reviewer guidelines (https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper) and this comment (https://github.com/JuliaCon/proceedings-review/issues/117#issuecomment-1345551688) a section named "Statement of need" seems necessary. Can you confirm? If so, @hdavid16 will have to make it explicit.
