### 2023-08-13 #117 @joaquimg
> Content

This seems to be the best place to comment on paper presentation.

I have a few main points:

1) Models in section 2.2 are confusing because they are not self-contained. They are simply groups of formulas that show up as needed, and the text dictates how to combine them. Although it is possible to understand, it could be more straightforward.

For instance, instead of having a paragraph:
"... this model is given by (13), (16)-(19)..." and only show the additional (16)-(19)
one could do:
"... this model is given by (16)-(20)..." and show (16)-(20) where (16) repeats (13) or adds it explicitly:
```
(13)                      (16)
new eq...                 (17)
new eq...                 (18)
new eq...                 (19)
new eq...                 (20)
```

2) In "Note: any negated Boolean variables, ¬Yi, are replaced with 1 − yi in the reformulation"
It is not clear if should be a term in the sum (this seems wrong) or if there would be and additional variable zi = 1 − yi (this looks good).

3) sections 2.4.2 and 2.4.3 seem to be not very related to the remainder of the paper.
They are not mentioned as future work. So they might be better as one sentence comment in a section the focuses more on  2.4.1 which is the relevant piece for the paper.

4) From paper and readme it is not clear what is the difference between `@disjunction` or `add_disjunction!`.
From the example, it is not clear if the chosen usage was necessary to define the problem or if it was an arbitrary choice.

5) reference [11] seems off.

6) fix JuMP capitalization in references [9] and [10]
