### 2023-09-06 5b973e7
extend operator_to_set. update allowed unicode logical ops
