+++ src/DisjunctiveProgramming.jl
@@ -24,8 +24,8 @@ include("optimize.jl")
 # export logic operators (NOTE: these don't get exported although defined in logic.jl)
 # TODO investigate why this is
 export ∨, ∧, ¬
-export ⇔, <-->
-export ⇒ # --> cannot be used because it clashes with indicator constraint parsing in JuMP.
+export ⇔ # <--> cannot be used because it clashes with reified constraint parsing in JuMP.
+export ⟹ # --> cannot be used because it clashes with indicator constraint parsing in JuMP.
 
 # Define additional stuff that should not be exported
 const _EXCLUDE_SYMBOLS = [Symbol(@__MODULE__), :eval, :include]
