+++ src/constraints.jl
@@ -430,99 +430,99 @@ end
 #         end
 #     end
 # end
-"""
-    function JuMP.parse_constraint_call(
-        _error::Function, 
-        is_vectorized::Bool, 
-        ::Val{:⇒}, 
-        lhs, 
-        rhs
-    )
+# """
+#     function JuMP.parse_constraint_call(
+#         _error::Function, 
+#         is_vectorized::Bool, 
+#         ::Val{:⟹}, 
+#         lhs, 
+#         rhs
+#     )
 
-Extend `JuMP.parse_constraint_call` to avoid order of precedence when 
-using the logical operator ⇒ when building a 
-logical proposition constraint.
-"""
-function JuMP.parse_constraint_call(
-    _error::Function, 
-    is_vectorized::Bool, 
-    ::Val{:⇒}, 
-    lhs, 
-    rhs
-)
-    op = :⇒
-    error_msg = 
-        "The MOI set in the logical constraint $lhs $op $rhs was not identified. " *
-        "If using the `func in set` notation, the set `MOI.EqualTo{Bool}(value)` " *
-        "should be preceded by `in` or `∈`. Otherwise `func == Bool` should be used " *
-        "to define the constraint."
-    rhs0, set0 = rhs.args[2:3]
-    if rhs.args[1] in (:in, :∈) && set0.head == :call && set0.args[2] isa Bool    
-        set = set0
-    elseif rhs.args[1] == :(==) && set0 isa Bool
-        set = _MOI.EqualTo(set0)
-    else
-        _error(error_msg)
-    end
-    func = :($op($lhs, $rhs0))
-    parse_code = :()
-    build_code = :(JuMP.build_constraint($(_error), $(esc(func)), $(esc(set))))
-    return parse_code, build_code
-end
+# Extend `JuMP.parse_constraint_call` to avoid order of precedence when 
+# using the logical operator ⟹ when building a 
+# logical proposition constraint.
+# """
+# function JuMP.parse_constraint_call(
+#     _error::Function, 
+#     is_vectorized::Bool, 
+#     ::Val{:⟹}, 
+#     lhs, 
+#     rhs
+# )
+#     op = :⟹
+#     error_msg = 
+#         "The MOI set in the logical constraint $lhs $op $rhs was not identified. " *
+#         "If using the `func in set` notation, the set `MOI.EqualTo{Bool}(value)` " *
+#         "should be preceded by `in` or `∈`. Otherwise `func == Bool` should be used " *
+#         "to define the constraint."
+#     rhs0, set0 = rhs.args[2:3]
+#     if rhs.args[1] in (:in, :∈) && set0.head == :call && set0.args[2] isa Bool    
+#         set = set0
+#     elseif rhs.args[1] == :(==) && set0 isa Bool
+#         set = _MOI.EqualTo(set0)
+#     else
+#         _error(error_msg)
+#     end
+#     func = :($op($lhs, $rhs0))
+#     parse_code = :()
+#     build_code = :(JuMP.build_constraint($(_error), $(esc(func)), $(esc(set))))
+#     return parse_code, build_code
+# end
 
-"""
-    function JuMP.parse_constraint_call(
-        _error::Function, 
-        is_vectorized::Bool, 
-        ::Union{Val{:⇔}, Val{:(<-->)}}, 
-        lhs, 
-        rhs
-    )
+# """
+#     function JuMP.parse_constraint_call(
+#         _error::Function, 
+#         is_vectorized::Bool, 
+#         ::Union{Val{:⇔}, Val{:(<-->)}}, 
+#         lhs, 
+#         rhs
+#     )
 
-Extend `JuMP.parse_constraint_call` to avoid order of precedence when 
-using the logical operator `⇔` or `<-->` when building a 
-logical proposition constraint.
-"""
[…]
