+++ src/logic.jl
@@ -1,57 +1,39 @@
 ################################################################################
 #                              LOGIC OPERATORS
 ################################################################################
-
-# Define all the logic functions/operators that use 2+ arguments
-for (name, func) in (
-    (:⇔, :⇔), (:⇔, :iff), (:⇔, :equals), # \Leftrightarrow + tab
-    (:∨, :∨), (:∨, :lor), # \vee + tab
-    (:∧, :∧), (:∧, :land), # \wedge + tab
-)
-    # make an expression constructor
-    @eval begin 
-        function $func(
-            v1::Union{LogicalVariableRef, _LogicalExpr}, 
-            v2::Union{LogicalVariableRef, _LogicalExpr},
-            v...
-            )
-            return _LogicalExpr(Symbol($name), Any[v1, v2, v...])
-        end
-    end
+function _op_fallback(name)
+    error("`$name` is only supported for logical expressions")
 end
 
-# Define all the logic functions/operators that use 2 arguments
-for (name, func) in (
-    (:⟹, :⟹), (:⟹, :implies), # \Rightarrow + tab
-    (:Ξ, :Ξ), (:Ξ, :exactly), # \Xi + tab
-    (:Λ, :Λ), (:Λ, :atmost), # \Lambda + tab
-    (:Γ, :Γ), (:Γ, :atleast) # \Gamma + tab
-)
-    # make an expression constructor
+# Define all the logical operators TODO rename the functions as desired
+for (name, alt) in (
+    (:⇔, :iff), # \Leftrightarrow + tab
+    (:⟹, :implies) # Longrightarrow + tab
+    )
+    # make operators
     @eval begin 
-        function $func(
-            v1::Union{LogicalVariableRef, _LogicalExpr}, 
-            v2::Union{LogicalVariableRef, _LogicalExpr}
-            )
-            return _LogicalExpr(Symbol($name), Any[v1, v2])
-        end
+        const $name = JuMP.NonlinearOperator($(Meta.quot(name)), (vs...) -> _op_fallback($(Meta.quot(name))))
+        const $alt = JuMP.NonlinearOperator($(Meta.quot(name)), (vs...) -> _op_fallback($(Meta.quot(alt))))
     end
 end
-
-# Define all the logic functions/operators that use 1 argument
-for (name, func) in ((:¬, :¬), (:¬, :lneg))
-    # make an expression constructor
+for (name, alt, func) in (
+    (:∨, :logical_or, :(|)), # \vee + tab
+    (:∧, :logical_and, :(&)), # \wedge + tab
+    (:¬, :logical_negation, :(!)), # \neg + tab
+    # (:Ξ, :exactly, :exactly), # \Xi + tab --> These should probably just be specified via constraint sets so there aren't 2 different ways
+    # (:Λ, :atmost, :atmost), # \Lambda + tab
+    # (:Γ, :atleast, :atleast) # \Gamma + tab
+    )
+    # make operators
     @eval begin 
-        function $func(v::Union{LogicalVariableRef, _LogicalExpr})
-            return _LogicalExpr(Symbol($name), Any[v])
-        end
+        const $name = JuMP.NonlinearOperator($(Meta.quot(name)), $func)
+        const $alt = JuMP.NonlinearOperator($(Meta.quot(name)), $func)
     end
 end
 
 ################################################################################
 #                            CONJUNCTIVE NORMAL FORM
 ################################################################################
-
 """
 
 """
@@ -76,7 +58,7 @@ function _eliminate_equivalence(lvar::LogicalVariableRef)
     return lvar
 end
[…]
