+++ src/reformulate.jl
@@ -136,7 +136,7 @@ function _reformulate_disjunctive_constraint(
     con::JuMP.ScalarConstraint{T, S}, 
     bvref::JuMP.VariableRef,
     method::Hull
-) where {T <: JuMP.NonlinearExpr, S <: _MOI.LessThan}
+) where {T <: JuMP.GenericNonlinearExpr, S <: _MOI.LessThan}
     #TODO: need to pass _error to build_constraint
     ϵ = method.value
     con_func = _disaggregate_nl_expression(model, con.func, bvref, method)
@@ -168,7 +168,7 @@ function _reformulate_disjunctive_constraint(
     con::JuMP.ScalarConstraint{T, S}, 
     bvref::JuMP.VariableRef,
     method::Hull
-) where {T <: JuMP.NonlinearExpr, S <: _MOI.GreaterThan}
+) where {T <: JuMP.GenericNonlinearExpr, S <: _MOI.GreaterThan}
     #TODO: need to pass _error to build_constraint
     ϵ = method.value
     con_func = _disaggregate_nl_expression(model, con.func, bvref, method)
@@ -201,7 +201,7 @@ function _reformulate_disjunctive_constraint(
     con::JuMP.ScalarConstraint{T, S}, 
     bvref::JuMP.VariableRef,
     method::Hull
-) where {T <: JuMP.NonlinearExpr, S <: _MOI.Interval}
+) where {T <: JuMP.GenericNonlinearExpr, S <: _MOI.Interval}
     #TODO: need to pass _error to build_constraint
     ϵ = method.value
     con_func = _disaggregate_nl_expression(model, con.func, bvref, method)
@@ -235,7 +235,7 @@ function _reformulate_disjunctive_constraint(
     con::JuMP.ScalarConstraint{T, S}, 
     bvref::JuMP.VariableRef,
     method::Hull
-    ) where {T <: JuMP.NonlinearExpr, S <: _MOI.EqualTo}
+    ) where {T <: JuMP.GenericNonlinearExpr, S <: _MOI.EqualTo}
     #TODO: need to pass _error to build_constraint
     ϵ = method.value
     con_func = _disaggregate_nl_expression(model, con.func, bvref, method)
@@ -295,21 +295,21 @@ function _reformulate_logical_constraint(model::JuMP.Model, lexpr::_LogicalExpr,
 end
 
 function _reformulate_selector(model::JuMP.Model, ::MOIAtLeast, val::Number, lvrefs::Vector{LogicalVariableRef})
-    bvrefs = Vector{Any}(_indicator_to_binary_ref.(lvrefs))
+    bvrefs = _indicator_to_binary_ref.(lvrefs)
     return JuMP.add_constraint(model,
-        JuMP.build_constraint(error, JuMP.NonlinearExpr(:+, bvrefs), _MOI.GreaterThan(val))
+        JuMP.build_constraint(error, JuMP.@expression(model, sum(bvrefs)), _MOI.GreaterThan(val))
     )
 end
 function _reformulate_selector(model::JuMP.Model, ::MOIAtMost, val::Number, lvrefs::Vector{LogicalVariableRef})
-    bvrefs = Vector{Any}(_indicator_to_binary_ref.(lvrefs))
+    bvrefs = _indicator_to_binary_ref.(lvrefs)
     return JuMP.add_constraint(model,
-        JuMP.build_constraint(error, JuMP.NonlinearExpr(:+, bvrefs), _MOI.LessThan(val))
+        JuMP.build_constraint(error, JuMP.@expression(model, sum(bvrefs)), _MOI.LessThan(val))
     )
 end
[…]
