### 2023-09-08 2dbc184
pass fix and start values from logical to bin vars
