### 2023-09-08 b4ac164
update NonlinearOperator arg order; clean up
