+++ src/constraints.jl
@@ -417,152 +417,12 @@ end
 ################################################################################
 #                              LOGICAL CONSTRAINTS
 ################################################################################
-# _first_order_ops = (
-#     (:Ξ, :exactly),
-#     (:Λ, :atmost), 
-#     (:Γ, :atleast)
-# )
-# for (ops, set) in zip(_first_order_ops, [Exactly, AtMost, AtLeast])
-#     for op in ops
-#         function JuMP.parse_constraint_call(_error::Function, ::Bool, ::Val{op}, val, lvec)
-#             build_code = :(JuMP.build_constraint($(_error), $(esc(lvec)), $set($(esc(val)))))
-#             return :(), build_code
-#         end
-#     end
-# end
-# """
-#     function JuMP.parse_constraint_call(
-#         _error::Function, 
-#         is_vectorized::Bool, 
-#         ::Val{:⟹}, 
-#         lhs, 
-#         rhs
-#     )
-
-# Extend `JuMP.parse_constraint_call` to avoid order of precedence when 
-# using the logical operator ⟹ when building a 
-# logical proposition constraint.
-# """
-# function JuMP.parse_constraint_call(
-#     _error::Function, 
-#     is_vectorized::Bool, 
-#     ::Val{:⟹}, 
-#     lhs, 
-#     rhs
-# )
-#     op = :⟹
-#     error_msg = 
-#         "The MOI set in the logical constraint $lhs $op $rhs was not identified. " *
-#         "If using the `func in set` notation, the set `MOI.EqualTo{Bool}(value)` " *
-#         "should be preceded by `in` or `∈`. Otherwise `func == Bool` should be used " *
-#         "to define the constraint."
-#     rhs0, set0 = rhs.args[2:3]
-#     if rhs.args[1] in (:in, :∈) && set0.head == :call && set0.args[2] isa Bool    
-#         set = set0
-#     elseif rhs.args[1] == :(==) && set0 isa Bool
-#         set = _MOI.EqualTo(set0)
-#     else
-#         _error(error_msg)
-#     end
-#     func = :($op($lhs, $rhs0))
-#     parse_code = :()
-#     build_code = :(JuMP.build_constraint($(_error), $(esc(func)), $(esc(set))))
-#     return parse_code, build_code
-# end
-
-# """
-#     function JuMP.parse_constraint_call(
-#         _error::Function, 
-#         is_vectorized::Bool, 
-#         ::Union{Val{:⇔}, Val{:(<-->)}}, 
-#         lhs, 
-#         rhs
-#     )
-
-# Extend `JuMP.parse_constraint_call` to avoid order of precedence when 
-# using the logical operator `⇔` or `<-->` when building a 
-# logical proposition constraint.
-# """
-# function JuMP.parse_constraint_call(
-#     _error::Function, 
-#     is_vectorized::Bool, 
-#     ::Union{Val{:⇔}, Val{:(<-->)}}, 
-#     lhs, 
-#     rhs
-# )
-#     op = :⇔
-#     error_msg = 
-#         "The MOI set in the logical constraint $lhs $op $rhs was not identified. " *
-#         "If using the `func in set` notation, the set `MOI.EqualTo{Bool}(value)` " *
-#         "should be preceded by `in` or `∈`. Otherwise `func == Bool` should be used " *
-#         "to define the constraint."
-#     op_args = [lhs] #initialize
-#     tmp = rhs
-#     set = _MOI.EqualTo(true) 
-#     while true
-#         set0 = tmp.args[3]
-#         if tmp.args[1] in (:⇔, :(<-->))
-#             push!(op_args, tmp.args[2])
-#             tmp = set0
[…]
