### 2023-09-11 7df62f8
extend BigM for VectorConstraints
