+++ src/bigm.jl
@@ -1,69 +1,76 @@
-# TODO extend for VectorConstraints and possible other constraint types
+# Get Big-M value for a particular constraint
+function _get_M_value(method::BigM, func::JuMP.AbstractJuMPScalar, set::_MOI.AbstractSet)
+    if method.tighten
+        return _get_tight_M(method, func, set)
+    else
+        return _get_M(method, func, set)
+    end
+end
 
 # Get the tightest Big-M value for a particular constraint
-function _get_tight_M(method::BigM, con::JuMP.ScalarConstraint{T, S}) where {T, S <: Union{_MOI.LessThan, _MOI.GreaterThan}}
-    M = min(method.value, _calculate_tight_M(con))
-    if isinf(M)
-        error("A finite Big-M value must be used. The value given was $M.")
-    end
-
-    return M
-end
-function _get_tight_M(method::BigM, con::JuMP.ScalarConstraint{T, S}) where {T, S <: Union{_MOI.Interval, _MOI.EqualTo}}
-    M = min.([method.value, method.value], _calculate_tight_M(con))
+function _get_tight_M(method::BigM, func::JuMP.AbstractJuMPScalar, set::_MOI.AbstractSet)
+    M = min.(method.value, _calculate_tight_M(func, set)) #broadcast for when S <: MOI.Interval or MOI.EqualTo or MOI.Zeros
     if any(isinf.(M))
-        error("A finite Big-M value must be used. The value given was $M.")
+        error("A finite Big-M value must be used. The value obtained was $M.")
     end
-
     return M
 end
 
 # Get user-specified Big-M value
-function _get_M(method::BigM, ::JuMP.ScalarConstraint{T, S}) where {T, S <: Union{_MOI.LessThan, _MOI.GreaterThan}}
+function _get_M(method::BigM, ::JuMP.AbstractJuMPScalar, ::Union{_MOI.LessThan, _MOI.GreaterThan, _MOI.Nonnegatives, _MOI.Nonpositives})
     M = method.value
     if isinf(M)
         error("A finite Big-M value must be used. The value given was $M.")
     end
-
     return M
 end
-function _get_M(method::BigM, ::JuMP.ScalarConstraint{T, S}) where {T, S <: Union{_MOI.Interval, _MOI.EqualTo}}
+function _get_M(method::BigM, ::JuMP.AbstractJuMPScalar, ::Union{_MOI.Interval, _MOI.EqualTo, _MOI.Zeros})
     M = method.value
     if isinf(M)
         error("A finite Big-M value must be used. The value given was $M.")
     end
-
-    return (M, M)
+    return [M, M]
 end
 
 # Apply interval arithmetic on a linear constraint to infer the tightest Big-M value from the bounds on the constraint.
-function _calculate_tight_M(con::JuMP.ScalarConstraint{JuMP.AffExpr, S}) where {S <: _MOI.LessThan}
-    return _interval_arithmetic_LessThan(con, -con.set.upper)
+function _calculate_tight_M(func::JuMP.AffExpr, set::_MOI.LessThan)
+    return _interval_arithmetic_LessThan(func, -set.upper)
 end
-function _calculate_tight_M(con::JuMP.ScalarConstraint{JuMP.AffExpr, S}) where {S <: _MOI.GreaterThan}
-    return _interval_arithmetic_GreaterThan(con, -con.set.lower)
+function _calculate_tight_M(func::JuMP.AffExpr, set::_MOI.GreaterThan)
+    return _interval_arithmetic_GreaterThan(func, -set.lower)
 end
[…]
