### 2023-09-11 f4ae4b9
use @expression for all reformulations; bug fixes; improve code
