+++ src/reformulate.jl
@@ -17,9 +17,7 @@ end
 ################################################################################
 #                              DISJUNCTIONS
 ################################################################################
-"""
-
-"""
+# disjunctions
 function _reformulate_disjunctions(model::JuMP.Model, method::AbstractReformulationMethod)
     for (disj_idx, disj) in _disjunctions(model)
         _reformulate_disjuncts(model, disj_idx, disj, method)
@@ -44,9 +42,7 @@ function _reformulate_disjuncts(model::JuMP.Model, disj_idx::DisjunctionIndex, d
     end
 end
 
-"""
-
-"""
+# individual disjuncts
 function _reformulate_disjunct(model::JuMP.Model, d::Disjunct, method::AbstractReformulationMethod)
     #reformulate each constraint and add to the model
     ind_idx = JuMP.index(d.indicator) #logical indicator
@@ -54,14 +50,12 @@ function _reformulate_disjunct(model::JuMP.Model, d::Disjunct, method::AbstractR
     bvref = JuMP.VariableRef(model, bv_idx)
     for con in d.constraints
         cdata = _disjunct_constraints(model)[JuMP.index(con)]
-        _reformulate_disjunctive_constraint(model, cdata.constraint, bvref, method)
+        _reformulate_disjunct_constraint(model, cdata.constraint, bvref, method)
     end
 end
 
-"""
-
-"""
-function _reformulate_disjunctive_constraint(
+# BigM: individual disjunct constraints
+function _reformulate_disjunct_constraint(
     model::JuMP.Model,
     con::JuMP.ScalarConstraint{T, S}, 
     bvref::JuMP.VariableRef,
@@ -69,23 +63,29 @@ function _reformulate_disjunctive_constraint(
 ) where {T, S <: _MOI.LessThan}
     #TODO: need to pass _error to build_constraint
     M = _get_M_value(method, con.func, con.set)
+    new_func = JuMP.@expression(model, con.func - M*(1-bvref))
     JuMP.add_constraint(model,
-        JuMP.build_constraint(error, con.func - M*(1-bvref), con.set)
+        JuMP.build_constraint(error, new_func, con.set)
     )
 end
-function _reformulate_disjunctive_constraint(
+function _reformulate_disjunct_constraint(
     model::JuMP.Model,
     con::JuMP.VectorConstraint{T, S}, 
     bvref::JuMP.VariableRef,
     method::BigM
 ) where {T, S <: _MOI.Nonpositives}
     #TODO: need to pass _error to build_constraint
-    M = [_get_M_value(method, func, con.set) for func in con.func]
+    M = JuMP.@expression(model, [i=1:con.set.dimension],
+        _get_M_value(method, con.func[i], con.set)
+    )
+    new_func = JuMP.@expression(model, [i=1:con.set.dimension], 
+        con.func[i] - M[i]*(1-bvref)
+    )
     JuMP.add_constraint(model,
-        JuMP.build_constraint(error, [func - M[i]*(1-bvref) for (i,func) in enumerate(con.func)], con.set)
+        JuMP.build_constraint(error, new_func, con.set)
     )
 end
-function _reformulate_disjunctive_constraint(
+function _reformulate_disjunct_constraint(
     model::JuMP.Model, 
     con::JuMP.ScalarConstraint{T, S}, 
     bvref::JuMP.VariableRef,
@@ -93,23 +93,29 @@ function _reformulate_disjunctive_constraint(
[…]
