### 2023-09-11 db5d95a
extend hull reform to vectorconstraints; cleanup
