+++ src/reformulate.jl
@@ -1,3 +1,8 @@
+_set_value(set::_MOI.LessThan) = set.upper
+_set_value(set::_MOI.GreaterThan) = set.lower
+_set_value(set::_MOI.EqualTo) = set.value
+_set_values(set::_MOI.Interval) = (set.lower, set.upper)
+
 ################################################################################
 #                              LOGICAL VARIABLES
 ################################################################################
@@ -23,13 +28,12 @@ function _reformulate_disjunctions(model::JuMP.Model, method::AbstractReformulat
         _reformulate_disjuncts(model, disj_idx, disj, method)
     end
 end
-
+# disjuncts
 function _reformulate_disjuncts(model::JuMP.Model, disj_idx::DisjunctionIndex, disj::ConstraintData{T}, method::Union{BigM,Indicator}) where {T<:Disjunction}
     for d in disj.constraint.disjuncts
         _reformulate_disjunct(model, d, method)
     end
 end
-
 function _reformulate_disjuncts(model::JuMP.Model, disj_idx::DisjunctionIndex, disj::ConstraintData{T}, method::Hull) where {T<:Disjunction}
     disj_vrefs = _get_disjunction_variables(disj)
     _update_variable_bounds.(disj_vrefs) 
@@ -70,10 +74,10 @@ function _reformulate_disjunct_constraint(
 end
 function _reformulate_disjunct_constraint(
     model::JuMP.Model,
-    con::JuMP.VectorConstraint{T, S}, 
+    con::JuMP.VectorConstraint{T, S, R}, 
     bvref::JuMP.VariableRef,
     method::BigM
-) where {T, S <: _MOI.Nonpositives}
+) where {T, S <: _MOI.Nonpositives, R}
     #TODO: need to pass _error to build_constraint
     M = JuMP.@expression(model, [i=1:con.set.dimension],
         _get_M_value(method, con.func[i], con.set)
@@ -100,10 +104,10 @@ function _reformulate_disjunct_constraint(
 end
 function _reformulate_disjunct_constraint(
     model::JuMP.Model, 
-    con::JuMP.VectorConstraint{T, S}, 
+    con::JuMP.VectorConstraint{T, S, R}, 
     bvref::JuMP.VariableRef,
     method::BigM,
-) where {T, S <: _MOI.Nonnegatives}
+) where {T, S <: _MOI.Nonnegatives, R}
     #TODO: need to pass _error to build_constraint
     M = JuMP.@expression(model, [i=1:con.set.dimension],
         _get_M_value(method, con.func[i], con.set)
@@ -120,41 +124,25 @@ function _reformulate_disjunct_constraint(
     con::JuMP.ScalarConstraint{T, S}, 
     bvref::JuMP.VariableRef,
     method::BigM
-) where {T, S <: _MOI.Interval}
+) where {T, S <: Union{_MOI.Interval, _MOI.EqualTo}}
     #TODO: need to pass _error to build_constraint
     M = _get_M_value(method, con.func, con.set)
     new_func_gt = JuMP.@expression(model, con.func + M[1]*(1-bvref))
     new_func_lt = JuMP.@expression(model, con.func - M[2]*(1-bvref))
+    set_values = _set_values(con.set)
     JuMP.add_constraint(model,
-        JuMP.build_constraint(error, new_func_gt, _MOI.GreaterThan(con.set.lower))
+        JuMP.build_constraint(error, new_func_gt, _MOI.GreaterThan(set_values[1]))
     )
     JuMP.add_constraint(model,
[…]
