### 2023-09-12 d7adb3a
refactor to simplify objects stored in gdpdata -use tmp storage for diaggregated variables -remove Disjunct struct -add reformulation_vars/const objects
