+++ src/model.jl
@@ -60,7 +60,7 @@ end
 """
 function disjunction_indicators(disjunction::DisjunctionRef) 
     model, idx = disjunction.model, disjunction.index
-    ind_idxs = _disjunction_to_indicators(model)[idx]
+    ind_idxs = _disjunctions(model)[idx].constraint
     return LogicalVariableRef.(model, ind_idxs) # TODO account for nested disjunctions
 end
 
@@ -81,14 +81,6 @@ function _disjunctions(model::JuMP.Model)
     return model.ext[:GDP].disjunctions
 end
 
-function _disjunction_to_indicators(model::JuMP.Model)
-    return model.ext[:GDP].disjunction_to_indicators
-end
-
-function _indicator_to_disjunction(model::JuMP.Model)
-    return model.ext[:GDP].indicator_to_disjunction
-end
-
 function _indicator_to_constraints(model::JuMP.Model)
     return model.ext[:GDP].indicator_to_constraints
 end
@@ -101,21 +93,18 @@ function _indicator_to_binary(model::JuMP.Model)
     return model.ext[:GDP].indicator_to_binary
 end
 
-function _global_to_disjunct_variable(model::JuMP.Model)
-    return model.ext[:GDP].global_to_disjunct_variable
-end
-
-function _global_to_disjunction_variables(model::JuMP.Model)
-    return model.ext[:GDP].global_to_disjunction_variables
-end
-
 function _variable_bounds(model::JuMP.Model)
     return model.ext[:GDP].variable_bounds
 end
 
-"""
+function _reformulation_variables(model::JuMP.Model)
+    return model.ext[:GDP].reformulation_variables
+end
 
-"""
+function _reformulation_constraints(model::JuMP.Model)
+    return model.ext[:GDP].reformulation_constraints
+end
+
 function _indicator_to_binary_ref(lvref::LogicalVariableRef)
     model, idx = lvref.model, lvref.index
     return JuMP.VariableRef(model, _indicator_to_binary(model)[idx])
@@ -130,15 +119,13 @@ function JuMP.copy_extension_data(data::GDPData, new_model::JuMP.AbstractModel, 
         data.logical_variables, 
         _copy(data.logical_constraints, new_model), 
         _copy(data.disjunct_constraints, new_model),
-        _copy(data.disjunctions, new_model), 
-        data.disjunction_to_indicators,
-        data.indicator_to_disjunction,
+        data.disjunctions,
         data.indicator_to_binary,
         data.indicator_to_constraints,
         data.constraint_to_indicator,
-        data.global_to_disjunct_variable,
-        data.global_to_disjunction_variables,
         data.variable_bounds,
+        data.reformulation_variables,
+        data.reformulation_constraints,
         data.solution_method, 
         data.ready_to_optimize,
     )
