+++ src/hull.jl
@@ -5,10 +5,13 @@
 """
 
 """
-function _get_disjunction_variables(disj::ConstraintData{Disjunction})
+function _get_disjunction_variables(model::JuMP.Model, disj::ConstraintData{Disjunction})
     vars = Set{JuMP.VariableRef}()
-    for d in disj.constraint.disjuncts
-        _interrogate_variables(v -> push!(vars, v), d)
+    for ind_idx in disj.constraint.disjuncts
+        for cidx in _indicator_to_constraints(model)[ind_idx]
+            cdata = _disjunct_constraints(model)[cidx]
+            _interrogate_variables(v -> push!(vars, v), cdata.constraint)
+        end
     end
     return vars
 end
@@ -56,20 +59,15 @@ end
 function _interrogate_variables(interrogator::Function, con::JuMP.ScalarConstraint)
     _interrogate_variables(interrogator, con.func)
 end
+function _interrogate_variables(interrogator::Function, con::JuMP.VectorConstraint)
+    for func in con.func
+        _interrogate_variables(interrogator, func)
+    end
+end
 
 # AbstractArray
 function _interrogate_variables(interrogator::Function, arr::AbstractArray)
-    _interrogate_variables.(interrogator, ex)
-    return
-end
-
-# Disjunct
-function _interrogate_variables(interrogator::Function, d::Disjunct)
-    for con in d.constraints
-        model, idx = con.model, con.index
-        cdata = _disjunct_constraints(model)[idx]
-        _interrogate_variables(interrogator, cdata.constraint)
-    end
+    _interrogate_variables.(interrogator, arr)
     return
 end
 
@@ -93,52 +91,49 @@ function _update_variable_bounds(vref::JuMP.VariableRef)
         _variable_bounds(model)[idx] = (lb, ub)
     end
 end
-function _disaggregate_variables(model::JuMP.Model, d::Disjunct, vrefs::Set{JuMP.VariableRef})
+function _disaggregate_variables(model::JuMP.Model, ind_idx::LogicalVariableIndex, vrefs::Set{JuMP.VariableRef}, disag_vars::_Hull)
     #create disaggregated variables for that disjunct
     for vref in vrefs
         JuMP.is_binary(vref) && continue #skip binary variables
-        _disaggregate_variable(model, d, vref) #create disaggregated var
+        _disaggregate_variable(model, ind_idx, vref, disag_vars) #create disaggregated var for that disjunct
     end
 end
-function _disaggregate_variable(model::JuMP.Model, d::Disjunct, vref::JuMP.VariableRef)
+function _disaggregate_variable(model::JuMP.Model, ind_idx::LogicalVariableIndex, vref::JuMP.VariableRef, disag_vars::_Hull)
     #create disaggregated (disjunct) vref
     v_idx = JuMP.index(vref) #variable
     lb, ub = _variable_bounds(model)[v_idx]
     dvar = JuMP.ScalarVariable(_variable_info(lower_bound = lb, upper_bound = ub))
-    dvref = JuMP.add_variable(model, dvar, "$(vref)_$(d.indicator)")
+    lvref = LogicalVariableRef(model, ind_idx)
+    dvref = JuMP.add_variable(model, dvar, "$(vref)_$(lvref)")
     dv_idx = JuMP.index(dvref) #disaggregated (disjunct) variable
+    push!(_reformulation_variables(model), dv_idx)
     #get binary indicator variable
-    ind_idx = JuMP.index(d.indicator) #logical indicator
[…]
