### 2023-09-12 e57b7e4
extend JuMP.delete for GDP constraints and vars
