### 2023-09-13 107ac0d
support vector constraints for disjunctconstraint
