### 2023-09-13 aa47a1c
update optimize_hook
