### 2023-09-13 ee9ff23
delete reformulated constraints and vars from previous optimize call
