+++ src/reformulate.jl
@@ -69,10 +69,12 @@ function _reformulate_disjunct_constraint(
 ) where {T, S <: _MOI.LessThan}
     M = _get_M_value(method, con.func, con.set)
     new_func = JuMP.@expression(model, con.func - M*(1-bvref))
-    reform_con = JuMP.add_constraint(model,
+    reform_con = JuMP.add_constraint(model, 
         JuMP.build_constraint(error, new_func, con.set)
     )
-    push!(_reformulation_constraints(model), JuMP.index(reform_con))
+    push!(_reformulation_constraints(model), 
+        (JuMP.index(reform_con), JuMP.ScalarShape())
+    )
 end
 function _reformulate_disjunct_constraint(
     model::JuMP.Model,
@@ -89,7 +91,7 @@ function _reformulate_disjunct_constraint(
     reform_con = JuMP.add_constraint(model,
         JuMP.build_constraint(error, new_func, con.set)
     )
-    push!(_reformulation_constraints(model), JuMP.index(reform_con))
+    push!(_reformulation_constraints(model), (JuMP.index(reform_con), JuMP.VectorShape()))
 end
 function _reformulate_disjunct_constraint(
     model::JuMP.Model, 
@@ -102,7 +104,7 @@ function _reformulate_disjunct_constraint(
     reform_con = JuMP.add_constraint(model,
         JuMP.build_constraint(error, new_func, con.set)
     )
-    push!(_reformulation_constraints(model), JuMP.index(reform_con))
+    push!(_reformulation_constraints(model), (JuMP.index(reform_con), JuMP.ScalarShape()))
 end
 function _reformulate_disjunct_constraint(
     model::JuMP.Model, 
@@ -119,7 +121,7 @@ function _reformulate_disjunct_constraint(
     reform_con = JuMP.add_constraint(model,
         JuMP.build_constraint(error, new_func, con.set)
     )
-    push!(_reformulation_constraints(model), JuMP.index(reform_con))
+    push!(_reformulation_constraints(model), (JuMP.index(reform_con), JuMP.VectorShape()))
 end
 function _reformulate_disjunct_constraint(
     model::JuMP.Model, 
@@ -137,7 +139,10 @@ function _reformulate_disjunct_constraint(
     reform_con_lt = JuMP.add_constraint(model,
         JuMP.build_constraint(error, new_func_lt, _MOI.LessThan(set_values[2]))
     )
-    push!(_reformulation_constraints(model), JuMP.index(reform_con_gt), JuMP.index(reform_con_lt))
+    push!(_reformulation_constraints(model), 
+        (JuMP.index(reform_con_gt), JuMP.ScalarShape()), 
+        (JuMP.index(reform_con_lt), JuMP.ScalarShape())
+    )
 end
 function _reformulate_disjunct_constraint(
     model::JuMP.Model, 
@@ -160,7 +165,10 @@ function _reformulate_disjunct_constraint(
     reforn_con_np = JuMP.add_constraint(model,
         JuMP.build_constraint(error, new_func_np, _MOI.Nonpositives(con.set.dimension))
     )
-    push!(_reformulation_constraints(model), JuMP.index(reform_con_nn), JuMP.index(reforn_con_np))
+    push!(_reformulation_constraints(model), 
+        (JuMP.index(reform_con_nn), JuMP.VectorShape()), 
+        (JuMP.index(reform_con_np), JuMP.VectorShape())
+    )
 end
 
 # Hull: individual disjunct constraints
@@ -176,7 +184,7 @@ function _reformulate_disjunct_constraint(
     reform_con = JuMP.add_constraint(model,
[…]
