### 2023-09-13 6a6c2e8
expose function for reformulating a model
