+++ examples/ex3.jl
@@ -10,22 +10,17 @@ m = GDPModel()
 @constraint(m, x >= 5, DisjunctConstraint(Y[2]))
 @disjunction(m, Y)
 @constraint(m, Y in Exactly(1)) #logical constraint
-DisjunctiveProgramming._reformulate_logical_variables(m)
-DisjunctiveProgramming._reformulate_logical_constraints(m)
+@objective(m, Max, x)
 print(m)
-# Feasibility
+# Max x
 # Subject to
-#  Y[1] + Y[2] = 1
 #  x ≥ -5
 #  x ≤ 10
-#  Y[1] binary
-#  Y[2] binary
 
 ##
-m_bigm = copy(m)
-DisjunctiveProgramming._reformulate_disjunctions(m_bigm, BigM())
-print(m_bigm)
-# Feasibility
+reformulate_model(m, BigM())
+print(m)
+# Max x
 # Subject to
 #  (exp(x) - 3.0) + (-1000000000 Y[2] + 1000000000) ≥ 0
 #  (exp(x) - 2.0) - (-1000000000 Y[1] + 1000000000) ≤ 0
@@ -38,15 +33,14 @@ print(m_bigm)
 #  Y[2] binary
 
 ##
-m_hull = copy(m)
-DisjunctiveProgramming._reformulate_disjunctions(m_hull, Hull())
-print(m_hull)
-# Feasibility
+reformulate_model(m, Hull())
+print(m)
+# Max x
 # Subject to
 #  (((0.999999 Y[2] + 1.0e-6) * (exp(x_Y[2] / (0.999999 Y[2] + 1.0e-6)) - 3.0)) - (2.0e-6 Y[2] - 2.0e-6)) - (0) ≥ 0
 #  (((0.999999 Y[1] + 1.0e-6) * (exp(x_Y[1] / (0.999999 Y[1] + 1.0e-6)) - 2.0)) - (1.0e-6 Y[1] - 1.0e-6)) - (0) ≤ 0
-#  Y[1] + Y[2] = 1
 #  x aggregation : -x + x_Y[1] + x_Y[2] = 0
+#  Y[1] + Y[2] = 1
 #  3 Y[1] + x_Y[1] ≥ 0
 #  -5 Y[2] + x_Y[2] ≥ 0
 #  x_Y[1] lower bounding : -5 Y[1] - x_Y[1] ≤ 0
