+++ examples/ex4.jl
@@ -8,19 +8,18 @@ using DisjunctiveProgramming
 m = GDPModel()
 @variable(m, Y[1:4], LogicalVariable)
 @constraint(m, ¬((Y[1] ∧ ¬Y[2]) ⇔ (Y[3] ∨ Y[4])) ∈ MOI.EqualTo(true))
-DisjunctiveProgramming._reformulate_logical_variables(m)
-DisjunctiveProgramming._reformulate_logical_constraints(m)
+reformulate_model(m, BigM())
 print(m)
 # Feasibility
 # Subject to
 #  -Y[1] + Y[2] - Y[3] ≥ -1
-#  Y[2] ≥ 0
-#  Y[3] ≥ 0
 #  -Y[1] + Y[2] - Y[4] ≥ -1
-#  -Y[2] + Y[3] + Y[4] ≥ 0
+#  Y[4] ≥ 0
 #  Y[1] + Y[3] + Y[4] ≥ 1
 #  -Y[1] ≥ -1
-#  Y[4] ≥ 0
+#  Y[2] ≥ 0
+#  -Y[2] + Y[3] + Y[4] ≥ 0
+#  Y[3] ≥ 0
 #  Y[1] binary
 #  Y[2] binary
 #  Y[3] binary
