+++ examples/ex1.jl
@@ -1,46 +1,31 @@
 using JuMP
 using DisjunctiveProgramming
+using HiGHS
 
-## Example 1: 
+## Example 1a: Linear GDP 
 
 # Disjunction Method 1: Assign Logical Variables Explicitly
-m = GDPModel()
+m = GDPModel(HiGHS.Optimizer)
 @variable(m, -5 ≤ x ≤ 10)
 @variable(m, Y[1:2], LogicalVariable)
 @constraint(m, 0 ≤ x ≤ 3, DisjunctConstraint(Y[1]))
 @constraint(m, 5 ≤ x, DisjunctConstraint(Y[2]))
 @constraint(m, x ≤ 9, DisjunctConstraint(Y[2]))
 @disjunction(m, [Y[1], Y[2]])
-
-# Disjunction Method 2: Create Logical Variables from Disjunctions
-# m = GDPModel()
-# @variable(m, -5 ≤ x ≤ 10)
-# @constraint(m, disjunct_1_con, 0 ≤ x ≤ 3, DisjunctConstraint)
-# @constraint(m, disjunct_2_con_a, 5 ≤ x, DisjunctConstraint)
-# @constraint(m, disjunct_2_con_b, x ≤ 9, DisjunctConstraint)
-# @disjunction(m, disjunction, [[disjunct_1_con], [disjunct_2_con_a, disjunct_2_con_b]])
-# Y = disjunction_indicators(disjunction)
-
-# Logical Constraint Method 1: Use func in set notation
-@constraint(m, Y in Exactly(1))
+@constraint(m, Y in Exactly(1)) 
+@objective(m, Max, x)
 
 # Reformulate logical variables and logical constraints
-DisjunctiveProgramming._reformulate_logical_variables(m)
-DisjunctiveProgramming._reformulate_logical_constraints(m)
 print(m)
-# Feasibility
+# Max x
 # Subject to
-#  Y[1] + Y[2] = 1
 #  x ≥ -5
 #  x ≤ 10
-#  Y[1] binary
-#  Y[2] binary
 
 ## BigM reformulation
-m_bigm = copy(m)
-DisjunctiveProgramming._reformulate_disjunctions(m_bigm, BigM())
-print(m_bigm)
-# Feasibility
+optimize!(m, method = BigM())
+print(m)
+# Max x
 # Subject to
 #  Y[1] + Y[2] = 1
 #  x - 5 Y[1] ≥ -5
@@ -53,13 +38,12 @@ print(m_bigm)
 #  Y[2] binary
 
 ## Hull reformulation
-m_hull = copy(m)
-DisjunctiveProgramming._reformulate_disjunctions(m_hull, Hull())
-print(m_hull)
-# Feasibility
+optimize!(m, method = Hull())
+print(m)
+# Max x
 # Subject to
-#  Y[1] + Y[2] = 1
 #  x aggregation : -x + x_Y[1] + x_Y[2] = 0
+#  Y[1] + Y[2] = 1
 #  x_Y[1] ≥ 0
 #  x_Y[1] lower bounding : -5 Y[1] - x_Y[1] ≤ 0
 #  x_Y[1] upper bounding : -10 Y[1] + x_Y[1] ≤ 0
@@ -77,17 +61,29 @@ print(m_hull)
 #  Y[1] binary
 #  Y[2] binary
 
+## Example 1b: Same as Example 1a, but using alternate syntax for disjunction creation and reformulation to MIP via indicator constraints.
+
+# Disjunction Method 2: Create Logical Variables from Disjunctions
[…]
