### 2023-09-13 7fb4540
support indicator reformulation with vector constraints and nonlinear constraints
