+++ test/runtests.jl
@@ -1,0 +1,4 @@
+using DisjunctiveProgramming
+using Test
+
+include("test_aqua.jl")
