+++ src/hull.jl
@@ -83,19 +83,19 @@ function _disaggregate_expression(model::JuMP.Model, aff::JuMP.AffExpr, bvref::J
     return new_expr
 end
 # quadratic expression
-# TODO review what happens when there are bilinear terms with binary variables involved...
+# TODO review what happens when there are bilinear terms with binary variables involved since these are not being disaggregated (e.g., complementarity constraints; though likely irrelevant)...
 function _disaggregate_expression(model::JuMP.Model, quad::JuMP.QuadExpr, bvref::JuMP.VariableRef, method::_Hull)
     #get affine part
     new_expr = zero(JuMP.NonlinearExpr)
      # TODO update when JuMP supports add_to_expression! for NonlinearExpr
-    new_expr += _disaggregate_expression(model, quad.aff, bvref, method)
+    JuMP._MA.add_mul!!(new_expr, _disaggregate_expression(model, quad.aff, bvref, method))
     #get nonlinear part
     ϵ = method.value
     for (pair, coeff) in quad.terms
         da_ref = method.disjunct[pair.a, bvref]
         db_ref = method.disjunct[pair.b, bvref]
          # TODO update when JuMP supports add_to_expression! for NonlinearExpr
-        new_expr += coeff * da_ref * db_ref / ((1-ϵ)*bvref+ϵ)
+         JuMP._MA.add_mul!!(new_expr, coeff * da_ref * db_ref / ((1-ϵ)*bvref+ϵ))
     end
 
     return new_expr
@@ -115,7 +115,7 @@ end
 function _disaggregate_nl_expression(model::JuMP.Model, aff::JuMP.AffExpr, bvref::JuMP.VariableRef, method::_Hull)
     new_expr = zero(JuMP.NonlinearExpr)
      # TODO update when JuMP supports add_to_expression! for NonlinearExpr
-    new_expr += JuMP.AffExpr(aff.constant)
+     JuMP._MA.add_mul!!(new_expr, JuMP.AffExpr(aff.constant))
     ϵ = method.value
     for (vref, coeff) in aff.terms
         if JuMP.is_binary(vref) #keep any binary variables undisaggregated
@@ -124,7 +124,7 @@ function _disaggregate_nl_expression(model::JuMP.Model, aff::JuMP.AffExpr, bvref
             dvref = method.disjunct[vref, bvref]
         end
          # TODO update when JuMP supports add_to_expression! for NonlinearExpr
-        new_expr += coeff * dvref / ((1-ϵ)*bvref+ϵ)
+         JuMP._MA.add_mul!!(new_expr, coeff * dvref / ((1-ϵ)*bvref+ϵ))
     end
     return new_expr
 end
@@ -138,7 +138,7 @@ function _disaggregate_nl_expression(model::JuMP.Model, quad::JuMP.QuadExpr, bvr
         da_ref = method.disjunct[pair.a, bvref]
         db_ref = method.disjunct[pair.b, bvref]
          # TODO update when JuMP supports add_to_expression! for NonlinearExpr
-        new_expr += coeff * da_ref * db_ref / ((1-ϵ)*bvref+ϵ)^2
+         JuMP._MA.add_mul!!(new_expr, coeff * da_ref * db_ref / ((1-ϵ)*bvref+ϵ)^2)
     end
     return new_expr
 end
