### 2023-09-15 38d919c
review and cleanup codebase
