+++ src/variables.jl
@@ -1,11 +1,7 @@
 ################################################################################
 #                              LOGICAL VARIABLES
 ################################################################################
-"""
-    _variable_info(binary::Bool=false, lower_bound::Float64=-Inf, upper_bound::Float64=Inf)
-
-Wrapper to build `JuMP.VariableInfo` for 
-"""
+# helper function to create the variable info used when creating reformulation variables
 function _variable_info(;
     binary::Bool=false, 
     lower_bound::Float64=-Inf, upper_bound::Float64=Inf, 
@@ -74,65 +70,44 @@ function JuMP.add_variable(
     return LogicalVariableRef(model, idx)
 end
 
-# Basic base extensions
+# Base extensions
 Base.copy(v::LogicalVariableRef) = v
 Base.broadcastable(v::LogicalVariableRef) = Ref(v)
 Base.length(v::LogicalVariableRef) = 1
-
-"""
-
-"""
-JuMP.owner_model(vref::LogicalVariableRef) = vref.model
-
-"""
-
-"""
-JuMP.index(vref::LogicalVariableRef) = vref.index
-
 function Base.:(==)(v::LogicalVariableRef, w::LogicalVariableRef)
     return v.model === w.model && v.index == w.index
 end
-JuMP.isequal_canonical(v::LogicalVariableRef, w::LogicalVariableRef) = v == w
-
 function Base.getindex(map::JuMP.ReferenceMap, vref::LogicalVariableRef)
     return LogicalVariableRef(map.model, JuMP.index(vref))
 end
 
-"""
+# JuMP extensions
+JuMP.owner_model(vref::LogicalVariableRef) = vref.model
 
-"""
+JuMP.index(vref::LogicalVariableRef) = vref.index
+
+JuMP.isequal_canonical(v::LogicalVariableRef, w::LogicalVariableRef) = v == w
+
 function JuMP.is_valid(model::JuMP.Model, vref::LogicalVariableRef)
     return model === JuMP.owner_model(vref)
 end
 
-"""
-
-"""
 function JuMP.name(vref::LogicalVariableRef)
     data = gdp_data(JuMP.owner_model(vref))
     return data.logical_variables[JuMP.index(vref)].name
 end
 
-"""
-
-"""
 function JuMP.set_name(vref::LogicalVariableRef, name::String)
     data = gdp_data(JuMP.owner_model(vref))
     data.logical_variables[JuMP.index(vref)].name = name
     return
 end
 
-"""
-
-"""
 function JuMP.start_value(vref::LogicalVariableRef)
     data = gdp_data(JuMP.owner_model(vref))
     return data.logical_variables[JuMP.index(vref)].variable.start
 end
 
-"""
-
-"""
 function JuMP.set_start_value(
     vref::LogicalVariableRef, 
     value::Union{Nothing, Bool}
@@ -144,17 +119,11 @@ function JuMP.set_start_value(
     return
 end
 
-"""
-
-"""
 function JuMP.fix_value(vref::LogicalVariableRef)
     data = gdp_data(JuMP.owner_model(vref))
     return data.logical_variables[JuMP.index(vref)].variable.fix_value
 end
 
-"""
-
-"""
 function JuMP.fix(vref::LogicalVariableRef, value::Bool)
     data = gdp_data(JuMP.owner_model(vref))
     var = data.logical_variables[JuMP.index(vref)].variable
@@ -163,9 +132,6 @@ function JuMP.fix(vref::LogicalVariableRef, value::Bool)
     return
 end
 
-"""
-
-"""
 function JuMP.unfix(vref::LogicalVariableRef)
     data = gdp_data(JuMP.owner_model(vref))
[…]
