### 2023-09-15 1402750
add tests for logical vars; fix bugs
