+++ test/runtests.jl
@@ -1,4 +1,8 @@
 using DisjunctiveProgramming
+using JuMP
 using Test
 
+const DP = DisjunctiveProgramming
+
 include("test_aqua.jl")
+include("test_lvar.jl")
