### 2023-09-15 98665e5
add lvar tests
