+++ src/reformulate.jl
@@ -22,21 +22,39 @@ end
 # disjunctions
 function _reformulate_disjunctions(model::JuMP.Model, method::AbstractReformulationMethod)
     for (_, disj) in _disjunctions(model)
-        _reformulate_disjunction(model, disj, method)
+        reformulate_disjunction(model, disj, method)
     end
 end
 
 # disjuncts
-function _reformulate_disjunction(model::JuMP.Model, disj::ConstraintData{T}, method::Union{BigM,Indicator}) where {T<:Disjunction}
-    for d in disj.constraint.disjuncts
+"""
+    reformulate_disjunction(
+        model::JuMP.Model, 
+        disj::ConstraintData{T},
+        method::AbstractReformulationMethod
+    ) where {T<:Disjunction}
+
+Reformulate a disjunction using the specified `method`. Current reformulation methods include
+`BigM`, `Hull`, and `Indicator`. This method can be extended for other reformualtion techniques.
+
+The `disj` field is the `ConstraintData` object for the disjunction, stored in the 
+`disjunctions` field of the `GDPData` object.
+"""
+function reformulate_disjunction(model::JuMP.Model, disj::ConstraintData{T}, method::BigM) where {T<:Disjunction}
+    for d in disj.constraint.indicators
         _reformulate_disjunct(model, d, method)
     end
 end
-function _reformulate_disjunction(model::JuMP.Model, disj::ConstraintData{T}, method::Hull) where {T<:Disjunction}
+function reformulate_disjunction(model::JuMP.Model, disj::ConstraintData{T}, method::Indicator) where {T<:Disjunction}
+    for d in disj.constraint.indicators
+        _reformulate_disjunct(model, d, method)
+    end
+end
+function reformulate_disjunction(model::JuMP.Model, disj::ConstraintData{T}, method::Hull) where {T<:Disjunction}
     disj_vrefs = _get_disjunction_variables(model, disj)
     _update_variable_bounds.(disj_vrefs)
     hull = _Hull(method.value, disj_vrefs)
-    for d in disj.constraint.disjuncts #reformulate each disjunct
+    for d in disj.constraint.indicators #reformulate each disjunct
         _disaggregate_variables(model, d, disj_vrefs, hull) #disaggregate variables for that disjunct
         _reformulate_disjunct(model, d, hull)
     end
@@ -52,275 +70,12 @@ function _reformulate_disjunct(model::JuMP.Model, ind_idx::LogicalVariableIndex,
     bvref = JuMP.VariableRef(model, bv_idx)
     for cidx in _indicator_to_constraints(model)[ind_idx]
         cdata = _disjunct_constraints(model)[cidx]
-        reformulate_disjunct_constraint(model, cdata.constraint, bvref, method)
+        _reformulate_disjunct_constraint(model, cdata.constraint, bvref, method)
     end
 end
 
-################################################################################
-#                              BIG-M REFORMULATION
-################################################################################
-function reformulate_disjunct_constraint(
-    model::JuMP.Model,
-    con::JuMP.ScalarConstraint{T, S}, 
-    bvref::JuMP.VariableRef,
-    method::BigM
-) where {T, S <: _MOI.LessThan}
-    M = _get_M_value(method, con.func, con.set)
-    new_func = JuMP.@expression(model, con.func - M*(1-bvref))
-    reform_con = JuMP.add_constraint(model, 
[…]
