### 2023-09-19 b61f701
add nested disjunction reformulations: - _reformulate_disjunct_constraint has a `nested` arg to indicate if reformulated constraint is added - disjunction is now a vector of logicvariablerefs to allow interrogating a nested constraint - gdp data includes a vector of disjunctionindices for the nested disjunctions
