+++ src/datatypes.jl
@@ -56,34 +56,34 @@ end
 ################################################################################
 #                        LOGICAL SELECTOR (CARDINALITY) SETS
 ################################################################################
-abstract type MOISelector <: _MOI.AbstractVectorSet end
+abstract type _MOISelector <: _MOI.AbstractVectorSet end
 
 """
-    MOIAtLeast{T<:Union{Int,LogicalVariableRef}} <: MOISelector
+    _MOIAtLeast{T<:Union{Int,LogicalVariableRef}} <: _MOISelector
 
 MOI level set for AtLeast constraints, see [`AtLeast`](@ref) for recommended syntax.
 """
-struct MOIAtLeast{T<:Union{Int,LogicalVariableRef}} <: MOISelector
+struct _MOIAtLeast{T<:Union{Int,LogicalVariableRef}} <: _MOISelector
     value::T
     dimension::Int
 end
 
 """
-    MOIAtMost{T<:Union{Int,LogicalVariableRef}} <: MOISelector
+    _MOIAtMost{T<:Union{Int,LogicalVariableRef}} <: _MOISelector
 
 MOI level set for AtMost constraints, see [`AtMost`](@ref) for recommended syntax.
 """
-struct MOIAtMost{T<:Union{Int,LogicalVariableRef}} <: MOISelector
+struct _MOIAtMost{T<:Union{Int,LogicalVariableRef}} <: _MOISelector
     value::T 
     dimension::Int
 end
 
 """
-    MOIExactly{T<:Union{Int,LogicalVariableRef}} <: MOISelector
+    _MOIExactly{T<:Union{Int,LogicalVariableRef}} <: _MOISelector
 
 MOI level set for Exactly constraints, see [`Exactly`](@ref) for recommended syntax.
 """
-struct MOIExactly{T<:Union{Int,LogicalVariableRef}} <: MOISelector
+struct _MOIExactly{T<:Union{Int,LogicalVariableRef}} <: _MOISelector
     value::T 
     dimension::Int
 end
@@ -92,7 +92,7 @@ end
 """
     AtLeast{T<:Union{Int,LogicalVariableRef}} <: JuMP.AbstractVectorSet
 
-Convenient alias for using [`MOIAtLeast`](@ref).
+Convenient alias for using [`_MOIAtLeast`](@ref).
 """
 struct AtLeast{T<:Union{Int,LogicalVariableRef}} <: JuMP.AbstractVectorSet
     value::T
@@ -101,7 +101,7 @@ end
 """
     AtMost{T<:Union{Int,LogicalVariableRef}} <: JuMP.AbstractVectorSet
 
-Convenient alias for using [`MOIAtMost`](@ref).
+Convenient alias for using [`_MOIAtMost`](@ref).
 """
 struct AtMost{T<:Union{Int,LogicalVariableRef}} <: JuMP.AbstractVectorSet
     value::T
@@ -110,16 +110,16 @@ end
 """
     Exactly <: JuMP.AbstractVectorSet
 
-Convenient alias for using [`MOIExactly`](@ref).
+Convenient alias for using [`_MOIExactly`](@ref).
 """
 struct Exactly{T<:Union{Int,LogicalVariableRef}} <: JuMP.AbstractVectorSet
     value::T 
 end
 
 # Extend JuMP.moi_set as needed
-JuMP.moi_set(set::AtLeast, dim::Int) = MOIAtLeast(set.value, dim)
-JuMP.moi_set(set::AtMost, dim::Int) = MOIAtMost(set.value, dim)
-JuMP.moi_set(set::Exactly, dim::Int) = MOIExactly(set.value, dim)
[…]
