+++ src/bigm.jl
@@ -107,86 +107,108 @@ function _reformulate_disjunct_constraint(
     model::JuMP.Model,
     con::JuMP.ScalarConstraint{T, S}, 
     bvref::JuMP.VariableRef,
-    method::BigM
+    method::BigM,
+    name::String,
+    nested::Bool
 ) where {T, S <: _MOI.LessThan}
     M = _get_M_value(method, con.func, con.set)
     new_func = JuMP.@expression(model, con.func - M*(1-bvref))
-    reform_con = JuMP.add_constraint(model, 
-        JuMP.build_constraint(error, new_func, con.set)
-    )
-    push!(_reformulation_constraints(model), 
-        (JuMP.index(reform_con), JuMP.ScalarShape())
-    )
+    reform_con = JuMP.build_constraint(error, new_func, con.set)
+    if !nested
+        con_name = isempty(name) ? name : string(name, "_", bvref)
+        _add_reformulated_constraint(model, reform_con, con_name)
+    end
+
+    return [reform_con]
 end
 function _reformulate_disjunct_constraint(
     model::JuMP.Model,
     con::JuMP.VectorConstraint{T, S, R}, 
     bvref::JuMP.VariableRef,
-    method::BigM
+    method::BigM,
+    name::String,
+    nested::Bool
 ) where {T, S <: _MOI.Nonpositives, R}
     M = [_get_M_value(method, func, con.set) for func in con.func]
     new_func = JuMP.@expression(model, [i=1:con.set.dimension], 
         con.func[i] - M[i]*(1-bvref)
     )
-    reform_con = JuMP.add_constraint(model,
-        JuMP.build_constraint(error, new_func, con.set)
-    )
-    push!(_reformulation_constraints(model), (JuMP.index(reform_con), JuMP.VectorShape()))
+    reform_con = JuMP.build_constraint(error, new_func, con.set)
+    if !nested
+        con_name = isempty(name) ? name : string(name, "_", bvref)
+        _add_reformulated_constraint(model, reform_con, con_name)
+    end
+
+    return [reform_con]
 end
 function _reformulate_disjunct_constraint(
     model::JuMP.Model, 
     con::JuMP.ScalarConstraint{T, S}, 
     bvref::JuMP.VariableRef,
     method::BigM,
+    name::String,
+    nested::Bool
 ) where {T, S <: _MOI.GreaterThan}
     M = _get_M_value(method, con.func, con.set)
     new_func = JuMP.@expression(model, con.func + M*(1-bvref))
-    reform_con = JuMP.add_constraint(model,
-        JuMP.build_constraint(error, new_func, con.set)
-    )
-    push!(_reformulation_constraints(model), (JuMP.index(reform_con), JuMP.ScalarShape()))
+    reform_con = JuMP.build_constraint(error, new_func, con.set)
+    if !nested
+        con_name = isempty(name) ? name : string(name, "_", bvref)
+        _add_reformulated_constraint(model, reform_con, con_name)
+    end
+
+    return [reform_con]
 end
 function _reformulate_disjunct_constraint(
     model::JuMP.Model, 
     con::JuMP.VectorConstraint{T, S, R}, 
     bvref::JuMP.VariableRef,
     method::BigM,
+    name::String,
+    nested::Bool
 ) where {T, S <: _MOI.Nonnegatives, R}
     M = [_get_M_value(method, func, con.set) for func in con.func]
     new_func = JuMP.@expression(model, [i=1:con.set.dimension], 
         con.func[i] + M[i]*(1-bvref)
     )
-    reform_con = JuMP.add_constraint(model,
[…]
