+++ src/reformulate.jl
@@ -106,9 +106,9 @@ function _reformulate_disjunct_constraint(
     ref_cons = reformulate_disjunction(model, disj, method, true)
     new_ref_cons = Vector{JuMP.AbstractConstraint}()
     for ref_con in ref_cons
-        push!(new_ref_cons, _reformulate_disjunct_constraint(model, ref_con, bvref, method, name, false)...)
+        push!(new_ref_cons, _reformulate_disjunct_constraint(model, ref_con, bvref, method, name, nested)...) #nested = false only for the parent disjunction (only adds the constraints when they have moved up all the way)
     end
-
+    
     return new_ref_cons
 end
 
