### 2023-09-20 72cd989
bug fix for nested indicator reformulation
