### 2023-09-20 4fb7e3d
clean nested indicator reformulation
