+++ src/indicator.jl
@@ -52,33 +52,4 @@ function _reformulate_disjunct_constraint(
     end
     
     return [con]
-    # # NOTE The following should not be required if the user adds the appropriate cardinality constraints to the nested disjunction indicators
-    # #create intermediate binary variable
-    # bvar = JuMP.@variable(model, binary = true, base_name = "$(bvref)_∧_$(con.func[1])")
-    # push!(_reformulation_variables(model), JuMP.index(bvar))
-    # #reformulate nested indicator constraint
-    # reform_cons = [
-    #     JuMP.build_constraint(error, [1*bvar, con.func[2]], con.set), # replace indicator variable with new bvar
-    #     #add logic constraints: bvar <=> bvref ∧ con.func[1]
-    #     JuMP.build_constraint(error, # {bvref ∧ con.func[1]} => bvar
-    #         JuMP.@expression(model, (1 - bvref) + (1 - con.func[1]) + bvar),
-    #         _MOI.GreaterThan(1)
-    #     ),
-    #     JuMP.build_constraint(error, # bvar => {bvref ∧ con.func[1]}
-    #         JuMP.@expression(model, (1 - bvar) + bvref),
-    #         _MOI.GreaterThan(1)
-    #     ),
-    #     JuMP.build_constraint(error, # bvar => {bvref ∧ con.func[1]}
-    #         JuMP.@expression(model, (1 - bvar) + con.func[1]),
-    #         _MOI.GreaterThan(1)
-    #     )
-    # ]
-    # for (i, reform_con) in enumerate(reform_cons)
-    #     if !nested || i >= 2 # always add the logic constraints to the model
-    #         con_name = isempty(name) ? name : string(name, "_", bvref)
-    #         _add_reformulated_constraint(model, reform_con, con_name)
-    #     end
-    # end
-
-    # return [reform_cons[1]]
 end
