### 2023-09-20 a416bc5
remove copy methods
