### 2023-09-20 8ec0573
improve nested disjunction code
