+++ src/hull.jl
@@ -1,29 +1,33 @@
 ################################################################################
 #                              VARIABLE DISAGGREGATION
 ################################################################################
+function _query_variable_bounds(model::JuMP.Model, method::Hull)
+    for var in JuMP.all_variables(model)
+        method.variable_bounds[var] = _update_variable_bounds(var)
+    end
+end
+
 function _update_variable_bounds(vref::JuMP.VariableRef)
-    model, idx = vref.model, vref.index
-    if (vref in keys(_variable_bounds(model))) || JuMP.is_binary(vref) #skip if binary or if bounds already stored
-        return
+    if JuMP.is_binary(vref) #not used
+        lb, ub = 0, 1
     elseif !JuMP.has_lower_bound(vref) || !JuMP.has_upper_bound(vref)
         error("Variable $vref must have both lower and upper bounds defined when using the Hull reformulation.")
     else
         lb = min(0, JuMP.lower_bound(vref))
         ub = max(0, JuMP.upper_bound(vref))
-        _variable_bounds(model)[idx] = (lb, ub)
     end
+    return lb, ub
 end
-function _disaggregate_variables(model::JuMP.Model, lvref::LogicalVariableRef, vrefs::Set{JuMP.VariableRef}, disag_vars::_Hull)
+function _disaggregate_variables(model::JuMP.Model, lvref::LogicalVariableRef, vrefs::Set{JuMP.VariableRef}, method::_Hull)
     #create disaggregated variables for that disjunct
     for vref in vrefs
         JuMP.is_binary(vref) && continue #skip binary variables
-        _disaggregate_variable(model, lvref, vref, disag_vars) #create disaggregated var for that disjunct
+        _disaggregate_variable(model, lvref, vref, method) #create disaggregated var for that disjunct
     end
 end
-function _disaggregate_variable(model::JuMP.Model, lvref::LogicalVariableRef, vref::JuMP.VariableRef, disag_vars::_Hull)
-    #create disaggregated (disjunct) vref
-    v_idx = JuMP.index(vref) #variable
-    lb, ub = _variable_bounds(model)[v_idx]
+function _disaggregate_variable(model::JuMP.Model, lvref::LogicalVariableRef, vref::JuMP.VariableRef, method::_Hull)
+    #create disaggregated vref
+    lb, ub = method.variable_bounds[vref]
     lv_idx = JuMP.index(lvref)
     dvref = JuMP.@variable(model, base_name = "$(vref)_$(lvref)", lower_bound = lb, upper_bound = ub)
     dv_idx = JuMP.index(dvref) #disaggregated (disjunct) variable
@@ -32,8 +36,11 @@ function _disaggregate_variable(model::JuMP.Model, lvref::LogicalVariableRef, vr
     bv_idx = _indicator_to_binary(model)[lv_idx]
     bvref = JuMP.VariableRef(model, bv_idx)
     #temp storage
-    push!(disag_vars.disjunction[vref], dvref)
-    disag_vars.disjunct[vref, bvref] = dvref
+    if !haskey(method.disjunction_variables, vref)
+        method.disjunction_variables[vref] = Vector{JuMP.VariableRef}()
+    end
+    push!(method.disjunction_variables[vref], dvref)
+    method.disjunct_variables[vref, bvref] = dvref
     #create bounding constraints
     dvname = JuMP.name(dvref)
     new_con_lb_ref = JuMP.add_constraint(model,
@@ -55,15 +62,10 @@ end
 ################################################################################
[…]
