+++ src/optimize.jl
@@ -9,28 +9,3 @@ function _optimize_hook(
     end
     return JuMP.optimize!(model; ignore_optimize_hook = true)
 end
-
-"""
-    reformulate_model(model::JuMP.Model, method::AbstractSolutionMethod)
-
-Reformulate a `GDPModel` using the specified `method`. Prior to reformulation,
-all previous reformulation variables and constraints are deleted.
-"""
-function reformulate_model(model::JuMP.Model, method::AbstractSolutionMethod)
-    #clear all previous reformulations
-    for (cidx, cshape) in _reformulation_constraints(model)
-        JuMP.delete(model, JuMP.ConstraintRef(model, cidx, cshape))
-    end
-    empty!(gdp_data(model).reformulation_constraints)
-    for vidx in _reformulation_variables(model)
-        JuMP.delete(model, JuMP.VariableRef(model, vidx))
-    end
-    empty!(gdp_data(model).reformulation_variables)
-    #reformulate
-    _reformulate_logical_variables(model)
-    _reformulate_disjunctions(model, method)
-    _reformulate_logical_constraints(model)
-    #set solution method
-    _set_solution_method(model, method)
-    _set_ready_to_optimize(model, true)
-end

[…]
