+++ src/reformulate.jl
@@ -20,13 +20,9 @@ function reformulate_model(model::JuMP.Model, method::AbstractSolutionMethod)
 end
 
 function _clear_reformulations(model::JuMP.Model)
-    for (cidx, cshape) in _reformulation_constraints(model)
-        JuMP.delete(model, JuMP.ConstraintRef(model, cidx, cshape))
-    end
+    JuMP.delete.(model, _reformulation_constraints(model))
+    JuMP.delete.(model, _reformulation_variables(model))
     empty!(gdp_data(model).reformulation_constraints)
-    for vidx in _reformulation_variables(model)
-        JuMP.delete(model, JuMP.VariableRef(model, vidx))
-    end
     empty!(gdp_data(model).reformulation_variables)
 end
 
@@ -42,9 +38,8 @@ function _reformulate_logical_variables(model::JuMP.Model)
         if JuMP.is_fixed(lvref)
             JuMP.fix(bvref, JuMP.fix_value(lvref))
         end
-        bv_idx = JuMP.index(bvref)
-        push!(_reformulation_variables(model), bv_idx)
-        _indicator_to_binary(model)[lv_idx] = bv_idx
+        push!(_reformulation_variables(model), bvref)
+        _indicator_to_binary(model)[lvref] = bvref
     end
 end
 
@@ -59,7 +54,7 @@ function _reformulate_all_disjunctions(model::JuMP.Model, method::AbstractReform
         for (i, ref_con) in enumerate(ref_cons)
             name = isempty(disj.name) ? "" : string(disj.name,"_$i")
             cref = JuMP.add_constraint(model, ref_con, name)
-            push!(_reformulation_constraints(model), (JuMP.index(cref), _map_con_shape(ref_con)))
+            push!(_reformulation_constraints(model), cref)
         end
     end
 end
@@ -114,15 +109,13 @@ function reformulate_disjunction(model::JuMP.Model, disj::Disjunction, method::_
 end
 
 # individual disjuncts
-function _reformulate_disjunct(model::JuMP.Model, ind_ref::LogicalVariableRef, method::AbstractReformulationMethod)
+function _reformulate_disjunct(model::JuMP.Model, lvref::LogicalVariableRef, method::AbstractReformulationMethod)
     #reformulate each constraint and add to the model
-    lv_idx = JuMP.index(ind_ref)
-    bv_idx = _indicator_to_binary(model)[lv_idx]
-    bvref = JuMP.VariableRef(model, bv_idx)
+    bvref = _indicator_to_binary(model)[lvref]
     ref_cons = Vector{JuMP.AbstractConstraint}()
-    for cidx in _indicator_to_constraints(model)[lv_idx]
-        cdata = _index_to_constraint(model, cidx)
-        push!(ref_cons, reformulate_disjunct_constraint(model, cdata.constraint, bvref, method)...)
+    for cref in _indicator_to_constraints(model)[lvref]
+        con = JuMP.constraint_object(cref)
+        push!(ref_cons, reformulate_disjunct_constraint(model, con, bvref, method)...)
     end
 
     return ref_cons
@@ -155,15 +148,6 @@ function reformulate_disjunct_constraint(
     method::AbstractReformulationMethod
 )
     error("$method reformulation for constraint $con is not supported yet.")
-end
-
-function _add_reformulated_constraint(
-    model::JuMP.Model,
-    con::JuMP.AbstractConstraint,
-    name::String
-)
-    reform_con = JuMP.add_constraint(model, con, name)
-    push!(_reformulation_constraints(model), (JuMP.index(reform_con), _map_con_shape(con)))
 end
 
 ################################################################################
