+++ examples/ex6.jl
@@ -27,5 +27,5 @@ m = GDPModel()
 @constraint(m, z in Exactly(w[1]))
 
 ##
-reformulate_model(m, Indicator())
+reformulate_model(m, BigM())
 print(m)
