+++ src/constraints.jl
@@ -318,6 +318,7 @@ function _disjunction(
     dref = _create_disjunction(_error, model, structure, name, true)
     obj = JuMP.constraint_object(dref)
     _add_indicator_var(_DisjunctConstraint(obj, tag.indicator), dref, model)
+    return dref
 end
 
 # General fallback for additional arguments
