### 2023-09-21 94f832b
Update src/constraints.jl
Co-authored-by: Joshua Pulsipher <33848311+pulsipher@users.noreply.github.com>
