### 2023-09-21 27b95c9
Update src/reformulate.jl
Co-authored-by: Joshua Pulsipher <33848311+pulsipher@users.noreply.github.com>
