### 2023-09-21 2782434
Update src/indicator.jl
Co-authored-by: Joshua Pulsipher <33848311+pulsipher@users.noreply.github.com>
