+++ src/indicator.jl
@@ -20,12 +20,7 @@ function reformulate_disjunct_constraint(
     method::Indicator
 ) where {T, S}
     set = _vec_to_scalar_set(con.set)
-    reform_cons = Vector{JuMP.VectorConstraint}()
-    for (i, func) in enumerate(con.func)
-        reform_con = JuMP.build_constraint(error, [1*bvref, func], _MOI.Indicator{_MOI.ACTIVATE_ON_ONE}(set))
-        push!(reform_cons, reform_con)
-    end
-
+    return [JuMP.build_constraint(error, [1*bvref, f], _MOI.Indicator{_MOI.ACTIVATE_ON_ONE}(set)) for f in con.func]
     return reform_cons
 end
 #nested indicator reformulation
