### 2023-09-21 9aa317a
check only continuous vars used in disj constr
