+++ src/DisjunctiveProgramming.jl
@@ -19,7 +19,6 @@ include("reformulate.jl")
 include("bigm.jl")
 include("hull.jl")
 include("indicator.jl")
-include("optimize.jl")
 
 # Define additional stuff that should not be exported
 const _EXCLUDE_SYMBOLS = [Symbol(@__MODULE__), :eval, :include]
