+++ test/Project.toml
@@ -1,5 +1,0 @@
-[deps]
-Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
-HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
-JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
