### 2023-09-22 b999c1d
Update reformulate.jl
Co-authored-by: Joshua Pulsipher <33848311+pulsipher@users.noreply.github.com>
