+++ test/runtests.jl
@@ -7,3 +7,4 @@ const DP = DisjunctiveProgramming
 include("aqua.jl")
 include("logical_variables.jl")
 include("model.jl")
+include("variable_interrogation.jl")
