### 2023-09-22 5667bae
add solve example tests
