+++ test/runtests.jl
@@ -8,3 +8,4 @@ include("aqua.jl")
 include("logical_variables.jl")
 include("model.jl")
 include("variable_interrogation.jl")
+include("solve.jl")
