### 2023-09-22 7e839fb
fix MOI sets for selectors
