+++ test/runtests.jl
@@ -5,7 +5,7 @@ using Test
 const DP = DisjunctiveProgramming
 
 include("aqua.jl")
-include("logical_variables.jl")
+include("variables/logical.jl")
+include("variables/query.jl")
 include("model.jl")
-include("variable_interrogation.jl")
 include("solve.jl")
