### 2023-09-28 6aea798
IsTrue() for propositions; add tests
