+++ test/runtests.jl
@@ -5,7 +5,8 @@ using Test
 const DP = DisjunctiveProgramming
 
 include("aqua.jl")
-include("variables/logical.jl")
 include("variables/query.jl")
+include("variables/logical.jl")
+include("constraints/logical.jl")
 include("model.jl")
 include("solve.jl")

[…]
