### 2023-09-28 1a45fc3
add some logic constraint tests
