+++ test/variables/query.jl
@@ -91,7 +91,7 @@ function test_interrogate_proposition_constraint()
     @variable(m, w[1:5], LogicalVariable)
     ex = (implies(w[1], w[2]) ∧ w[3]) ⇔ (¬w[4] ∨ y)
     @constraint(m, con, ex in IsTrue())
-    obj = JuMP.constraint_object(con)
+    obj = constraint_object(con)
     vars = DP._get_constraint_variables(m, obj)
     @test w[1] in vars
     @test w[2] in vars
@@ -107,7 +107,7 @@ function test_interrogate_selector_constraint()
     @variable(m, y, LogicalVariable)
     @variable(m, w[1:5], LogicalVariable)
     @constraint(m, con, w[1:4] in AtMost(y))
-    obj = JuMP.constraint_object(con)
+    obj = constraint_object(con)
     vars = DP._get_constraint_variables(m, obj)
     @test w[1] in vars
     @test w[2] in vars
