+++ test/constraints/bigm.jl
@@ -3,6 +3,6 @@ function test_eq_scalar_affine_bigm()
     @variable(model, x)
     @variable(model, y, LogicalVariable)
     @constraint(model, con, 3*x == 1, DisjunctConstraint(y))
-    _reformulate_logical_variables(model)
+    
     
 end
