+++ test/constraints/proposition.jl
@@ -222,8 +222,10 @@ function test_lvar_cnf_functions()
     @variable(model, y, LogicalVariable)
     @test DP._eliminate_equivalence(y) == y
     @test DP._eliminate_implication(y) == y
-    @test DP._move_negation_inward(y) == y
-    @test DP._negate(y) == ¬y
+    @test DP._move_negations_inward(y) == y
+    neg_y = DP._negate(y)
+    @test neg_y.head == :¬
+    @test neg_y.args[1] == y
     @test DP._distribute_and_over_or(y) == y
     @test DP._flatten(y) == y
 end
@@ -237,8 +239,8 @@ function test_eliminate_equivalence()
     @test length(new_ex.args) == 2
     @test new_ex.args[1].head == :⟹
     @test new_ex.args[2].head == :⟹
-    @test new_ex.args[1].args == y
-    @test new_ex.args[2].args == y
+    @test Set(new_ex.args[1].args) == Set{Any}(y)
+    @test Set(new_ex.args[2].args) == Set{Any}(y)
 end
 
 function test_eliminate_equivalence_flat()
@@ -432,4 +434,79 @@ function test_to_cnf()
     @variable(model, y[1:3], LogicalVariable)
     ex = iff(y...)
     new_ex = DP._to_cnf(ex)
+    @test new_ex.head == :∧
+    for arg in new_ex.args
+        @test arg.head == :∨
+    end
+    @test (y[1] in new_ex.args[1].args && y[2] in new_ex.args[1].args && y[3] in new_ex.args[1].args) ||
+        (y[1] in new_ex.args[2].args && y[2] in new_ex.args[2].args && y[3] in new_ex.args[2].args) ||
+        (y[1] in new_ex.args[3].args && y[2] in new_ex.args[3].args && y[3] in new_ex.args[3].args) ||
+        (y[1] in new_ex.args[4].args && y[2] in new_ex.args[4].args && y[3] in new_ex.args[4].args) ||
+        (y[1] in new_ex.args[5].args && y[2] in new_ex.args[5].args && y[3] in new_ex.args[5].args) ||
+        (y[1] in new_ex.args[6].args && y[2] in new_ex.args[6].args && y[3] in new_ex.args[6].args)
+
+    @test (y[1] in new_ex.args[1].args && !(y[2] in new_ex.args[1].args) && !(y[3] in new_ex.args[1].args)) ||
+        (y[1] in new_ex.args[2].args && !(y[2] in new_ex.args[2].args) && !(y[3] in new_ex.args[2].args)) ||
+        (y[1] in new_ex.args[3].args && !(y[2] in new_ex.args[3].args) && !(y[3] in new_ex.args[3].args)) ||
+        (y[1] in new_ex.args[4].args && !(y[2] in new_ex.args[4].args) && !(y[3] in new_ex.args[4].args)) ||
+        (y[1] in new_ex.args[5].args && !(y[2] in new_ex.args[5].args) && !(y[3] in new_ex.args[5].args)) ||
+        (y[1] in new_ex.args[6].args && !(y[2] in new_ex.args[6].args) && !(y[3] in new_ex.args[6].args))
+
+    @test (!(y[1] in new_ex.args[1].args) && y[2] in new_ex.args[1].args && !(y[3] in new_ex.args[1].args)) ||
+        (!(y[1] in new_ex.args[2].args) && y[2] in new_ex.args[2].args && !(y[3] in new_ex.args[2].args)) ||
+        (!(y[1] in new_ex.args[3].args) && y[2] in new_ex.args[3].args && !(y[3] in new_ex.args[3].args)) ||
[…]
