### 2023-09-29 b5a69b4
add test for disjunctions
