+++ test/disjunction.jl
@@ -1,0 +1,220 @@
+function test_disjunction_add_fail()
+    model = GDPModel()
+    @variable(model, x)
+    @variable(model, y[1:2], LogicalVariable)
+    @constraint(model, x == 5, DisjunctConstraint(y[1]))
+
+    @test_macro_throws ErrorException @disjunction(model) #not enough arguments
+    @test_macro_throws UndefVarError @disjunction(model, y) #unassociated indicator
+    @test_macro_throws UndefVarError @disjunction(GDPModel(), y) #wrong model
+    @test_macro_throws ErrorException @disjunction(Model(), y) #not a GDPModel
+    @test_macro_throws UndefVarError @disjunction(model, [y[1], y[1]]) #duplicate indicator
+    @test_macro_throws ErrorException @disjunction(model, y[1]) #no disjunction expression
+    @test_macro_throws UndefVarError @disjunction(model, y, "random_arg") #unrecognized extra argument
+    @test_macro_throws UndefVarError @disjunction(model, "ABC") #unrecognized structure
+    @test_macro_throws ErrorException @disjunction(model, begin y end) #@disjunctions (plural)
+    @test_macro_throws UndefVarError @disjunction(model, x, y) #name x already exists
+
+    @constraint(model, x == 10, DisjunctConstraint(y[2]))
+    @disjunction(model, disj, y)
+    @test_macro_throws UndefVarError @disjunction(model, disj, y) #duplicate name
+
+    # "Invalid name $name"
+
+    # "Expression $name should not be used as a disjunction name. Use " *
+    #            "the \"anonymous\" syntax $name = @disjunction(model, " *
+    #            "...) instead."
+    @test_macro_throws ErrorException @disjunction(model, [model=1:2], y) #index name can't be same as model name
+    # "Index $(model) is the same symbol as the model. Use a ",
+    # "different name for the index."
+end
+
+function test_disjunction_add_success()
+    model = GDPModel()
+    @variable(model, x)
+    @variable(model, y[1:2], LogicalVariable)
+    @constraint(model, x == 5, DisjunctConstraint(y[1]))
+    @constraint(model, x == 10, DisjunctConstraint(y[2]))
+    disj = @disjunction(model, y)
+    @disjunction(model, disj2, y)
+    @test owner_model(disj) == model
+    @test is_valid(model, disj)
+    @test index(disj) == DisjunctionIndex(1)
+    @test name(disj) == ""
+    @test name(disj2) == "disj2"
+    @test haskey(DP._disjunctions(model), index(disj))
+    @test haskey(DP._disjunctions(model), index(disj2))
+    @test DP._disjunctions(model)[index(disj)] == DP._constraint_data(disj)
+    @test !constraint_object(disj).nested
+    @test constraint_object(disj).indicators == y
+    @test disj == copy(disj)
+end
+
+function test_disjunction_add_nested()
+    model = GDPModel()
+    @variable(model, x)
+    @variable(model, y[1:2], LogicalVariable)
+    @variable(model, z[1:2], LogicalVariable)
+    @constraint(model, x <= 5, DisjunctConstraint(y[1]))
+    @constraint(model, x >= 5, DisjunctConstraint(y[2]))
+    @disjunction(model, inner, y, DisjunctConstraint(z[1]))
+    @constraint(model, x <= 10, DisjunctConstraint(z[1]))
+    @constraint(model, x >= 10, DisjunctConstraint(z[2]))
+    @disjunction(model, outer, z)
+
+    @test is_valid(model, inner)
[…]
