### 2023-09-30 a6d34a4
Fix disjunction macro logic
