### 2023-10-01 02af76a
generalize disjunction for containers of logicalvariablerefs
