### 2023-10-02 040fc15
Clean up and simply macro
