+++ src/macros.jl
@@ -181,11 +181,11 @@ macro disjunction(model, args...)
         _error("Invalid syntax. Did you mean to use `@disjunctions`?")
     end
 
-# TODO: three cases lead to problems when julia variables are used for DisjunctConstraint tags
-# which violate the cases considered in the table further below. The three cases are
-# (i) @disjunction(m, Y[1, :], tag[1]) --> gets confused for @disjunction(m, name[...], Y[1, :]) (Case 2 below)
-# (ii) @disjunction(m, Y, tagref) --> gets confused for @disjunction(m, name, Y) (Case 1 below)
-# (iii) @disjunction(m, Y[1, :], tagref) --> gets confused for @disjunction(m, name[...], Y) (Case 2 below)
+    # TODO: three cases lead to problems when julia variables are used for DisjunctConstraint tags
+    # which violate the cases considered in the table further below. The three cases are
+    # (i) @disjunction(m, Y[1, :], tag[1]) --> gets confused for @disjunction(m, name[...], Y[1, :]) (Case 2 below)
+    # (ii) @disjunction(m, Y, tagref) --> gets confused for @disjunction(m, name, Y) (Case 1 below)
+    # (iii) @disjunction(m, Y[1, :], tagref) --> gets confused for @disjunction(m, name[...], Y) (Case 2 below)
 
     # Determine if a reference/container argument was given by the user
     # There are 9 cases to consider:
