### 2023-10-09 762a877
get variable bounds up front for BigM and store This allows customizations when nesting
