+++ src/bigm.jl
@@ -2,18 +2,18 @@
 #                              BIG-M VALUE
 ################################################################################
 # Get Big-M value for a particular constraint
-function _get_M_value(method::BigM, func::JuMP.AbstractJuMPScalar, set::_MOI.AbstractSet)
+function _get_M_value(func::JuMP.AbstractJuMPScalar, set::_MOI.AbstractSet, method::BigM)
     if method.tighten
-        M = _get_tight_M(method, func, set)
+        M = _get_tight_M(func, set, method)
     else
-        M = _get_M(method, func, set)
+        M = _get_M(func, set, method)
     end
     return M
 end
 
 # Get the tightest Big-M value for a particular constraint
-function _get_tight_M(method::BigM, func::JuMP.AbstractJuMPScalar, set::_MOI.AbstractSet)
-    M = min.(method.value, _calculate_tight_M(func, set)) #broadcast for when S <: MOI.Interval or MOI.EqualTo or MOI.Zeros
+function _get_tight_M(func::JuMP.AbstractJuMPScalar, set::_MOI.AbstractSet, method::BigM)
+    M = min.(method.value, _calculate_tight_M(func, set, method)) #broadcast for when S <: MOI.Interval or MOI.EqualTo or MOI.Zeros
     if any(isinf.(M))
         error("A finite Big-M value must be used. The value obtained was $M.")
     end
@@ -21,14 +21,14 @@ function _get_tight_M(method::BigM, func::JuMP.AbstractJuMPScalar, set::_MOI.Abs
 end
 
 # Get user-specified Big-M value
-function _get_M(method::BigM, ::JuMP.AbstractJuMPScalar, ::Union{_MOI.LessThan, _MOI.GreaterThan, _MOI.Nonnegatives, _MOI.Nonpositives})
+function _get_M(::JuMP.AbstractJuMPScalar, ::Union{_MOI.LessThan, _MOI.GreaterThan, _MOI.Nonnegatives, _MOI.Nonpositives}, method::BigM)
     M = method.value
     if isinf(M)
         error("A finite Big-M value must be used. The value given was $M.")
     end
     return M
 end
-function _get_M(method::BigM, ::JuMP.AbstractJuMPScalar, ::Union{_MOI.Interval, _MOI.EqualTo, _MOI.Zeros})
+function _get_M(::JuMP.AbstractJuMPScalar, ::Union{_MOI.Interval, _MOI.EqualTo, _MOI.Zeros}, method::BigM)
     M = method.value
     if isinf(M)
         error("A finite Big-M value must be used. The value given was $M.")
@@ -37,64 +37,79 @@ function _get_M(method::BigM, ::JuMP.AbstractJuMPScalar, ::Union{_MOI.Interval, 
 end
 
 # Apply interval arithmetic on a linear constraint to infer the tightest Big-M value from the bounds on the constraint.
-function _calculate_tight_M(func::JuMP.AffExpr, set::_MOI.LessThan)
-    return _interval_arithmetic_LessThan(func, -set.upper)
+function _calculate_tight_M(func::JuMP.AffExpr, set::_MOI.LessThan, method::BigM)
+    return _interval_arithmetic_LessThan(func, -set.upper, method)
 end
-function _calculate_tight_M(func::JuMP.AffExpr, set::_MOI.GreaterThan)
-    return _interval_arithmetic_GreaterThan(func, -set.lower)
+function _calculate_tight_M(func::JuMP.AffExpr, set::_MOI.GreaterThan, method::BigM)
+    return _interval_arithmetic_GreaterThan(func, -set.lower, method)
 end
[…]
