+++ test/constraints/hull.jl
@@ -1,0 +1,538 @@
+function test_default_hull()
+    method = Hull()
+    @test method.value == 1e-6
+end
+
+function test_set_hull()
+    method = Hull(0.001)
+    @test method.value == 0.001
+end
+
+function test_query_variable_bounds()
+    model = GDPModel()
+    @variable(model, 10 <= x <= 100)
+    @variable(model, -100 <= y <= -10)
+    method = Hull()
+    DP._query_variable_bounds(model, method)
+    @test haskey(method.variable_bounds, x)
+    @test haskey(method.variable_bounds, y)
+    @test method.variable_bounds[x] == (0, 100)
+    @test method.variable_bounds[y] == (-100, 0)
+end
+
+function test_query_variable_bounds_error1()
+    model = GDPModel()
+    @variable(model, x <= 100)
+    method = Hull()
+    @test_throws ErrorException DP._query_variable_bounds(model, method)
+end
+
+function test_query_variable_bounds_error2()
+    model = GDPModel()
+    @variable(model, -100 <= x)
+    method = Hull()
+    @test_throws ErrorException DP._query_variable_bounds(model, method)
+end
+
+function test_disaggregate_variables()
+    model = GDPModel()
+    @variable(model, 10 <= x <= 100)
+    @variable(model, y, Bin)
+    @variable(model, z, LogicalVariable)
+    vrefs = Set([x,y])
+    DP._reformulate_logical_variables(model)
+    method = DP._Hull(Hull(1e-3, Dict(x => (0., 100.))), vrefs)
+    DP._disaggregate_variables(model, z, vrefs, method)
+
+    refvars = DP._reformulation_variables(model)
+    @test length(refvars) == 2
+    zbin = variable_by_name(model, "z")
+    @test zbin in refvars
+    x_z = variable_by_name(model, "x_z")
+    @test x_z in refvars
+    @test has_lower_bound(x_z) && lower_bound(x_z) == 0
+    @test has_upper_bound(x_z) && upper_bound(x_z) == 100
+    @test name(x_z) == "x_z"
+    x_z_lb = constraint_by_name(model, "x_z_lower_bound") |> constraint_object
+    @test x_z_lb.func == -x_z
+    @test x_z_lb.set == MOI.LessThan(0.)
+    x_z_ub = constraint_by_name(model, "x_z_upper_bound") |> constraint_object
+    @test x_z_ub.func == -100zbin + x_z
+    @test x_z_ub.set == MOI.LessThan(0.)
+end
+
+function test_aggregate_variable()
+    model = GDPModel()
+    @variable(model, 10 <= x <= 100)
+    @variable(model, z, LogicalVariable)
+    vrefs = Set([x])
+    DP._reformulate_logical_variables(model)
+    method = DP._Hull(Hull(1e-3, Dict(x => (0., 100.))), vrefs)
+    DP._disaggregate_variables(model, z, vrefs, method)
+    refcons = Vector{JuMP.AbstractConstraint}() 
+    DP._aggregate_variable(model, refcons, x, method)
+    @test length(refcons) == 1
+    @test refcons[1].func == -x + sum(method.disjunction_variables[x])
+    @test refcons[1].set == MOI.EqualTo(0.)
+end
+
+function test_disaggregate_expression_affine()
+    model = GDPModel()
+    @variable(model, 10 <= x <= 100)
+    @variable(model, z, LogicalVariable)
+    DP._reformulate_logical_variables(model)
[…]
