+++ test/constraints/bigm.jl
@@ -1,8 +1,329 @@
-function test_eq_scalar_affine_bigm()
+function test_default_bigm()
+    method = BigM()
+    @test method.value == 1e9
+    @test method.tighten
+end
+
+function test_default_tighten_bigm()
+    method = BigM(100)
+    @test method.value == 100
+    @test method.tighten
+end
+
+function test_set_bigm()
+    method = BigM(1e6, false)
+    @test method.value == 1e6
+    @test !method.tighten
+end
+
+function test_get_M_1sided()
+    model = GDPModel()
+    @variable(model, x)
+    @variable(model, y, LogicalVariable)
+    @constraint(model, con, 3*x <= 1, DisjunctConstraint(y))
+    cobj = constraint_object(con)
+    M = DP._get_M(cobj.func, cobj.set, BigM(100, false))
+    @test M == 100
+    @test_throws ErrorException DP._get_M(cobj.func, cobj.set, BigM(Inf, false))
+end
+
+function test_get_tight_M_1sided()
+    model = GDPModel()
+    @variable(model, x)
+    @variable(model, y, LogicalVariable)
+    @constraint(model, con, 3*x <= 1, DisjunctConstraint(y))
+    cobj = constraint_object(con)
+
+    method = BigM(100)
+    DP._query_variable_bounds(model, method)
+    M = DP._get_tight_M(cobj.func, cobj.set, method)
+    @test M == 100
+
+    method = BigM(Inf)
+    DP._query_variable_bounds(model, method)
+    @test_throws ErrorException DP._get_tight_M(cobj.func, cobj.set, method)
+
+    set_upper_bound(x, 10)
+    method = BigM()
+    DP._query_variable_bounds(model, method)
+    M = DP._get_tight_M(cobj.func, cobj.set, method)
+    @test M == 29
+end
+
+function test_get_M_2sided()
     model = GDPModel()
     @variable(model, x)
     @variable(model, y, LogicalVariable)
     @constraint(model, con, 3*x == 1, DisjunctConstraint(y))
+    cobj = constraint_object(con)
+
+    method = BigM(100)
+    M = DP._get_M(cobj.func, cobj.set, method)
+    @test M[1] == 100
+    @test M[2] == 100
+
+    method = BigM(Inf)
+    @test_throws ErrorException DP._get_M(cobj.func, cobj.set, method)
+end
+
+function test_get_tight_M_2sided()
+    model = GDPModel()
+    @variable(model, x)
+    @variable(model, y, LogicalVariable)
+    @constraint(model, con, 3*x == 1, DisjunctConstraint(y))
+    cobj = constraint_object(con)
     
-    
+    method = BigM(100)
+    DP._query_variable_bounds(model, method)
+    M = DP._get_tight_M(cobj.func, cobj.set, method)
+    @test M[1] == 100
+    @test M[2] == 100
+
+    method = BigM(Inf)
+    DP._query_variable_bounds(model, method)
+    @test_throws ErrorException DP._get_tight_M(cobj.func, cobj.set, method)
+
+    set_lower_bound(x, -10)
+    set_upper_bound(x, 10)
+    method = BigM()
+    DP._query_variable_bounds(model, method)
+    M = DP._get_tight_M(cobj.func, cobj.set, method)
+    @test M[1] == 31
+    @test M[2] == 29
+end
+
+function test_interval_arithmetic_LessThan()
+    model = Model()
+    @variable(model, x[1:3])
+    @variable(model, y, Bin)
[…]
