+++ test/runtests.jl
@@ -32,6 +32,9 @@ include("variables/query.jl")
 include("variables/logical.jl")
 include("constraints/selector.jl")
 include("constraints/proposition.jl")
+include("constraints/disjunct.jl")
 include("constraints/indicator.jl")
+include("constraints/bigm.jl")
+include("constraints/hull.jl")
 include("disjunction.jl")
 include("solve.jl")
