### 2023-10-10 4136a4a
support empty disjuncts
