### 2023-10-10 ad5c87c
update readme and ex2
