+++ examples/ex2.jl
@@ -1,23 +1,23 @@
-# https://optimization.mccormick.northwestern.edu/index.php/Disjunctive_inequalities
+# https://optimization.cbe.cornell.edu/index.php?title=Disjunctive_inequalities#Big-M_Reformulation[1][2]
 using JuMP
 using DisjunctiveProgramming
 using HiGHS
 
 m = GDPModel(HiGHS.Optimizer)
-@variable(m, -5 ≤ x[1:2] ≤ 10)
+@variable(m, 0 ≤ x[1:2] ≤ 20)
 @variable(m, Y[1:2], LogicalVariable)
-@constraint(m, [i = 1:2], 0 ≤ x[i] ≤ [3,4][i], DisjunctConstraint(Y[1]))
-@constraint(m, [i = 1:2], [5,4][i] ≤ x[i] ≤ [9,6][i], DisjunctConstraint(Y[2]))
+@constraint(m, [i = 1:2], [2,5][i] ≤ x[i] ≤ [6,9][i], DisjunctConstraint(Y[1]))
+@constraint(m, [i = 1:2], [8,10][i] ≤ x[i] ≤ [11,15][i], DisjunctConstraint(Y[2]))
 @disjunction(m, Y)
 @constraint(m, Y in Exactly(1)) #logical constraint
 @objective(m, Max, sum(x))
 print(m)
 # Max x[1] + x[2]
 # Subject to
-#  x[1] ≥ -5
-#  x[2] ≥ -5
-#  x[1] ≤ 10
-#  x[2] ≤ 10
+#  x[1] ≥ 0
+#  x[2] ≥ 0
+#  x[1] ≤ 20
+#  x[2] ≤ 20
 
 ##
 optimize!(m, method = BigM(100, false)) #specify M value and disable M-tightening
@@ -25,18 +25,18 @@ print(m)
 # Max x[1] + x[2]
 # Subject to
 #  Y[1] + Y[2] = 1
-#  x[1] - 100 Y[1] ≥ -100
-#  x[2] - 100 Y[1] ≥ -100
-#  x[1] - 100 Y[2] ≥ -95
-#  x[2] - 100 Y[2] ≥ -96
-#  x[1] + 100 Y[1] ≤ 103
-#  x[2] + 100 Y[1] ≤ 104
-#  x[1] + 100 Y[2] ≤ 109
-#  x[2] + 100 Y[2] ≤ 106
-#  x[1] ≥ -5
-#  x[2] ≥ -5
-#  x[1] ≤ 10
-#  x[2] ≤ 10
+#  x[1] - 100 Y[1] ≥ -98
+#  x[2] - 100 Y[1] ≥ -95
+#  x[1] - 100 Y[2] ≥ -92
+#  x[2] - 100 Y[2] ≥ -90
+#  x[1] + 100 Y[1] ≤ 106
+#  x[2] + 100 Y[1] ≤ 109
+#  x[1] + 100 Y[2] ≤ 111
+#  x[2] + 100 Y[2] ≤ 115
+#  x[1] ≥ 0
+#  x[2] ≥ 0
+#  x[1] ≤ 20
+#  x[2] ≤ 20
 #  Y[1] binary
 #  Y[2] binary
 
[…]
