+++ README.md
@@ -1,5 +1,5 @@
 # DisjunctiveProgramming.jl
-Generalized Disjunctive Programming extension to JuMP
+Generalized Disjunctive Programming (GDP) extension to JuMP, based on the GDP modeling paradigm described in [Perez and Grossmann, 2023](https://arxiv.org/abs/2303.04375).
 
 ## Installation
 
@@ -8,80 +8,206 @@ using Pkg
 Pkg.add("DisjunctiveProgramming")
 ```
 
+## Model
+
+A generalized disjunctive programming (GDP) model is created using `GDPModel()`, where the optimizer can be passed at model creation, along with other keyword arguments supported by JuMP Models. 
+
+```julia
+using DisjunctiveProgramming
+using HiGHS
+
+model = GDPModel(HiGHS.Optimizer)
+```
+
+A `GDPModel` is a `JuMP Model` with a `GDPData` field in the model's `.ext` dictionary, which stores the following:
+
+- `Logical Variables`: Indicator variables used for the various disjuncts involved in the model's disjunctions.
+- `Logical Constraints`: Selector (cardinality) or proposition (Boolean) constraints describing the relationships between the logical variables.
+- `Disjunct Constraints`: Constraints associated with each disjunct in the model.
+- `Disjunctions`: Disjunction constraints.
+- `Solution Method`: The reformulation technique or solution method. Currently supported methods include Big-M, Hull, and Indicator Constraints.
+- `Reformulation Variables`: List of JuMP variables created when reformulating a GDP model into a MIP model.
+- `Reformulation Constraints`: List of constraints created when reformulating a GDP model into a MIP model.
+- `Ready to Optimize`: Flag indicating if the model can be optimized.
+
+Additionally, the following mapping dictionaries are stored in `GDPData`:
+
+- `Indicator to Binary`: Maps the Logical variables to their respective reformulated Binary variables.
+- `Indicator to Constraints`: Maps the Logical variables to the disjunct constraints associated with them.
+
+A GDP Model's `GDPData` can be accessed via:
+
+```julia
+data = gdp_data(model)
+```
+
+## Logical Variables
+
+Logical variables are JuMP `AbstractVariable`s with two fields: `fix_value` and `start_value`. These can be optionally specified at variable creation. Logical variables are created with the `@variable` JuMP macro by adding the tag `LogicalVariable` as the last keyword argument. As with the regular `@variable` macro, variables can be named and indexed: 
+
+```julia
+@variable(model, Y[1:3], LogicalVariable)
+```
+
+## Logical Constraints
+
+Two types of logical constraints are supported:
+
+1. `Selector` or cardinality constraints: A subset of Logical variables is passed and `Exactly`, `AtMost`, or `AtLeast` `n` of these is allowed to be `True`. These constraints are specified with the `func` $\in$ `set` notation in `MathOptInterface` in a `@constraint` JuMP macro. It is not assumed that disjunctions have an `Exactly(1)` constraint enforced on their disjuncts upon creation. This constraint must be explicitly specified.
+
+    ```julia
+    @constraint(model, [Y[1], Y[2]] in Exactly(1))
+    ```
+
+2. `Proposition` or Boolean constraints: These describe the relationships between Logical variables via Boolean algebra. Supported logical operators include:
+
[…]
