### 2023-10-10 bc2a3bc
Update logic heads to be consistent with JuMP
