+++ test/constraints/proposition.jl
@@ -30,7 +30,7 @@ function test_negation_add_success()
     @test constraint_object(c1).func isa DP._LogicalExpr
     @test constraint_object(c2).func isa DP._LogicalExpr
     @test constraint_object(c1).func.head == 
-            constraint_object(c2).func.head == :¬
+            constraint_object(c2).func.head == :!
     @test constraint_object(c1).func.args ==
             constraint_object(c2).func.args == Any[y]
     @test constraint_object(c1).set == 
@@ -45,7 +45,7 @@ function test_implication_add_success()
     @constraint(model, c2, (y[1] ⟹ y[2]) in IsTrue())
     @test_macro_throws ErrorException @constraint(model, y[1] ⟹ y[2] in IsTrue())
     @test constraint_object(c1).func.head == 
-            constraint_object(c2).func.head == :⟹
+            constraint_object(c2).func.head == :(=>)
     @test constraint_object(c1).func.args ==
             constraint_object(c2).func.args == Vector{Any}(y)
     @test constraint_object(c1).set == 
@@ -59,7 +59,7 @@ function test_equivalence_add_success()
     @constraint(model, c2, (y[1] ⇔ y[2]) in IsTrue())
     @test_macro_throws ErrorException @constraint(model, y[1] ⇔ y[2] in IsTrue())
     @test constraint_object(c1).func.head == 
-            constraint_object(c2).func.head == :⇔
+            constraint_object(c2).func.head == :(==)
     @test constraint_object(c1).func.args ==
             constraint_object(c2).func.args == Vector{Any}(y)
     @test constraint_object(c1).set == 
@@ -77,7 +77,7 @@ function test_intersection_and_flatten_add_success()
     @test is_valid(model, c3)
     @test constraint_object(c1).func.head == 
             constraint_object(c2).func.head == 
-            constraint_object(c3).func.head == :∧
+            constraint_object(c3).func.head == :&&
     @test Set(constraint_object(c1).func.args) ==
             Set(constraint_object(c2).func.args) ==
             Set(DP._flatten(constraint_object(c3).func).args)
@@ -97,7 +97,7 @@ function test_union_and_flatten_add_success()
     @test is_valid(model, c3)
     @test constraint_object(c1).func.head == 
             constraint_object(c2).func.head == 
-            constraint_object(c3).func.head == :∨
+            constraint_object(c3).func.head == :||
     @test Set(constraint_object(c1).func.args) ==
             Set(constraint_object(c2).func.args) ==
             Set(DP._flatten(constraint_object(c3).func).args)
@@ -243,7 +243,7 @@ function test_lvar_cnf_functions()
     @test DP._eliminate_implication(y) == y
     @test DP._move_negations_inward(y) == y
     neg_y = DP._negate(y)
-    @test neg_y.head == :¬
+    @test neg_y.head == :!
     @test neg_y.args[1] == y
     @test DP._distribute_and_over_or(y) == y
     @test DP._flatten(y) == y
@@ -254,10 +254,10 @@ function test_eliminate_equivalence()
     @variable(model, y[1:2], LogicalVariable)
     ex = y[1] ⇔ y[2]
     new_ex = DP._eliminate_equivalence(ex)
-    @test new_ex.head == :∧
+    @test new_ex.head == :&&
     @test length(new_ex.args) == 2
[…]
