+++ src/logic.jl
@@ -6,32 +6,26 @@ function _op_fallback(name)
 end
 
 # Define all the logical operators
-const _logical_operators = (
-    :∨, :logical_or, # \vee + tab
-    :∧, :logical_and, # \wedge + tab
-    :¬, :logical_not, # \neg + tab
-    :⇔, :iff, :⇔, # \Leftrightarrow + tab
-    :⟹, :implies, :⟹ # Longrightarrow + tab
-)
-for (name, alt) in (
-    (:⇔, :iff), # \Leftrightarrow + tab
-    (:⟹, :implies) # Longrightarrow + tab
+const _LogicalOperatorHeads = (:(==), :(=>), :||, :&&, :!)
+for (name, alt, head) in (
+    (:⇔, :iff, :(==)), # \Leftrightarrow + tab
+    (:⟹, :implies, :(=>)) # Longrightarrow + tab
     )
     # make operators
     @eval begin 
-        const $name = JuMP.NonlinearOperator((vs...) -> _op_fallback($(Meta.quot(name))), $(Meta.quot(name)))
-        const $alt = JuMP.NonlinearOperator((vs...) -> _op_fallback($(Meta.quot(alt))), $(Meta.quot(name)))
+        const $name = JuMP.NonlinearOperator((vs...) -> _op_fallback($(Meta.quot(name))), $(Meta.quot(head)))
+        const $alt = JuMP.NonlinearOperator((vs...) -> _op_fallback($(Meta.quot(alt))), $(Meta.quot(head)))
     end
 end
-for (name, alt, func) in (
-    (:∨, :logical_or, :(|)), # \vee + tab
-    (:∧, :logical_and, :(&)), # \wedge + tab
-    (:¬, :logical_not, :(!)) # \neg + tab
+for (name, alt, head, func) in (
+    (:∨, :logical_or, :||, :(|)), # \vee + tab
+    (:∧, :logical_and, :&&, :(&)), # \wedge + tab
+    (:¬, :logical_not, :!, :(!)) # \neg + tab
     )
     # make operators
     @eval begin 
-        const $name = JuMP.NonlinearOperator($func, $(Meta.quot(name)))
-        const $alt = JuMP.NonlinearOperator($func, $(Meta.quot(name)))
+        const $name = JuMP.NonlinearOperator($func, $(Meta.quot(head)))
+        const $alt = JuMP.NonlinearOperator($func, $(Meta.quot(head)))
     end
 end
 
@@ -55,19 +49,19 @@ function _eliminate_equivalence(lvar::LogicalVariableRef)
     return lvar
 end
 function _eliminate_equivalence(lexpr::_LogicalExpr)
-    if lexpr.head == :⇔
+    if lexpr.head == :(==)
         A = _eliminate_equivalence(lexpr.args[1])
         if length(lexpr.args) > 2 
-            nested = _LogicalExpr(:⇔, Vector{Any}(lexpr.args[2:end]))
+            nested = _LogicalExpr(:(==), Vector{Any}(lexpr.args[2:end]))
             B = _eliminate_equivalence(nested)
         elseif length(lexpr.args) == 2
             B = _eliminate_equivalence(lexpr.args[2])
         else
             error("The equivalence logic operator `⇔` must have at least two arguments.")
         end
-        new_lexpr = _LogicalExpr(:∧, Any[
-            _LogicalExpr(:⟹, Any[A, B]),
-            _LogicalExpr(:⟹, Any[B, A])
+        new_lexpr = _LogicalExpr(:&&, Any[
+            _LogicalExpr(:(=>), Any[A, B]),
+            _LogicalExpr(:(=>), Any[B, A])
         ])
     else
[…]
