### 2023-10-11 f91885c
merge https://github.com/pulsipher/DisjunctiveProgramming.jl/tree/structure_overall
