+++ src/constraints.jl
@@ -1,0 +1,541 @@
+################################################################################
+#                         HELPER SET MAPPING FUNCTIONS
+################################################################################
+# helper functions to get the value of an MOI set
+_set_value(set::_MOI.LessThan) = set.upper
+_set_value(set::_MOI.GreaterThan) = set.lower
+_set_value(set::_MOI.EqualTo) = set.value
+_set_values(set::_MOI.EqualTo) = (set.value, set.value)
+_set_values(set::_MOI.Interval) = (set.lower, set.upper)
+
+# helper functions to reformulate vector constraints to indicator constraints
+_vec_to_scalar_set(set::_MOI.Nonpositives) = _MOI.LessThan(0)
+_vec_to_scalar_set(set::_MOI.Nonnegatives) = _MOI.GreaterThan(0)
+_vec_to_scalar_set(set::_MOI.Zeros) = _MOI.EqualTo(0)
+
+# helper functions to map jump selector to moi selector sets
+_jump_to_moi_selector(set::Exactly) = _MOIExactly
+_jump_to_moi_selector(set::AtLeast) = _MOIAtLeast
+_jump_to_moi_selector(set::AtMost) = _MOIAtMost
+
+# helper functions to map selectors to scalar sets
+_vec_to_scalar_set(set::_MOIExactly) = _MOI.EqualTo
+_vec_to_scalar_set(set::_MOIAtLeast) = _MOI.GreaterThan
+_vec_to_scalar_set(set::_MOIAtMost) = _MOI.LessThan
+
+################################################################################
+#                         BOILERPLATE EXTENSION METHODS
+################################################################################
+for (RefType, loc) in ((:DisjunctConstraintRef, :disjunct_constraints), 
+                       (:DisjunctionRef, :disjunctions), 
+                       (:LogicalConstraintRef, :logical_constraints))
+    @eval begin
+        @doc """
+            JuMP.owner_model(cref::$($RefType))
+
+        Return the model to which `cref` belongs.
+        """
+        JuMP.owner_model(cref::$RefType) = cref.model
+
+        @doc """
+            JuMP.index(cref::$($RefType))
+
+        Return the index constraint associated with `cref`.
+        """
+        JuMP.index(cref::$RefType) = cref.index
+
+        @doc """
+            JuMP.is_valid(model::JuMP.Model, cref::$($RefType))
+
+        Return `true` if `cref` refers to a valid constraint in the `GDP model`.
+        """
+        function JuMP.is_valid(model::JuMP.Model, cref::$RefType) # TODO: generalize for AbstractModel
+            return model === JuMP.owner_model(cref)
+        end
+
+        # Get the ConstraintData object
+        function _constraint_data(cref::$RefType)
+            return gdp_data(JuMP.owner_model(cref)).$loc[JuMP.index(cref)]
+        end
+
+        @doc """
+            JuMP.name(cref::$($RefType))
+
+        Get a constraint's name attribute.
+        """
+        function JuMP.name(cref::$RefType)
+            return _constraint_data(cref).name
+        end
+
+        @doc """
+            JuMP.set_name(cref::$($RefType), name::String)
+
+        Set a constraint's name attribute.
+        """
+        function JuMP.set_name(cref::$RefType, name::String)
+            _constraint_data(cref).name = name
+            _set_ready_to_optimize(JuMP.owner_model(cref), false)
+            return
+        end
+
+        @doc """
+            JuMP.constraint_object(cref::$($RefType))
+
+        Return the underlying constraint data for the constraint 
+            referenced by `cref`.
[…]
