+++ test/variables/query.jl
@@ -1,0 +1,168 @@
+function test_interrogate_non_variables()
+    vars = Set()
+    f = Base.Fix1(push!, vars) #interrogator
+    #numbers
+    DP._interrogate_variables(f, 1)
+    @test isempty(vars)
+    empty!(vars)
+    DP._interrogate_variables(f, 1.0)
+    @test isempty(vars)
+    empty!(vars)
+    #strings/symbols
+    @test_throws ErrorException DP._interrogate_variables(f, "a")
+    @test_throws ErrorException DP._interrogate_variables(f, :a)
+end
+
+function test_interrogate_variables()
+    vars = Set()
+    f = Base.Fix1(push!, vars) #interrogator
+    m = GDPModel()
+    @variable(m, x)
+    @variable(m, y, LogicalVariable)
+    DP._interrogate_variables(f, [x, y])
+    @test x in vars
+    @test y in vars
+    @test length(vars) == 2
+end
+
+function test_interrogate_affexpr()
+    vars = Set()
+    f = Base.Fix1(push!, vars) #interrogator
+    m = GDPModel()
+    @variable(m, x)
+    @variable(m, y, LogicalVariable)
+    @variable(m, z)
+    DP._interrogate_variables(f, x + y + z)
+    @test x in vars
+    @test y in vars
+    @test z in vars
+    @test length(vars) == 3
+end
+
+function test_interrogate_quadexpr()
+    vars = Set()
+    f = Base.Fix1(push!, vars) #interrogator
+    m = GDPModel()
+    @variable(m, x)
+    @variable(m, y, LogicalVariable)
+    @variable(m, z)
+    DP._interrogate_variables(f, x^2 + x*y + z + 1)
+    @test x in vars
+    @test y in vars
+    @test z in vars
+    @test length(vars) == 3
+    empty!(vars)
+end
+
+function test_interrogate_nonlinear_expr()
+    vars = Set()
+    f = Base.Fix1(push!, vars) #interrogator
+    m = GDPModel()
+    @variable(m, x)
+    @variable(m, y, LogicalVariable)
+    @variable(m, z)
+    DP._interrogate_variables(f, sin(exp(x^2 + 1)) + cos(x) + y + 2)
+    @test x in vars
+    @test y in vars
+    @test !(z in vars)
+    @test length(vars) == 2
+end
+
+function test_interrogate_logical_expr()
+    vars = Set()
+    f = Base.Fix1(push!, vars) #interrogator
+    m = GDPModel()
+    @variable(m, y, LogicalVariable)
+    @variable(m, w[1:5], LogicalVariable)
+    ex = (implies(w[1], w[2]) ∧ w[3]) ⇔ (¬w[4] ∨ y)
+    DP._interrogate_variables(f, ex)
+    @test w[1] in vars
+    @test w[2] in vars
+    @test w[3] in vars
+    @test w[4] in vars
+    @test !(w[5] in vars)
+    @test y in vars
+    @test length(vars) == 5
+end
+
+function test_interrogate_proposition_constraint()
+    m = GDPModel()
+    @variable(m, y, LogicalVariable)
+    @variable(m, w[1:5], LogicalVariable)
+    ex = (implies(w[1], w[2]) ∧ w[3]) ⇔ (¬w[4] ∨ y)
+    @constraint(m, con, ex in IsTrue())
+    obj = constraint_object(con)
+    vars = DP._get_constraint_variables(m, obj)
+    @test w[1] in vars
+    @test w[2] in vars
+    @test w[3] in vars
+    @test w[4] in vars
+    @test !(w[5] in vars)
+    @test y in vars
[…]
