+++ Project.toml
@@ -9,7 +9,7 @@ Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 
 [compat]
 JuMP = "1.15"
-Reexport = "0.2, 1"
+Reexport = "1"
 julia = "1.6"
 
 [extras]
