### 2023-10-12 2231536
Improve allocation
