+++ src/moi.jl
@@ -14,30 +14,31 @@
 #                            REFRORMULATION METHODS
 ################################################################################
 # Helper methods to handle recursively flattening the disjuncts
-function _constr_set!(funcs, con::JuMP.AbstractConstraint)
+function _constr_set!(model, funcs, con::JuMP.AbstractConstraint)
     append!(funcs, JuMP.jump_function(con))
     return JuMP.moi_set(con)
 end
-function _constr_set!(funcs, con::Disjunction)
-    inner_funcs, set = _disjunction_to_set(con)
+function _constr_set!(model, funcs, con::Disjunction)
+    inner_funcs, set = _disjunction_to_set(model, con)
     append!(funcs, inner_funcs)
     return set
 end
 
 # Create the vectors needed for a disjunction vector constraint
-function _disjunction_to_set(d::Disjunction)
+function _disjunction_to_set(model::JuMP.Model, d::Disjunction)
     # allocate memory for the storage vectors
     num_disjuncts = length(d.indicators)
-    funcs = sizehint!(JuMP.AbstractJuMPScalar[], num_disjuncts)
+    constr_mappings = _indicator_to_constraints(model)
+    num_constrs = sum(length(constr_mappings[lvref]) for lvref in d.indicators)
+    funcs = sizehint!(JuMP.AbstractJuMPScalar[], num_disjuncts + num_constrs)
     sets = Vector{Vector{_MOI.AbstractSet}}(undef, num_disjuncts)
     d_idxs = Vector{Int}(undef, num_disjuncts)
     # iterate over the underlying disjuncts to fill in the storage vectors
     for (i, lvref) in enumerate(d.indicators)
-        model = JuMP.owner_model(lvref)
         push!(funcs, _indicator_to_binary(model)[lvref])
         d_idxs[i] = length(funcs)
-        crefs = _indicator_to_constraints(model)[lvref]
-        sets[i] = map(c -> _constr_set!(funcs, JuMP.constraint_object(c)), crefs)
+        crefs = constr_mappings(model)[lvref]
+        sets[i] = map(c -> _constr_set!(model, funcs, JuMP.constraint_object(c)), crefs)
     end
     # convert the `sets` type to be concrete if possible (TODO benchmark if this is worth it)
     SetType = typeof(first(sets))
@@ -53,6 +54,6 @@ function reformulate_disjunction(
     d::Disjunction, 
     ::MOIDisjunction
     )
-    funcs, set = _disjunction_to_set(d)
+    funcs, set = _disjunction_to_set(model, d)
     return [JuMP.VectorConstraint(funcs, set, JuMP.VectorShape())]
 end
