### 2023-10-13 ce7a890
add Logical alias for Logical Variable. Closes #73
