+++ test/disjunction.jl
@@ -1,7 +1,7 @@
 function test_disjunction_add_fail()
     model = GDPModel()
     @variable(model, x)
-    @variable(model, y[1:2], LogicalVariable)
+    @variable(model, y[1:2], Logical)
     @constraint(model, x == 5, DisjunctConstraint(y[1]))
 
     @test_macro_throws ErrorException @disjunction(model) #not enough arguments
@@ -26,7 +26,7 @@ end
 function test_disjunction_add_success()
     model = GDPModel()
     @variable(model, x)
-    @variable(model, y[1:2], LogicalVariable)
+    @variable(model, y[1:2], Logical)
     @constraint(model, x == 5, DisjunctConstraint(y[1]))
     @constraint(model, x == 10, DisjunctConstraint(y[2]))
     disj = @disjunction(model, y)
@@ -47,8 +47,8 @@ end
 function test_disjunction_add_nested()
     model = GDPModel()
     @variable(model, x)
-    @variable(model, y[1:2], LogicalVariable)
-    @variable(model, z[1:2], LogicalVariable)
+    @variable(model, y[1:2], Logical)
+    @variable(model, z[1:2], Logical)
     @constraint(model, x <= 5, DisjunctConstraint(y[1]))
     @constraint(model, x >= 5, DisjunctConstraint(y[2]))
     @disjunction(model, inner, y, DisjunctConstraint(z[1]))
@@ -69,7 +69,7 @@ end
 function test_disjunction_add_array()
     model=GDPModel()
     @variable(model, x)
-    @variable(model, y[1:2, 1:3, 1:4], LogicalVariable)
+    @variable(model, y[1:2, 1:3, 1:4], Logical)
     @constraint(model, con[i=1:2, j=1:3, k=1:4], x==i+j+k, DisjunctConstraint(y[i,j,k]))
     @disjunction(model, disj[i=1:2, j=1:3], y[i,j,:])
 
@@ -83,7 +83,7 @@ function test_disjunciton_add_dense_axis()
     @variable(model, x)
     I = ["a", "b", "c"]
     J = [1, 2]
-    @variable(model, y[I, J, 1:4], LogicalVariable)
+    @variable(model, y[I, J, 1:4], Logical)
     @constraint(model, con[i=I, j=J, k=1:4], x==k, DisjunctConstraint(y[i,j,k]))
     @disjunction(model, disj[i=I, j=J], y[i,j,:])
 
@@ -96,7 +96,7 @@ end
 function test_disjunction_add_sparse_axis()
     model = GDPModel()
     @variable(model, x)
-    @variable(model, y[1:3, 1:3, 1:4], LogicalVariable)
+    @variable(model, y[1:3, 1:3, 1:4], Logical)
     @constraint(model, con[i=1:3, j=1:3, k=1:4; j > i], x==i+j+k, DisjunctConstraint(y[i,j,k]))
     @disjunction(model, disj[i=1:3, j=1:3; j > i], y[i,j,:])
 
@@ -109,8 +109,8 @@ end
 function test_disjunctions_add_success()
     model = GDPModel()
     @variable(model, x)
-    @variable(model, y[1:2], LogicalVariable)
-    @variable(model, z[1:2], LogicalVariable)
+    @variable(model, y[1:2], Logical)
+    @variable(model, z[1:2], Logical)
     @constraint(model, x <= 5, DisjunctConstraint(y[1]))
     @constraint(model, x >= 5, DisjunctConstraint(y[2]))
     @constraint(model, x <= 10, DisjunctConstraint(z[1]))
@@ -136,7 +136,7 @@ end
 function test_disjunction_set_name()
     model = GDPModel()
     @variable(model, x)
-    @variable(model, y[1:2], LogicalVariable)
[…]
