+++ test/variables/logical.jl
@@ -1,12 +1,12 @@
 # test creating, modifying, and reformulating logical variables
 function test_lvar_add_fail()
    model = Model()
-   @test_throws ErrorException @variable(model, y, LogicalVariable)
+   @test_throws ErrorException @variable(model, y, Logical)
 end
 
 function test_lvar_add_success()
     model = GDPModel()
-    @variable(model, y, LogicalVariable)
+    @variable(model, y, Logical)
     @test typeof(y) == LogicalVariableRef
     @test owner_model(y) == model
     @test is_valid(model, y)
@@ -24,14 +24,14 @@ end
 
 function test_lvar_add_array()
     model = GDPModel()
-    @variable(model, y[1:3, 1:2], LogicalVariable)
+    @variable(model, y[1:3, 1:2], Logical)
     @test y isa Array{LogicalVariableRef, 2}
     @test length(y) == 6
 end
 
 function test_lvar_add_dense_axis()
     model = GDPModel()
-    @variable(model, y[["a","b","c"],[1,2]], LogicalVariable)
+    @variable(model, y[["a","b","c"],[1,2]], Logical)
     @test y isa Containers.DenseAxisArray
     @test length(y) == 6
     @test y.axes[1] == ["a","b","c"]
@@ -41,7 +41,7 @@ end
 
 function test_lvar_add_sparse_axis()
     model = GDPModel()
-    @variable(model, y[i = 1:3, j = 1:3; j > i], LogicalVariable)
+    @variable(model, y[i = 1:3, j = 1:3; j > i], Logical)
     @test y isa Containers.SparseAxisArray
     @test length(y) == 3
     @test y.names == (:i, :j)
@@ -50,7 +50,7 @@ end
 
 function test_lvar_set_name()
     model = GDPModel()
-    @variable(model, y, LogicalVariable)
+    @variable(model, y, Logical)
     set_name(y, "z")
     @test name(y) == "z"
     #reformulate the variable
@@ -59,7 +59,7 @@ end
 
 function test_lvar_creation_start_value()
     model = GDPModel()
-    @variable(model, y, LogicalVariable, start = true)
+    @variable(model, y, Logical, start = true)
     @test start_value(y)
     #reformulate the variable
     test_lvar_reformulation(model, y)
@@ -67,7 +67,7 @@ end
 
 function test_lvar_set_start_value()
     model = GDPModel()
-    @variable(model, y, LogicalVariable)
+    @variable(model, y, Logical)
     @test isnothing(start_value(y))
     set_start_value(y, false)
     @test !start_value(y)
@@ -77,7 +77,7 @@ end
 
 function test_lvar_fix_value()
     model = GDPModel()
-    @variable(model, y, LogicalVariable)
+    @variable(model, y, Logical)
     @test isnothing(fix_value(y))
     fix(y, true)
     @test fix_value(y)
@@ -90,8 +90,8 @@ end
 
 function test_lvar_delete()
     model = GDPModel()
-    @variable(model, y, LogicalVariable)
-    @variable(model, z, LogicalVariable)
+    @variable(model, y, Logical)
+    @variable(model, z, Logical)
     @variable(model, x)
     @constraint(model, con, x <= 10, DisjunctConstraint(y))
     @constraint(model, con2, x >= 50, DisjunctConstraint(z))
@@ -115,7 +115,7 @@ end
 
 function test_lvar_reformulation()
     model = GDPModel()
-    @variable(model, y, LogicalVariable, start = false)
+    @variable(model, y, Logical, start = false)
     fix(y, true)
     test_lvar_reformulation(model, y)
 end

[…]
