### 2023-10-13 6a56b4d
remove using JuMP. remove jldoctests
