+++ docs/make.jl
@@ -1,12 +1,13 @@
-push!(LOAD_PATH,"../src/")
 using DisjunctiveProgramming
 using Documenter
 makedocs(
-         sitename = "DisjunctiveProgramming.jl",
-         modules  = [DisjunctiveProgramming],
-         pages=[
-                "Home" => "index.md"
-               ])
+    sitename = "DisjunctiveProgramming.jl",
+    modules  = [DisjunctiveProgramming],
+    pages=[
+        "Home" => "index.md"
+    ],
+    checkdocs = :none
+)
 deploydocs(;
     repo="github.com/hdavid16/DisjunctiveProgramming.jl",
 )
