+++ examples/ex6.jl
@@ -1,4 +1,3 @@
-using JuMP
 using DisjunctiveProgramming
 
 ##
