+++ docs/Project.toml
@@ -1,2 +1,3 @@
 [deps]
 DisjunctiveProgramming = "0d27d021-0159-4c7d-b4a7-9ccb5d9366cf"
+Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
