### 2023-10-14 e98764b
begin fixes to paper
