+++ paper/ref.bib
@@ -198,7 +198,7 @@ pages={183–198}
 } 
 
 @article{dunning_huchette_lubin_2017,
-title={Jump: A modeling language for mathematical optimization},
+title={JuMP: A modeling language for mathematical optimization},
 volume={59},
 number={2},
 journal={SIAM Review},
