### 2023-10-14 91ea54e
update example in readme for review
