+++ docs/src/api.md
@@ -1,0 +1,6 @@
+# API
+
+```@autodocs
+Modules = [Foo, Bar, Bar.Baz]
+Order   = [:function, :type]
+```
