+++ docs/src/index.md
@@ -1,2 +1,214 @@
-# Overview
-DisjunctiveProgramming.jl
+# DisjunctiveProgramming.jl
+Generalized Disjunctive Programming (GDP) extension to JuMP, based on the GDP modeling paradigm described in [Perez and Grossmann, 2023](https://arxiv.org/abs/2303.04375).
+
+![](assets/logo.png)
+
+[![codecov](https://codecov.io/gh/hdavid16/DisjunctiveProgramming.jl/graph/badge.svg?token=3FRPGMWF0J)](https://codecov.io/gh/hdavid16/DisjunctiveProgramming.jl)
+[![Docs](https://img.shields.io/badge/docs-stable-blue.svg)](https://hdavid16.github.io/InventoryManagement.jl/stable/)
+[![Docs](https://img.shields.io/badge/docs-latest-blue.svg)](https://hdavid16.github.io/InventoryManagement.jl/dev/)
+
+## Installation
+
+```julia
+using Pkg
+Pkg.add("DisjunctiveProgramming")
+```
+
+## Model
+
+A generalized disjunctive programming (GDP) model is created using `GDPModel()`, where the optimizer can be passed at model creation, along with other keyword arguments supported by JuMP Models. 
+
+```julia
+using DisjunctiveProgramming
+using HiGHS
+
+model = GDPModel(HiGHS.Optimizer)
+```
+
+A `GDPModel` is a `JuMP Model` with a `GDPData` field in the model's `.ext` dictionary, which stores the following:
+
+- `Logical Variables`: Indicator variables used for the various disjuncts involved in the model's disjunctions.
+- `Logical Constraints`: Selector (cardinality) or proposition (Boolean) constraints describing the relationships between the logical variables.
+- `Disjunct Constraints`: Constraints associated with each disjunct in the model.
+- `Disjunctions`: Disjunction constraints.
+- `Solution Method`: The reformulation technique or solution method. Currently supported methods include Big-M, Hull, and Indicator Constraints.
+- `Reformulation Variables`: List of JuMP variables created when reformulating a GDP model into a MIP model.
+- `Reformulation Constraints`: List of constraints created when reformulating a GDP model into a MIP model.
+- `Ready to Optimize`: Flag indicating if the model can be optimized.
+
+Additionally, the following mapping dictionaries are stored in `GDPData`:
+
+- `Indicator to Binary`: Maps the Logical variables to their respective reformulated Binary variables.
+- `Indicator to Constraints`: Maps the Logical variables to the disjunct constraints associated with them.
+
+A GDP Model's `GDPData` can be accessed via:
+
+```julia
+data = gdp_data(model)
+```
+
+## Logical Variables
+
+Logical variables are JuMP `AbstractVariable`s with two fields: `fix_value` and `start_value`. These can be optionally specified at variable creation. Logical variables are created with the `@variable` JuMP macro by adding the tag `Logical` as the last keyword argument. As with the regular `@variable` macro, variables can be named and indexed: 
+
+```julia
+@variable(model, Y[1:3], Logical)
+```
+
+## Logical Constraints
+
+Two types of logical constraints are supported:
+
[…]
