+++ paper/ref.bib
@@ -26,7 +26,8 @@
   number={1},
   pages={607--642},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-3-030-00148-3}
 }
 @article{vecchietti1999logmip,
   title={LOGMIP: a disjunctive 0--1 non-linear optimizer for process system models},
@@ -98,7 +99,8 @@ author={Bynum, Michael L. and Hackebeil, Gabriel A. and Hart, William E. and Lai
 edition={Third},
 volume={67},
 year={2021},
-publisher={Springer Science \& Business Media}
+publisher={Springer Science \& Business Media},
+doi={10.1007/978-3-030-68928-5}
 }
 
 @article{hart2011pyomo,
@@ -156,14 +158,6 @@ url = {https://www.sciencedirect.com/science/article/pii/S0098135422001946},
 author = {Fahad Matovu and Shuhaimi Mahadzir and Rasel Ahmed and Nor Erniza Mohammad Rozali},
 keywords = {Multilevel refrigeration, GDP Modelling, Synthesis and optimization, Logic based branch and bound},
 abstract = {The synthesis and optimization of multilevel refrigeration systems is challenging because it’s a highly non-linear, multi-variable, and multi-modal problem. This work presents a novel approach to develop a complete generalized disjunctive programming model for the synthesis and optimization of multilevel refrigeration systems. The model is based on the application of mass, energy, and thermodynamic model equations and developed to optimize the total shaft work requirement of the system. The presented model is very systematic, easily manageable, and can be extendable to include all design features for refrigeration systems. The model is solvable using advanced solution algorithms such as the Logic-based branch and bound because of it’s disjunctive nature. Thus the solution is sought in reduced space and avoids the full scale as is the case with Mixed integer linear/non-linear models. The model solution yields optimal temperature/pressure levels, mass flow rates, and the shaft work consumption. A case study of the precooling cycle of the propane pre-cooled mixed refrigerant (C3MR) LNG process is used to test the proposed model. The model results are validated by simulation using Aspen Hysys software. Shaft work savings of up to 12.3% are obtained from the model results against the base case. Preliminary estimations show that cost savings of up to 2.55 MM$/y are realizable against the base case.}
-}
-
-@article{kim2022generalized,
-  title={Generalized Disjunctive Programming-based, Mixed Integer Linear MPC Formulation for Optimal Operation of a District Energy System for PV Self-consumption and Grid Decarbonization: Field Implementation},
-  author={Kim, Donghun},
-  journal={International High Performance Buildings Conference},
-  url={https://docs.lib.purdue.edu/cgi/viewcontent.cgi?article=1395\&context=ihpbc},
-  year={2022}
 }
 
 @article{ZHOU202269,
@@ -257,7 +251,8 @@ place={Cham, Switzerland},
 title={Disjunctive programming},
 publisher={Springer},
 author={Balas, Egon},
-year={2018}
+year={2018},
+doi={10.1007/978-3-030-00148-3}
 }
 
 @article{TRESPALACIOS201598,
@@ -294,7 +289,8 @@ place={New York},
 title={Integer and combinatorial optimization},
 publisher={John Wiley and Sons},
 author={Nemhauser, George L.},
-year={1999}
+year={1999},
+doi={10.1002/9781118627372}
 } 
 
 @article{grossmann_lee_2003,
@@ -356,11 +352,13 @@ abstract = {Raman and Grossmann [Raman, R., \& Grossmann, I.E. (1994). Modeling 
[…]
