+++ docs/src/api.md
@@ -1,6 +1,6 @@
 # API
 
 ```@autodocs
-Modules = [Foo, Bar, Bar.Baz]
-Order   = [:function, :type]
+Modules = [DisjunctiveProgramming]
+Order   = [:type, :function]
 ```
