+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"#Overview","page":"Home","title":"Overview","text":"","category":"section"},{"location":"","page":"Home","title":"Home","text":"DisjunctiveProgramming.jl","category":"page"}]
+[{"location":"api/#API","page":"API","title":"API","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"Modules = [DisjunctiveProgramming]\nOrder   = [:type, :function]","category":"page"},{"location":"api/#DisjunctiveProgramming.AbstractReformulationMethod","page":"API","title":"DisjunctiveProgramming.AbstractReformulationMethod","text":"AbstractReformulationMethod <: AbstractSolutionMethod\n\nAn abstract type for reformulation approaches used to solve GDPModels.\n\n\n\n\n\n","category":"type"},{"location":"api/#DisjunctiveProgramming.AbstractSolutionMethod","page":"API","title":"DisjunctiveProgramming.AbstractSolutionMethod","text":"AbstractSolutionMethod\n\nAn abstract type for solution methods used to solve GDPModels.\n\n\n\n\n\n","category":"type"},{"location":"api/#DisjunctiveProgramming.AtLeast","page":"API","title":"DisjunctiveProgramming.AtLeast","text":"AtLeast{T<:Union{Int,LogicalVariableRef}} <: JuMP.AbstractVectorSet\n\nConvenient alias for using _MOIAtLeast.\n\n\n\n\n\n","category":"type"},{"location":"api/#DisjunctiveProgramming.AtMost","page":"API","title":"DisjunctiveProgramming.AtMost","text":"AtMost{T<:Union{Int,LogicalVariableRef}} <: JuMP.AbstractVectorSet\n\nConvenient alias for using _MOIAtMost.\n\n\n\n\n\n","category":"type"},{"location":"api/#DisjunctiveProgramming.BigM","page":"API","title":"DisjunctiveProgramming.BigM","text":"BigM <: AbstractReformulationMethod\n\nA type for using the big-M reformulation approach for disjunctive constraints.\n\nFields\n\nvalue::Float64: Big-M value (default = 1e9).\ntight::Bool: Attempt to tighten the Big-M value (default = true)?\n\n\n\n\n\n","category":"type"},{"location":"api/#DisjunctiveProgramming.ConstraintData","page":"API","title":"DisjunctiveProgramming.ConstraintData","text":"ConstraintData{C <: JuMP.AbstractConstraint}\n\nA type for storing constraint objects in GDPData and any meta-data  they possess.\n\nFields\n\nconstraint::C: The constraint.\nname::String: The name of the proposition.\n\n\n\n\n\n","category":"type"},{"location":"api/#DisjunctiveProgramming.DisjunctConstraint","page":"API","title":"DisjunctiveProgramming.DisjunctConstraint","text":"DisjunctConstraint\n\nUsed as a tag for constraints that will be used in disjunctions. This is done via  the following syntax:\n\njulia> @constraint(model, [constr_expr], DisjunctConstraint)\n\njulia> @constraint(model, [constr_expr], DisjunctConstraint(lvref))\n\nwhere lvref is a LogicalVariableRef that will ultimately be associated  with the disjunct the constraint is added to. If no lvref is given, then one is  generated when the disjunction is created.\n\n\n\n\n\n","category":"type"},{"location":"api/#DisjunctiveProgramming.DisjunctConstraintIndex","page":"API","title":"DisjunctiveProgramming.DisjunctConstraintIndex","text":"DisjunctConstraintIndex\n\nA
[…]
