### 2023-10-14 #117 @hdavid16
@joaquimg thank you for your review.

The package has just gone a complete rewrite and has now been released (v0.4.1).  With this new release:

- Installing should work on 1.9 without any errors.
- Docs and CI have also been added.
- DOIs have been fixed/added.
- Community guidelines added.
- Full example with `optimize!` call added to README and docs.

Regarding your main points:

1. Examples are self-contained now as you suggested.
2. Conversion of logical constraints to algebraic inequalities has been written in a more clear way.
3. Section 2.4 is referenced in future work.
4. Rewrite avoids ambiguous usage for nested disjunctions.
5. Reference fixed.
6. Capitalization in JuMP reference fixed.

I look forward to your feedback.
