### 2023-10-14 afed95f
remove unused function
