### 2023-10-14 37cc24d
remove unused function
