+++ test/runtests.jl
@@ -27,6 +27,7 @@ end
 
 include("aqua.jl")
 include("model.jl")
+include("jump.jl")
 include("variables/query.jl")
 include("variables/logical.jl")
 include("constraints/selector.jl")
@@ -35,5 +36,6 @@ include("constraints/disjunct.jl")
 include("constraints/indicator.jl")
 include("constraints/bigm.jl")
 include("constraints/hull.jl")
+include("constraints/fallback.jl")
 include("disjunction.jl")
 include("solve.jl")
