+++ src/constraints.jl
@@ -45,17 +45,17 @@ for (RefType, loc) in ((:DisjunctConstraintRef, :disjunct_constraints),
         JuMP.index(cref::$RefType) = cref.index
 
         @doc """
-            JuMP.is_valid(model::JuMP.Model, cref::$($RefType))
+            JuMP.is_valid(model::Model, cref::$($RefType))
 
         Return `true` if `cref` refers to a valid constraint in the `GDP model`.
         """
-        function JuMP.is_valid(model::JuMP.Model, cref::$RefType) # TODO: generalize for AbstractModel
-            return model === JuMP.owner_model(cref)
+        function JuMP.is_valid(model::Model, cref::$RefType) # TODO: generalize for AbstractModel
+            return model === owner_model(cref)
         end
 
         # Get the ConstraintData object
         function _constraint_data(cref::$RefType)
-            return gdp_data(JuMP.owner_model(cref)).$loc[JuMP.index(cref)]
+            return gdp_data(owner_model(cref)).$loc[index(cref)]
         end
 
         @doc """
@@ -74,7 +74,7 @@ for (RefType, loc) in ((:DisjunctConstraintRef, :disjunct_constraints),
         """
         function JuMP.set_name(cref::$RefType, name::String)
             _constraint_data(cref).name = name
-            _set_ready_to_optimize(JuMP.owner_model(cref), false)
+            _set_ready_to_optimize(owner_model(cref), false)
             return
         end
 
@@ -94,25 +94,25 @@ for (RefType, loc) in ((:DisjunctConstraintRef, :disjunct_constraints),
         end
         Base.copy(cref::$RefType) = cref
         @doc """
-            Base.getindex(map::JuMP.GenericReferenceMap, cref::$($RefType))
+            Base.getindex(map::GenericReferenceMap, cref::$($RefType))
 
         ...
         """
-        function Base.getindex(map::JuMP.ReferenceMap, cref::$RefType)
-            $RefType(map.model, JuMP.index(cref))
+        function Base.getindex(map::ReferenceMap, cref::$RefType)
+            $RefType(map.model, index(cref))
         end
     end
 end
 
 # Extend delete
 """
-    JuMP.delete(model::JuMP.Model, cref::DisjunctionRef)
+    JuMP.delete(model::Model, cref::DisjunctionRef)
 
 Delete a disjunction constraint from the `GDP model`.
 """
-function JuMP.delete(model::JuMP.Model, cref::DisjunctionRef)
-    @assert JuMP.is_valid(model, cref) "Disjunctive constraint does not belong to model."
-    cidx = JuMP.index(cref)
+function JuMP.delete(model::Model, cref::DisjunctionRef)
+    @assert is_valid(model, cref) "Disjunctive constraint does not belong to model."
+    cidx = index(cref)
     dict = _disjunctions(model)
     delete!(dict, cidx)
     _set_ready_to_optimize(model, false)
@@ -120,13 +120,13 @@ function JuMP.delete(model::JuMP.Model, cref::DisjunctionRef)
 end
 
 """
-    JuMP.delete(model::JuMP.Model, cref::DisjunctConstraintRef)
+    JuMP.delete(model::Model, cref::DisjunctConstraintRef)
 
 Delete a disjunct constraint from the `GDP model`.
 """
-function JuMP.delete(model::JuMP.Model, cref::DisjunctConstraintRef)
-    @assert JuMP.is_valid(model, cref) "Disjunctive constraint does not belong to model."
-    cidx = JuMP.index(cref)
+function JuMP.delete(model::Model, cref::DisjunctConstraintRef)
+    @assert is_valid(model, cref) "Disjunctive constraint does not belong to model."
[…]
