+++ test/constraints/hull.jl
@@ -39,9 +39,10 @@ function test_disaggregate_variables()
     @variable(model, 10 <= x <= 100)
     @variable(model, y, Bin)
     @variable(model, z, Logical)
-    vrefs = Set([x,y])
+    vrefs = Set{VariableRef}() #initialize empty set to check if method.disjunct_variables has variables added to it in _disaggregate_variable call
     DP._reformulate_logical_variables(model)
     method = DP._Hull(Hull(1e-3, Dict(x => (0., 100.))), vrefs)
+    vrefs = Set([x,y])
     DP._disaggregate_variables(model, z, vrefs, method)
 
     refvars = DP._reformulation_variables(model)
@@ -74,6 +75,38 @@ function test_aggregate_variable()
     @test length(refcons) == 1
     @test refcons[1].func == -x + sum(method.disjunction_variables[x])
     @test refcons[1].set == MOI.EqualTo(0.)
+end
+
+function test_disaggregate_expression_var_binary()
+    model = GDPModel()
+    @variable(model, x, Bin)
+    @variable(model, z, Logical)
+    DP._reformulate_logical_variables(model)
+    bvrefs = DP._indicator_to_binary(model)
+
+    vrefs = Set([x])
+    method = DP._Hull(Hull(1e-3, Dict(x => (0., 1.))), vrefs)
+    DP._disaggregate_variables(model, z, vrefs, method)
+    @test isnothing(variable_by_name(model, "x_z"))
+    
+    refexpr = DP._disaggregate_expression(model, x, bvrefs[z], method)
+    @test refexpr == x
+end
+
+function test_disaggregate_expression_var()
+    model = GDPModel()
+    @variable(model, 10 <= x <= 100)
+    @variable(model, z, Logical)
+    DP._reformulate_logical_variables(model)
+    bvrefs = DP._indicator_to_binary(model)
+
+    vrefs = Set([x])
+    method = DP._Hull(Hull(1e-3, Dict(x => (0., 100.))), vrefs)
+    DP._disaggregate_variables(model, z, vrefs, method)
+    
+    refexpr = DP._disaggregate_expression(model, x, bvrefs[z], method)
+    x_z = variable_by_name(model, "x_z")
+    @test refexpr == x_z
 end
 
 function test_disaggregate_expression_affine()
@@ -129,6 +162,21 @@ function test_disaggregate_nl_expression_c()
     
     refexpr = DP._disaggregate_nl_expression(model, 1, bvrefs[z], method)
     @test refexpr == 1
+end
+
+function test_disaggregate_nl_expression_var_binary()
+    model = GDPModel()
+    @variable(model, x, Bin)
+    @variable(model, z, Logical)
+    DP._reformulate_logical_variables(model)
+    bvrefs = DP._indicator_to_binary(model)
+
+    vrefs = Set([x])
+    method = DP._Hull(Hull(1e-3, Dict(x => (0., 1.))), vrefs)
+    DP._disaggregate_variables(model, z, vrefs, method)
+    
+    refexpr = DP._disaggregate_nl_expression(model, x, bvrefs[z], method)
+    @test refexpr == x
 end
 
 function test_disaggregate_nl_expression_var()
@@ -548,9 +596,12 @@ end
     test_query_variable_bounds_error2()
     test_disaggregate_variables()
     test_aggregate_variable()
+    test_disaggregate_expression_var_binary()
+    test_disaggregate_expression_var()
     test_disaggregate_expression_affine()
     test_disaggregate_expression_quadratic()
     test_disaggregate_nl_expression_c()
+    test_disaggregate_nl_expression_var_binary()
     test_disaggregate_nl_expression_var()
     test_disaggregate_nl_expression_aff()
     test_disaggregate_nl_expression_quad()
