### 2023-10-16 fae47b1
add Hull tests/bug fix
