+++ test/runtests.jl
@@ -37,5 +37,5 @@ include("constraints/indicator.jl")
 include("constraints/bigm.jl")
 include("constraints/hull.jl")
 include("constraints/fallback.jl")
-include("disjunction.jl")
+include("constraints/disjunction.jl")
 include("solve.jl")
