### 2023-10-26 c054578
Enable `exactly1` option and fix bugs
