+++ test/constraints/indicator.jl
@@ -13,10 +13,10 @@ function test_indicator_scalar_constraints()
     
     ref_cons = DP._reformulation_constraints(model)
     ref_cons_obj = constraint_object.(ref_cons)
-    @test length(ref_cons) == 6
+    @test length(ref_cons) == 7
     @test all(is_valid.(model, ref_cons))
-    @test all(isa.(ref_cons_obj, VectorConstraint))
-    @test all([cobj.set isa MOI.Indicator for cobj in ref_cons_obj])
+    @test all(isa.(ref_cons_obj[1:6], VectorConstraint))
+    @test all([cobj.set isa MOI.Indicator for cobj in ref_cons_obj[1:6]])
 end
 
 function test_indicator_vector_constraints()
@@ -32,10 +32,10 @@ function test_indicator_vector_constraints()
     
     ref_cons = DP._reformulation_constraints(model)
     ref_cons_obj = constraint_object.(ref_cons)
-    @test length(ref_cons) == 6
+    @test length(ref_cons) == 7
     @test all(is_valid.(model, ref_cons))
-    @test all(isa.(ref_cons_obj, VectorConstraint))
-    @test all([cobj.set isa MOI.Indicator for cobj in ref_cons_obj])
+    @test all(isa.(ref_cons_obj[1:6], VectorConstraint))
+    @test all([cobj.set isa MOI.Indicator for cobj in ref_cons_obj[1:6]])
 end
 
 function test_indicator_array()
@@ -44,7 +44,7 @@ function test_indicator_array()
     @variable(model, y[1:2], Logical)
     @constraint(model, [1:3, 1:2], x <= 6, Disjunct(y[1]))
     @constraint(model, [1:3, 1:2], x >= 6, Disjunct(y[2]))
-    @disjunction(model, y)
+    @disjunction(model, y, exactly1 = false)
     reformulate_model(model, Indicator())
 
     ref_cons = DP._reformulation_constraints(model)
@@ -61,7 +61,7 @@ function test_indicator_dense_axis()
     @variable(model, y[1:2], Logical)
     @constraint(model, [["a","b","c"],[1,2]], x <= 7, Disjunct(y[1]))
     @constraint(model, [["a","b","c"],[1,2]], x >= 7, Disjunct(y[2]))
-    @disjunction(model, y)
+    @disjunction(model, y, exactly1 = false)
     reformulate_model(model, Indicator())    
 
     ref_cons = DP._reformulation_constraints(model)
@@ -78,7 +78,7 @@ function test_indicator_sparse_axis()
     @variable(model, y[1:2], Logical)
     @constraint(model, [i = 1:3, j = 1:3; j > i], x <= 7, Disjunct(y[1]))
     @constraint(model, [i = 1:3, j = 1:3; j > i], x >= 7, Disjunct(y[2]))
-    @disjunction(model, y)
+    @disjunction(model, y, exactly1 = false)
     reformulate_model(model, Indicator()) 
 
     ref_cons = DP._reformulation_constraints(model)
@@ -96,10 +96,10 @@ function test_indicator_nested()
     @variable(model, z[1:2], Logical)
     @constraint(model, x <= 5, Disjunct(y[1]))
     @constraint(model, x >= 5, Disjunct(y[2]))
-    @disjunction(model, y, Disjunct(z[1]))
+    @disjunction(model, y, Disjunct(z[1]), exactly1 = false)
     @constraint(model, x <= 10, Disjunct(z[1]))
     @constraint(model, x >= 10, Disjunct(z[2]))
-    @disjunction(model, z)
+    @disjunction(model, z, exactly1 = false)
     reformulate_model(model, Indicator())
 
     ref_cons = DP._reformulation_constraints(model)
