+++ test/constraints/bigm.jl
@@ -288,10 +288,10 @@ function test_nested_bigm()
     @variable(model, z[1:2], Logical)
     @constraint(model, x <= 5, Disjunct(y[1]))
     @constraint(model, x >= 5, Disjunct(y[2]))
-    @disjunction(model, inner, y, Disjunct(z[1]))
+    @disjunction(model, inner, y, Disjunct(z[1]), exactly1 = false)
     @constraint(model, x <= 10, Disjunct(z[1]))
     @constraint(model, x >= 10, Disjunct(z[2]))
-    @disjunction(model, outer, z)
+    @disjunction(model, outer, z, exactly1 = false)
 
     reformulate_model(model, BigM())
     bvrefs = DP._indicator_to_binary(model)
