+++ src/macros.jl
@@ -144,7 +144,10 @@ which must be a `Vector` of `LogicalVariableRef`s.
     @disjunction(model, ref[i=..., j=..., ...], expr, kw_args...)
 
 Add a group of disjunction described by the expression `expr` parameterized
-by `i`, `j`, ..., which must be a `Vector` of `LogicalVariableRef`s.
+by `i`, `j`, ..., which must be a `Vector` of `LogicalVariableRef`s. 
+
+In both of the above calls, a [`Disjunct`](@ref) tag can be added to created 
+nested disjunctions.
 
 The recognized keyword arguments in `kw_args` are the following:
 -  `base_name`: Sets the name prefix used to generate constraint names. 
@@ -152,6 +155,10 @@ The recognized keyword arguments in `kw_args` are the following:
     the constraint names are set to `base_name[...]` for each index `...` 
         of the axes `axes`.
 -  `container`: Specify the container type.
+-  `exactly1`: Specify a `Bool` whether an exactly one constraint for the indicator 
+   variables should be added.
+
+To create disjunctions without macros, see [`disjunction`](@ref).
 """
 macro disjunction(model, args...)
     # prepare the model 

[…]
