### 2023-10-27 e5f686f
Fix docstrings and remove unnecessary GDPData method
