+++ src/reformulate.jl
@@ -2,12 +2,12 @@
 #                              REFORMULATE
 ################################################################################
 """
-    reformulate_model(model::Model, method::AbstractSolutionMethod)
+    reformulate_model(model::JuMP.Model, method::AbstractSolutionMethod)
 
 Reformulate a `GDPModel` using the specified `method`. Prior to reformulation,
 all previous reformulation variables and constraints are deleted.
 """
-function reformulate_model(model::Model, method::AbstractSolutionMethod)
+function reformulate_model(model::JuMP.Model, method::AbstractSolutionMethod)
     #clear all previous reformulations
     _clear_reformulations(model)
     #reformulate
@@ -19,7 +19,7 @@ function reformulate_model(model::Model, method::AbstractSolutionMethod)
     _set_ready_to_optimize(model, true)
 end
 
-function _clear_reformulations(model::Model)
+function _clear_reformulations(model::JuMP.Model)
     delete.(model, _reformulation_constraints(model))
     delete.(model, _reformulation_variables(model))
     empty!(gdp_data(model).reformulation_constraints)
@@ -30,7 +30,7 @@ end
 #                              LOGICAL VARIABLES
 ################################################################################
 # create binary (indicator) variables for logic variables.
-function _reformulate_logical_variables(model::Model)
+function _reformulate_logical_variables(model::JuMP.Model)
     for (lv_idx, lv_data) in _logical_variables(model)
         lv = lv_data.variable
         lvref = LogicalVariableRef(model, lv_idx)
@@ -47,7 +47,7 @@ end
 #                              DISJUNCTIONS
 ################################################################################
 # disjunctions
-function _reformulate_all_disjunctions(model::Model, method::AbstractReformulationMethod)
+function _reformulate_all_disjunctions(model::JuMP.Model, method::AbstractReformulationMethod)
     for (_, disj) in _disjunctions(model)
         disj.constraint.nested && continue #only reformulate top level disjunctions
         ref_cons = reformulate_disjunction(model, disj.constraint, method)
@@ -58,14 +58,14 @@ function _reformulate_all_disjunctions(model::Model, method::AbstractReformulati
         end
     end
 end
-function _reformulate_disjunctions(model::Model, method::AbstractReformulationMethod)
+function _reformulate_disjunctions(model::JuMP.Model, method::AbstractReformulationMethod)
     _reformulate_all_disjunctions(model, method)
 end
-function _reformulate_disjunctions(model::Model, method::BigM)
+function _reformulate_disjunctions(model::JuMP.Model, method::BigM)
     method.tighten && _query_variable_bounds(model, method)
     _reformulate_all_disjunctions(model, method)
 end
-function _reformulate_disjunctions(model::Model, method::Hull)
+function _reformulate_disjunctions(model::JuMP.Model, method::Hull)
     _query_variable_bounds(model, method)
     _reformulate_all_disjunctions(model, method)
 end
@@ -73,7 +73,7 @@ end
 # disjuncts
 """
     reformulate_disjunction(
-        model::Model, 
+        model::JuMP.Model, 
         disj::Disjunction,
         method::AbstractReformulationMethod
     ) where {T<:Disjunction}
@@ -85,7 +85,7 @@ The `disj` field is the `ConstraintData` object for the disjunction, stored in t
 `disjunctions` field of the `GDPData` object.
 """
 # generic fallback (e.g., BigM, Indicator)
-function reformulate_disjunction(model::Model, disj::Disjunction, method::AbstractReformulationMethod)
[…]
