+++ Project.toml
@@ -8,7 +8,7 @@ JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 
 [compat]
-JuMP = "1.15"
+JuMP = "1.16"
 Reexport = "1"
 julia = "1.6"
 

[…]
