+++ test/constraints/proposition.jl
@@ -13,6 +13,8 @@ function test_proposition_add_fail()
     @test_throws ErrorException @constraint(m, logical_or(y...) == 2)
     @test_throws ErrorException @constraint(m, logical_or(y...) <= 1)
     @test_throws ErrorException @constraint(m, sum(y) := true)
+    @test_throws ErrorException @constraint(m, y[1]^2 := true)
+    @test_throws ErrorException @constraint(m, sum(y) == true)
     @test_throws ErrorException @constraint(m, prod(y) := true)
     @test_throws ErrorException @constraint(m, sin(y[1]) := true)
     @test_throws ErrorException @constraint(m, logical_or(y...) == true)
