### 2023-10-28 335a13d
update readme and docstrings
