### 2023-10-28 2aa7118
minor changes to docstrings
