### 2023-10-28 f3ba855
doc and coverage fixes
