### 2023-10-29 b26e1ed
Refactor Logical Constraints to use `:=` (#91)
* refactor for :=

* Improve coverage and fix macro bug

* fix macro bug

* another bug fix

* fix bug

* doc and coverage fixes

* improve coverage

* add test
