+++ test/model.jl
@@ -11,8 +11,10 @@ function test_empty_model()
     @test isempty(DP._logical_constraints(model))
     @test isempty(DP._disjunct_constraints(model))
     @test isempty(DP._disjunctions(model))
+    @test isempty(DP._exactly1_constraints(model))
     @test isempty(DP._indicator_to_binary(model))
     @test isempty(DP._indicator_to_constraints(model))
+    @test isempty(DP._constraint_to_indicator(model))
     @test isempty(DP._reformulation_variables(model))
     @test isempty(DP._reformulation_constraints(model))
     @test isnothing(DP._solution_method(model))
