+++ test/constraints/proposition.jl
@@ -8,22 +8,25 @@ end
 function test_proposition_add_fail()
     m = GDPModel()
     @variable(m, y[1:3], Logical)
-    @test_throws ErrorException @constraint(m, y[1] in IsTrue())
-    @test_throws ErrorException @constraint(Model(), logical_or(y...) in IsTrue())
+    @test_throws ErrorException @constraint(m, y[1] := true)
+    @test_throws ErrorException @constraint(Model(), logical_or(y...) := true)
     @test_throws ErrorException @constraint(m, logical_or(y...) == 2)
     @test_throws ErrorException @constraint(m, logical_or(y...) <= 1)
-    @test_throws ErrorException @constraint(m, sum(y) in IsTrue())
-    @test_throws ErrorException @constraint(m, prod(y) in IsTrue())
-    @test_throws ErrorException @constraint(m, sin(y[1]) in IsTrue())
-    @test_throws MethodError @constraint(m, logical_or(y...) == IsTrue())
-    @test_throws AssertionError @constraint(GDPModel(), logical_or(y...) in IsTrue())
+    @test_throws ErrorException @constraint(m, sum(y) := true)
+    @test_throws ErrorException @constraint(m, y[1]^2 := true)
+    @test_throws ErrorException @constraint(m, sum(y) == true)
+    @test_throws ErrorException @constraint(m, prod(y) := true)
+    @test_throws ErrorException @constraint(m, sin(y[1]) := true)
+    @test_throws ErrorException @constraint(m, logical_or(y...) == true)
+    @test_throws VariableNotOwned @constraint(GDPModel(), logical_or(y...) := true)
+    @test_throws AssertionError add_constraint(m, ScalarConstraint(logical_or(y...), MOI.LessThan(42)))
 end
 
 function test_negation_add_success()
     model = GDPModel()
     @variable(model, y, Logical)
-    c1 = @constraint(model, logical_not(y) in IsTrue())
-    @constraint(model, c2, ¬y in IsTrue())
+    c1 = @constraint(model, logical_not(y) := true)
+    @constraint(model, c2, ¬y := true)
     @test is_valid(model, c1)
     @test is_valid(model, c2)
     @test owner_model(c1) == model
@@ -42,44 +45,46 @@ function test_negation_add_success()
     @test constraint_object(c1).func.args ==
             constraint_object(c2).func.args == Any[y]
     @test constraint_object(c1).set == 
-            constraint_object(c2).set == IsTrue()
+            constraint_object(c2).set == MOI.EqualTo{Bool}(true)
     @test c1 == copy(c1)
 end
 
 function test_implication_add_success()
     model = GDPModel()
     @variable(model, y[1:2], Logical)
-    @constraint(model, c1, implies(y...) in IsTrue())
-    @constraint(model, c2, (y[1] ⟹ y[2]) in IsTrue())
-    @test_macro_throws ErrorException @constraint(model, y[1] ⟹ y[2] in IsTrue())
+    @constraint(model, c1, implies(y...) := true)
+    @constraint(model, c2, y[1] ⟹ y[2] := true)
+    @constraint(model, c3, y[1] ⟹ y[2] := false)
     @test constraint_object(c1).func.head == 
             constraint_object(c2).func.head == :(=>)
     @test constraint_object(c1).func.args ==
             constraint_object(c2).func.args == Vector{Any}(y)
     @test constraint_object(c1).set == 
-            constraint_object(c2).set == IsTrue()
+            constraint_object(c2).set == MOI.EqualTo{Bool}(true)
+    @test constraint_object(c3).func.head == :!
[…]
