+++ src/constraints.jl
@@ -93,14 +93,14 @@ for (RefType, loc) in ((:DisjunctConstraintRef, :disjunct_constraints),
             return cref1.model === cref2.model && cref1.index == cref2.index
         end
         Base.copy(cref::$RefType) = cref
-        @doc """
-            Base.getindex(map::GenericReferenceMap, cref::$($RefType))
+        # @doc """
+        #     Base.getindex(map::GenericReferenceMap, cref::$($RefType))
 
-        ...
-        """
-        function Base.getindex(map::ReferenceMap, cref::$RefType)
-            $RefType(map.model, index(cref))
-        end
+        # ...
+        # """
+        # function Base.getindex(map::ReferenceMap, cref::$RefType)
+        #     $RefType(map.model, index(cref))
+        # end
     end
 end
 
@@ -432,19 +432,6 @@ end
 ################################################################################
 #                              LOGICAL CONSTRAINTS
 ################################################################################
-JuMP.operator_to_set(_error::Function, ::Val{:⟹}) = _error(
-    "Cannot use ⟹ in a MOI set (invalid right-hand side). If you are seeing this error, " *
-    "you likely added a logical constraint of the form A ⟹ B ∈ IsTrue(). " *
-    "Instead, you should enclose the constraint function in parenthesis: " *
-    "(A ⟹ B) ∈ IsTrue()."
-)
-JuMP.operator_to_set(_error::Function, ::Val{:⇔}) = _error(
-    "Cannot use ⇔ in a MOI set (invalid right-hand side). If you are seeing this error, " *
-    "you likely added a logical constraint of the form A ⇔ B ∈ IsTrue(). " *
-    "Instead, you should enclose the constraint function in parenthesis: " *
-    "(A ⇔ B) ∈ IsTrue()."
-)
-
 """
     function JuMP.build_constraint(
         _error::Function, 
@@ -468,17 +455,6 @@ model = GDPModel();
 @variable(model, Y[i = 1:2], LogicalVariable);
 @constraint(model, [Y[1], Y[2]] in Exactly(1));
 ```
-
-    JuMP.build_constraint(
-        _error::Function, 
-        func::_LogicalExpr,
-        set::IsTrue
-    )
-
-Extend `JuMP.build_constraint` to add logical propositional constraints to a [`GDPModel`](@ref). 
-This in combination with `JuMP.add_constraint` enables the use of 
-`@constraint(model, [name], logical_expr in IsTrue())` to define a Boolean expression that must
-either be true or false.
 """
 function JuMP.build_constraint( # Cardinality logical constraint
     _error::Function, 
@@ -497,31 +473,6 @@ function JuMP.build_constraint( # Cardinality logical constraint
     _error("Selector constraints can only be applied to a Vector or Container of LogicalVariableRefs.")
 end
 
-# Proposition logical constraint: _LogicalExpr
-function JuMP.build_constraint(
-    _error::Function, 
-    func::_LogicalExpr,
-    set::IsTrue
-    )
-    if !(func.head in _LogicalOperatorHeads)
-        _error("Unrecognized logical operator `$(func.head)`.")
-    else
-        return ScalarConstraint(func, set)
-    end
-end
-
-# Fallback for LogicalVariableRef in IsTrue
-function JuMP.build_constraint(
-    _error::Function, 
-    func::LogicalVariableRef,
-    set::IsTrue
-    )
-    _error(
-        "Logical propositions must be of the form `logical_expr in IsTrue()`. " *
[…]

[…]
