+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#API","page":"API","title":"API","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"Modules = [DisjunctiveProgramming]\nOrder   = [:macro, :function, :type]","category":"page"},{"location":"api/#DisjunctiveProgramming.@disjunction-Tuple{Any, Vararg{Any}}","page":"API","title":"DisjunctiveProgramming.@disjunction","text":"@disjunction(model, expr, kw_args...)\n\nAdd a disjunction described by the expression expr,  which must be a Vector of LogicalVariableRefs.\n\n@disjunction(model, ref[i=..., j=..., ...], expr, kw_args...)\n\nAdd a group of disjunction described by the expression expr parameterized by i, j, ..., which must be a Vector of LogicalVariableRefs. \n\nIn both of the above calls, a Disjunct tag can be added to create  nested disjunctions.\n\nThe recognized keyword arguments in kw_args are the following:\n\nbase_name: Sets the name prefix used to generate constraint names.    It corresponds to the constraint name for scalar constraints, otherwise,    the constraint names are set to base_name[...] for each index ...        of the axes axes.\ncontainer: Specify the container type.\nexactly1: Specify a Bool whether a constraint should be added to   only allow selecting one disjunct in the disjunction.\n\nTo create disjunctions without macros, see disjunction.\n\n\n\n\n\n","category":"macro"},{"location":"api/#DisjunctiveProgramming.@disjunctions-Tuple{Any, Any}","page":"API","title":"DisjunctiveProgramming.@disjunctions","text":"@disjunctions(model, args...)\n\nAdds groups of disjunctions at once, in the same fashion as the @disjunction macro.\n\nThe model must be the first argument, and multiple disjunctions can be added on multiple  lines wrapped in a begin ... end block.\n\nThe macro returns a tuple containing the disjunctions that were defined.\n\nExample\n\njulia model = GDPModel(); @variable(model, w); @variable(model, x); @variable(model, Y[1:4], LogicalVariable); @constraint(model, [i=1:2], w == i, Disjunct(Y[i])); @constraint(model, [i=3:4], x == i, Disjunct(Y[i])); @disjunctions(model, begin     [Y[1], Y[2]]     [Y[3], Y[4]] end);`\n\n\n\n\n\n","category":"macro"},{"location":"api/#Base.getindex-Tuple{ReferenceMap, DisjunctConstraintRef}","page":"API","title":"Base.getindex","text":"Base.getindex(map::GenericReferenceMap, cref::DisjunctConstraintRef)\n\n...\n\n\n\n\n\n","category":"method"},{"location":"api/#Base.getindex-Tuple{ReferenceMap, DisjunctionRef}","page":"API","title":"Base.getindex","text":"Base.getindex(map::GenericReferenceMap, cref::DisjunctionRef)\n\n...\n\n\n\n\n\n","category":"method"},{"location":"api/#Base.getindex-Tuple{ReferenceMap, LogicalConstraintRef}","page":"API","title":"Base.getindex","text":"Base.getindex(map::GenericReferenceMap, cref::LogicalConstraintRef)\n\n...\n\n\n\n\n\n","category":"method"},{"location":"api/#DisjunctiveProgramming.GDPModel-Tuple","page":"API","title":"DisjunctiveProgramming.GDPModel","text":"GDPModel([optimizer]; [kwargs...])::Model\n\nThe core model object for building general disjunction programming models.\n\n\n\n\n\n","category":"method"},{"locatio
[…]
