### 2023-10-30 642e12b
add checks to logical propositions to disallow propositions with only 1 logical variable
