### 2023-11-04 ba67fd3
Support `AbstractModel`s
