+++ test/constraints/hull.jl
@@ -1,37 +1,30 @@
 function test_default_hull()
-    method = Hull()
-    @test method.value == 1e-6
+    @test Hull().value == 1e-6
 end
 
 function test_set_hull()
-    method = Hull(0.001)
-    @test method.value == 0.001
+    @test Hull(0.001).value == 0.001
 end
 
 function test_query_variable_bounds()
     model = GDPModel()
     @variable(model, 10 <= x <= 100)
     @variable(model, -100 <= y <= -10)
-    method = Hull()
-    DP._query_variable_bounds(model, method)
-    @test haskey(method.variable_bounds, x)
-    @test haskey(method.variable_bounds, y)
-    @test method.variable_bounds[x] == (0, 100)
-    @test method.variable_bounds[y] == (-100, 0)
+    prep_bounds([x, y], model, Hull())
+    @test variable_bound_info(x) == (0, 100)
+    @test variable_bound_info(y) == (-100, 0)
 end
 
 function test_query_variable_bounds_error1()
     model = GDPModel()
     @variable(model, x <= 100)
-    method = Hull()
-    @test_throws ErrorException DP._query_variable_bounds(model, method)
+    @test_throws ErrorException set_variable_bound_info(x, Hull())
 end
 
 function test_query_variable_bounds_error2()
     model = GDPModel()
     @variable(model, -100 <= x)
-    method = Hull()
-    @test_throws ErrorException DP._query_variable_bounds(model, method)
+    @test_throws ErrorException set_variable_bound_info(x, Hull())
 end
 
 function test_disaggregate_variables()
@@ -39,11 +32,12 @@ function test_disaggregate_variables()
     @variable(model, 10 <= x <= 100)
     @variable(model, y, Bin)
     @variable(model, z, Logical)
-    vrefs = Set{VariableRef}() #initialize empty set to check if method.disjunct_variables has variables added to it in _disaggregate_variable call
-    DP._reformulate_logical_variables(model)
-    method = DP._Hull(Hull(1e-3, Dict(x => (0., 100.))), vrefs)
     vrefs = Set([x,y])
-    DP._disaggregate_variables(model, z, vrefs, method)
+    DP._reformulate_logical_variables(model)
+    @test prep_bounds(x, model, Hull()) isa Nothing
+    method = DP._Hull(Hull(1e-3), vrefs)
+    @test DP._disaggregate_variables(model, z, vrefs, method) isa Nothing
+    @test haskey(method.disjunct_variables, (x, DP._indicator_to_binary(model)[z]))
 
     refvars = DP._reformulation_variables(model)
     @test length(refvars) == 2
@@ -68,8 +62,9 @@ function test_aggregate_variable()
     @variable(model, z, Logical)
     vrefs = Set([x])
     DP._reformulate_logical_variables(model)
-    method = DP._Hull(Hull(1e-3, Dict(x => (0., 100.))), vrefs)
-    DP._disaggregate_variables(model, z, vrefs, method)
+    @test prep_bounds(x, model, Hull()) isa Nothing
+    method = DP._Hull(Hull(1e-3), vrefs)
+    @test DP._disaggregate_variables(model, z, vrefs, method) isa Nothing
     refcons = Vector{JuMP.AbstractConstraint}() 
     DP._aggregate_variable(model, refcons, x, method)
     @test length(refcons) == 1
@@ -85,8 +80,9 @@ function test_disaggregate_expression_var()
[…]
