+++ src/indicator.jl
@@ -3,20 +3,20 @@
 ################################################################################
 #scalar disjunct constraint
 function reformulate_disjunct_constraint(
-    model::Model,
+    ::JuMP.AbstractModel,
     con::ScalarConstraint{T, S},
-    bvref::VariableRef,
-    method::Indicator
+    bvref::JuMP.AbstractVariableRef,
+    ::Indicator
 ) where {T, S}
     reform_con = build_constraint(error, [1*bvref, con.func], _MOI.Indicator{_MOI.ACTIVATE_ON_ONE}(con.set))
     return [reform_con]
 end
 #vectorized disjunct constraint
 function reformulate_disjunct_constraint(
-    model::Model,
+    ::JuMP.AbstractModel,
     con::VectorConstraint{T, S},
-    bvref::VariableRef,
-    method::Indicator
+    bvref::JuMP.AbstractVariableRef,
+    ::Indicator
 ) where {T, S}
     set = _vec_to_scalar_set(con.set)
     return [
@@ -26,10 +26,10 @@ function reformulate_disjunct_constraint(
 end
 #nested indicator reformulation. NOTE: the user needs to provide the appropriate linking constraint for the logical variables for this to work (e.g. w in Exactly(y[1]) to link the parent disjunct y[1] to the nested disjunction w)
 function reformulate_disjunct_constraint(
-    model::Model,
+    ::JuMP.AbstractModel,
     con::VectorConstraint{T, S},
-    bvref::VariableRef,
-    method::Indicator
+    ::JuMP.AbstractVariableRef,
+    ::Indicator
 ) where {T, S <: _MOI.Indicator}
     return [con]
 end

[…]
