### 2023-11-05 a8ad8cd
fix optimize hook
