### 2023-11-06 14e960a
update docs and rename method keyword
