+++ src/model.jl
@@ -41,11 +41,11 @@ end
 # See https://github.com/jump-dev/JuMP.jl/blob/9ea1df38fd320f864ab4c93c78631d0f15939c0b/src/JuMP.jl#L718-L745
 function _optimize_hook(
     model::JuMP.AbstractModel; 
-    method::AbstractSolutionMethod,
+    gdp_method::AbstractSolutionMethod = BigM(),
     kwargs...
     ) # can add more kwargs if wanted
-    if !_ready_to_optimize(model) || _solution_method(model) != method
-        reformulate_model(model, method)
+    if !_ready_to_optimize(model) || _solution_method(model) != gdp_method
+        reformulate_model(model, gdp_method)
     end
     return JuMP.optimize!(model; ignore_optimize_hook = true, kwargs...)
 end
@@ -53,7 +53,6 @@ end
 ################################################################################
 #                              GDP DATA
 ################################################################################
-
 """
     gdp_data(model::JuMP.AbstractModel)::GDPData
 
