### 2023-11-06 0768056
add `binary_variable` and fix tests
