+++ test/constraints/hull.jl
@@ -33,16 +33,15 @@ function test_disaggregate_variables()
     @variable(model, y, Bin)
     @variable(model, z, Logical)
     vrefs = Set([x,y])
-    DP._reformulate_logical_variables(model)
     @test prep_bounds(x, model, Hull()) isa Nothing
     method = DP._Hull(Hull(1e-3), vrefs)
     @test DP._disaggregate_variables(model, z, vrefs, method) isa Nothing
     @test haskey(method.disjunct_variables, (x, DP._indicator_to_binary(model)[z]))
 
     refvars = DP._reformulation_variables(model)
-    @test length(refvars) == 2
+    @test length(refvars) == 1
     zbin = variable_by_name(model, "z")
-    @test zbin in refvars
+    @test zbin == binary_variable(z)
     x_z = variable_by_name(model, "x_z")
     @test x_z in refvars
     @test has_lower_bound(x_z) && lower_bound(x_z) == 0
@@ -61,7 +60,6 @@ function test_aggregate_variable()
     @variable(model, 10 <= x <= 100)
     @variable(model, z, Logical)
     vrefs = Set([x])
-    DP._reformulate_logical_variables(model)
     @test prep_bounds(x, model, Hull()) isa Nothing
     method = DP._Hull(Hull(1e-3), vrefs)
     @test DP._disaggregate_variables(model, z, vrefs, method) isa Nothing
@@ -76,7 +74,6 @@ function test_disaggregate_expression_var()
     model = GDPModel()
     @variable(model, 10 <= x <= 100)
     @variable(model, z, Logical)
-    DP._reformulate_logical_variables(model)
     bvrefs = DP._indicator_to_binary(model)
 
     vrefs = Set([x])
@@ -93,7 +90,6 @@ function test_disaggregate_expression_var_binary()
     model = GDPModel()
     @variable(model, x, Bin)
     @variable(model, z, Logical)
-    DP._reformulate_logical_variables(model)
     bvrefs = DP._indicator_to_binary(model)
 
     vrefs = Set([x])
@@ -109,7 +105,6 @@ function test_disaggregate_expression_affine()
     model = GDPModel()
     @variable(model, 10 <= x <= 100)
     @variable(model, z, Logical)
-    DP._reformulate_logical_variables(model)
     bvrefs = DP._indicator_to_binary(model)
 
     vrefs = Set([x])
@@ -128,7 +123,6 @@ function test_disaggregate_expression_affine_mip()
     @variable(model, 10 <= x <= 100)
     @variable(model, y, Bin)
     @variable(model, z, Logical)
-    DP._reformulate_logical_variables(model)
     bvrefs = DP._indicator_to_binary(model)
 
     vrefs = Set([x, y])
@@ -146,7 +140,6 @@ function test_disaggregate_expression_quadratic()
     model = GDPModel()
     @variable(model, 10 <= x <= 100)
     @variable(model, z, Logical)
-    DP._reformulate_logical_variables(model)
     bvrefs = DP._indicator_to_binary(model)
 
     vrefs = Set([x])
@@ -170,7 +163,6 @@ function test_disaggregate_nl_expression_c()
     model = GDPModel()
     @variable(model, 10 <= x <= 100)
     @variable(model, z, Logical)
-    DP._reformulate_logical_variables(model)
     bvrefs = DP._indicator_to_binary(model)
 
     vrefs = Set([x])
@@ -186,7 +178,6 @@ function test_disaggregate_nl_expression_var_binary()
     model = GDPModel()
     @variable(model, x, Bin)
[…]
