### 2023-11-16 #117 @matbesancon
@hdavid16 while we are waiting for confirmation from @joaquimg that the changes are satisfactory, one step to do is archiving on Zenodo the current version of the software and giving us the corresponding DOI that gets generated
