### 2023-11-16 d142456
set BigM as the default reformulation method
