### 2023-11-16 3fa57e7
add default reformulation method: BigM()
