+++ v0.4.1-zenodo/siteinfo.js
@@ -1,0 +1,1 @@
+var DOCUMENTER_CURRENT_VERSION = "v0.4.1-zenodo";
