### 2023-11-17 #117 @matbesancon
@hdavid16 for stylistic reasons, can we replace the successive equations by an `align` block? This will look much nicer
