### 2023-11-17 #117 @matbesancon
same for all optimization problems, prefer a template like:
```
\begin{align*}
\min_{x}\,\, & f(x) \\
\text{s.t. }\, & g(x) \leq 0 \\
& x_i \in X && \forall i \in .. \\
& x \in \mathcal{X}
\end{align*}
```
