### 2023-12-05 0caa95e
support `value` for logical variables
