### 2023-12-06 #117 @joaquimg
I do re-iterate @matbesancon comment.

Other than that, good to go on my side.

Great work and congratulations for the nice package!
