### 2023-12-06 97de4cd
return Bool values
