+++ test/solve.jl
@@ -17,19 +17,19 @@ function test_linear_gdp_example(m)
     @test termination_status(m) == MOI.OPTIMAL
     @test objective_value(m) ≈ 11
     @test value.(x) ≈ [9,2]
-    @test value(Y[1]) ≈ 0
-    @test value(Y[2]) ≈ 1
-    @test value(W[1]) ≈ 0
-    @test value(W[2]) ≈ 0
+    @test !value(Y[1])
+    @test value(Y[2])
+    @test !value(W[1])
+    @test !value(W[2])
 
     @test optimize!(m, gdp_method = Hull()) isa Nothing
     @test termination_status(m) == MOI.OPTIMAL
     @test objective_value(m) ≈ 11
     @test value.(x) ≈ [9,2]
-    @test value(Y[1]) ≈ 0
-    @test value(Y[2]) ≈ 1
-    @test value(W[1]) ≈ 0
-    @test value(W[2]) ≈ 0
+    @test !value(Y[1])
+    @test value(Y[2])
+    @test !value(W[1])
+    @test !value(W[2])
     @test value(variable_by_name(m, "x[1]_Y[1]")) ≈ 0
     @test value(variable_by_name(m, "x[1]_Y[2]")) ≈ 9
     @test value(variable_by_name(m, "x[1]_W[1]")) ≈ 0
