### 2023-12-06 d1e6cf4
Support `JuMP.value` for Logical Variables (#96)
* support `value` for logical variables

* return Bool values
