+++ src/variables.jl
@@ -267,6 +267,17 @@ function binary_variable(vref::LogicalVariableRef)
 end
 
 """
+    JuMP.value(vref::LogicalVariableRef)::Bool
+
+Returns the optimized value of `vref`. This dispatches on 
+`value(binary_variable(vref))` and then rounds to the closest 
+`Bool` value.
+"""
+function JuMP.value(vref::LogicalVariableRef)
+    return JuMP.value(binary_variable(vref)) >= 0.5
+end
+
+"""
     JuMP.delete(model::JuMP.AbstractModel, vref::LogicalVariableRef)::Nothing
 
 Delete the logical variable associated with `vref` from the `GDP model`.
