### 2023-12-06 1c89629
Update to use JuMP.SkipModelConvertScalarSetWrapper
