+++ src/constraints.jl
@@ -50,12 +50,12 @@ for (RefType, loc) in ((:DisjunctConstraintRef, :disjunct_constraints),
         Return `true` if `cref` refers to a valid constraint in the `GDP model`.
         """
         function JuMP.is_valid(model::JuMP.AbstractModel, cref::$RefType)
-            return model === owner_model(cref) && haskey(gdp_data(model).$loc, JuMP.index(cref))
+            return model === JuMP.owner_model(cref) && haskey(gdp_data(model).$loc, JuMP.index(cref))
         end
 
         # Get the ConstraintData object
         function _constraint_data(cref::$RefType)
-            return gdp_data(owner_model(cref)).$loc[index(cref)]
+            return gdp_data(JuMP.owner_model(cref)).$loc[JuMP.index(cref)]
         end
 
         @doc """
@@ -74,7 +74,7 @@ for (RefType, loc) in ((:DisjunctConstraintRef, :disjunct_constraints),
         """
         function JuMP.set_name(cref::$RefType, name::String)
             _constraint_data(cref).name = name
-            _set_ready_to_optimize(owner_model(cref), false)
+            _set_ready_to_optimize(JuMP.owner_model(cref), false)
             return
         end
 
@@ -110,13 +110,13 @@ end
 Delete a disjunction constraint from the `GDP model`.
 """
 function JuMP.delete(model::JuMP.AbstractModel, cref::DisjunctionRef)
-    @assert is_valid(model, cref) "Disjunction does not belong to model."
+    @assert JuMP.is_valid(model, cref) "Disjunction does not belong to model."
     if JuMP.constraint_object(cref).nested
         lvref = gdp_data(model).constraint_to_indicator[cref]
         filter!(Base.Fix2(!=, cref), _indicator_to_constraints(model)[lvref])
         delete!(gdp_data(model).constraint_to_indicator, cref)
     end
-    delete!(_disjunctions(model), index(cref))
+    delete!(_disjunctions(model), JuMP.index(cref))
     exactly1_dict = gdp_data(model).exactly1_constraints
     if haskey(exactly1_dict, cref)
         JuMP.delete(model, exactly1_dict[cref])
@@ -132,8 +132,8 @@ end
 Delete a disjunct constraint from the `GDP model`.
 """
 function JuMP.delete(model::JuMP.AbstractModel, cref::DisjunctConstraintRef)
-    @assert is_valid(model, cref) "Disjunctive constraint does not belong to model."
-    delete!(_disjunct_constraints(model), index(cref))
+    @assert JuMP.is_valid(model, cref) "Disjunctive constraint does not belong to model."
+    delete!(_disjunct_constraints(model), JuMP.index(cref))
     lvref = gdp_data(model).constraint_to_indicator[cref]
     filter!(Base.Fix2(!=, cref), _indicator_to_constraints(model)[lvref])
     delete!(gdp_data(model).constraint_to_indicator, cref)
@@ -147,8 +147,8 @@ end
 Delete a logical constraint from the `GDP model`.
 """
 function JuMP.delete(model::JuMP.AbstractModel, cref::LogicalConstraintRef)
-    @assert is_valid(model, cref) "Logical constraint does not belong to model."
-    delete!(_logical_constraints(model), index(cref))
+    @assert JuMP.is_valid(model, cref) "Logical constraint does not belong to model."
+    delete!(_logical_constraints(model), JuMP.index(cref))
     _set_ready_to_optimize(model, false)
     return 
 end
@@ -159,7 +159,7 @@ end
 function _check_expression(expr::Ex) where {Ex <: JuMP.AbstractJuMPScalar}
[…]

[…]
