### 2023-12-07 177c864
Update to use JuMP.SkipModelConvertScalarSetWrapper (#97)
