+++ Project.toml
@@ -9,7 +9,7 @@ Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 
 [compat]
 Aqua = "0.8"
-JuMP = "1.16"
+JuMP = "1.17"
 Reexport = "1"
 julia = "1.6"
 
