+++ src/hull.jl
@@ -51,7 +51,7 @@ function _disaggregate_variables(
     )
     #create disaggregated variables for that disjunct
     for vref in vrefs
-        if !requires_disaggregation(vref) || is_binary(vref) 
+        if !requires_disaggregation(vref) || JuMP.is_binary(vref) 
             continue # skip variables that don't require dissagregation
         end
         _disaggregate_variable(model, lvref, vref, method) #create disaggregated var for that disjunct
@@ -73,11 +73,11 @@ function _disaggregate_variable(
     push!(method.disjunction_variables[vref], dvref)
     method.disjunct_variables[vref, bvref] = dvref
     #create bounding constraints
-    dvname = name(dvref)
+    dvname = JuMP.name(dvref)
     lbname = isempty(dvname) ? "" : "$(dvname)_lower_bound"
     ubname = isempty(dvname) ? "" : "$(dvname)_upper_bound"
-    new_con_lb_ref = @constraint(model, lb*bvref - dvref <= 0, base_name = lbname)
-    new_con_ub_ref = @constraint(model, dvref - ub*bvref <= 0, base_name = ubname)
+    new_con_lb_ref = JuMP.@constraint(model, lb*bvref - dvref <= 0, base_name = lbname)
+    new_con_ub_ref = JuMP.@constraint(model, dvref - ub*bvref <= 0, base_name = ubname)
     push!(_reformulation_constraints(model), new_con_lb_ref, new_con_ub_ref)
     return dvref
 end
@@ -91,9 +91,9 @@ function _aggregate_variable(
     vref::JuMP.AbstractVariableRef, 
     method::_Hull
     )
-    is_binary(vref) && return #skip binary variables
-    con_expr = @expression(model, -vref + sum(method.disjunction_variables[vref]))
-    push!(ref_cons, build_constraint(error, con_expr, _MOI.EqualTo(0)))
+    JuMP.is_binary(vref) && return #skip binary variables
+    con_expr = JuMP.@expression(model, -vref + sum(method.disjunction_variables[vref]))
+    push!(ref_cons, JuMP.build_constraint(error, con_expr, _MOI.EqualTo(0)))
     return 
 end
 
@@ -107,7 +107,7 @@ function _disaggregate_expression(
     bvref::JuMP.AbstractVariableRef, 
     method::_Hull
     )
-    if is_binary(vref) || !haskey(method.disjunct_variables, (vref, bvref)) #keep any binary variables or nested disaggregated variables unchanged 
+    if JuMP.is_binary(vref) || !haskey(method.disjunct_variables, (vref, bvref)) #keep any binary variables or nested disaggregated variables unchanged 
         return vref #NOTE: not needed because nested constraint of the form `vref in MOI.AbstractScalarSet` gets reformulated to an affine expression.
     else #replace with disaggregated form
         return method.disjunct_variables[vref, bvref]
@@ -122,11 +122,11 @@ function _disaggregate_expression(
     )
     new_expr = @expression(model, aff.constant*bvref) #multiply constant by binary indicator variable
     for (vref, coeff) in aff.terms
-        if is_binary(vref) || !haskey(method.disjunct_variables, (vref, bvref)) #keep any binary variables or nested disaggregated variables unchanged 
-            add_to_expression!(new_expr, coeff*vref)
+        if JuMP.is_binary(vref) || !haskey(method.disjunct_variables, (vref, bvref)) #keep any binary variables or nested disaggregated variables unchanged 
[…]

[…]
