### 2023-12-07 01eb9e8
Add printing for constraints
