+++ test/runtests.jl
@@ -19,4 +19,5 @@ include("constraints/bigm.jl")
 include("constraints/hull.jl")
 include("constraints/fallback.jl")
 include("constraints/disjunction.jl")
+include("print.jl")
 include("solve.jl")
