+++ test/print.jl
@@ -1,0 +1,136 @@
+function test_disjunct_constraints()
+    # Set up the model
+    model = GDPModel()
+    @variable(model, x[1:2])
+    @variable(model, Y[1:2], Logical)
+    @constraint(model, c1, x[1]^2 >= 3.2, Disjunct(Y[1]))
+    c2 = @constraint(model, x[2] <= 2, Disjunct(Y[2]))
+
+    # Test plain printing
+    if Sys.iswindows()
+        show_test(MIME("text/plain"), c1, "c1 : x[1]² >= 3.2, if Y[1] = True")
+        show_test(MIME("text/plain"), c2, "x[2] <= 2, if Y[2] = True") 
+    else
+        show_test(MIME("text/plain"), c1, "c1 : x[1]² ≥ 3.2, if Y[1] = True")
+        show_test(MIME("text/plain"), c2, "x[2] ≤ 2, if Y[2] = True") 
+    end
+
+    # Test math mode string
+    c1_str = "x_{1}^2 \\geq 3.2, \\; \\text{if } Y_{1} = \\text{True}"
+    @test constraint_string(MIME("text/latex"), c1, in_math_mode = true) == c1_str
+    c2_str = "x_{2} \\leq 2, \\; \\text{if } Y_{2} = \\text{True}"
+    @test constraint_string(MIME("text/latex"), c2, in_math_mode = true) == c2_str
+
+    # Test LaTeX printing
+    show_test(MIME("text/latex"), c1, "\$\$ $(c1_str) \$\$")
+    show_test(MIME("text/latex"), c2, "\$\$ $(c2_str) \$\$") 
+end
+
+function test_disjunctions()
+    # Set up the model
+    model = GDPModel()
+    @variable(model, x[1:2])
+    @variable(model, Y[1:2], Logical)
+    @constraint(model, 2x[1]^2 >= 1, Disjunct(Y[1]))
+    @constraint(model, x[2] - 1 == 2.1, Disjunct(Y[2]))
+    @constraint(model, 0 <= x[1] <= 1, Disjunct(Y[2]))
+    @disjunction(model, d1, Y)
+    d2 = disjunction(model, Y)
+
+    # Test plain printing
+    if Sys.iswindows()
+        str = "[Y[1] --> {2 x[1]² >= 1}] or [Y[2] --> {x[2] == 3.1; x[1] in [0, 1]}]"
+        show_test(MIME("text/plain"), d1, "d1 : " * str)
+        show_test(MIME("text/plain"), d2, str) 
+    else
+        str = "[Y[1] ⟹ {2 x[1]² ≥ 1}] ⋁ [Y[2] ⟹ {x[2] = 3.1; x[1] ∈ [0, 1]}]"
+        show_test(MIME("text/plain"), d1, str)
+        show_test(MIME("text/plain"), d2, str) 
+    end
+
+    # Test math mode string
+    str = "\\begin{bmatrix}\n Y_{1}\\\\\n 2 x_{1}^2 \\geq 1\\end{bmatrix} \\bigvee \\begin{bmatrix}\n Y_{2}\\\\\n x_{2} = 3.1\\\\\n x_{1} \\in [0, 1]\\end{bmatrix}"
+    @test constraint_string(MIME("text/latex"), d1, in_math_mode = true) == str
+    @test constraint_string(MIME("text/latex"), d2, in_math_mode = true) == str
+
+    # Test LaTeX printing
[…]
