### 2023-12-08 d7e8e7a
Improve Printing for Disjunct/Logical Constraints and Disjunctions (#100)
* Add printing for constraints

* fix test function naming

* attempted linux fix

* replace fix attempt

* fix attempt 2

* fix coverage

* minor formatting fix
