+++ test/utilities.jl
@@ -18,6 +18,26 @@ macro test_macro_throws(errortype, m)
     end
 end
 
+# Helper function to test IO methods work correctly
+function show_test(mode, obj, exp_str::String; repl=:both)
+    if mode == MIME("text/plain")
+        repl != :show  && @test sprint(print, obj) == exp_str
+        repl != :print && @test sprint(show,  obj) == exp_str
+    else
+        @test sprint(show, "text/latex", obj) == exp_str
+    end
+end
+
+# Helper function for IO methods with different possibilities
+function show_test(mode, obj, exp_str::Vector{String}; repl=:both)
+    if mode == MIME("text/plain")
+        repl != :show  && @test sprint(print, obj) in exp_str
+        repl != :print && @test sprint(show,  obj) in exp_str
+    else
+        @test sprint(show, "text/latex", obj) in exp_str
+    end
+end
+
 # Helper functions to prepare variable bounds without reformulating
 function prep_bounds(vref, model, method)
     if requires_variable_bound_info(method)
