+++ src/DisjunctiveProgramming.jl
@@ -22,6 +22,7 @@ include("reformulate.jl")
 include("bigm.jl")
 include("hull.jl")
 include("indicator.jl")
+include("print.jl")
 
 # Define additional stuff that should not be exported
 const _EXCLUDE_SYMBOLS = [Symbol(@__MODULE__), :eval, :include]

[…]
