### 2023-12-11 98e4782
update equations to align* blocks
