### 2023-12-11 #117 @hdavid16
@joaquimg and @matbesancon the paper has been updated on `master` as you requested (replacing all equations/expressions with `align*` blocks).
